Get data from one specific senseBox sensor.

If the arguments fromDate and toDate are NOT given, the measurements from the last 48 h from now are downloaded. The maximum numbers of downloaded measurements are 10,000. The maximum time frame is one month. It is possible to download the data directly as a csv file, when argument CSV is TRUE.

get_senseBox_data(senseBoxId, sensorId = rep("all", times =
  length(unlist(senseBoxId))), fromDate = NULL, toDate = NULL,
  CSV = FALSE, POSIXct = TRUE)

Arguments

senseBoxId
characteror vector of character (required): senseBoxId or a vector of character of senseBoxIds

sensorId
character or list of character (required): sensorId or a list of senseBoxIds

fromDate
character (optional): Just show data with from date in format YYYY-mm-dd HH:MM:SS

toDate
character (optional): Just show data with up to date in format YYYY-mm-dd HH:MM:SS

CSV
logical (optional): Download data as .csv? NOT POSSIBLE UNTIL NOW!

POSIXct
logical (optional): Should the timestamp be translated into POSIXct?

Value

list

Function version

0.0.1

Examples

##=====================================
## Example: get data from sensorId
##=====================================

get_senseBox_data(
  senseBoxId = "5957b67494f0520011304cc1",
  sensorId = "5957b67494f0520011304cc4")

#> $`5957b67494f0520011304cc1`
#> $`5957b67494f0520011304cc1`$`Temperatur [°C]`
#>      value           createdAt
#> 1     4.21 2017-11-28 13:37:11
#> 2     4.27 2017-11-28 13:34:43
#> 3     4.37 2017-11-28 13:32:15
#> 4     4.36 2017-11-28 13:29:47
#> 5     4.34 2017-11-28 13:27:19
#> 6     4.36 2017-11-28 13:24:50
#> 7     4.51 2017-11-28 13:22:22
#> 8     4.58 2017-11-28 13:19:55
#> 9     4.60 2017-11-28 13:17:25
#> 10    4.59 2017-11-28 13:14:57
#> 11    4.48 2017-11-28 13:12:29
#> 12    4.34 2017-11-28 13:10:01
#> 13    4.54 2017-11-28 13:07:34
#> 14    4.57 2017-11-28 13:05:05
#> 15    4.31 2017-11-28 13:02:38
#> 16    4.26 2017-11-28 13:00:09
#> 17    4.39 2017-11-28 12:57:41
#> 18    4.28 2017-11-28 12:55:13
#> 19    4.74 2017-11-28 12:52:45
#> 20    4.50 2017-11-28 12:50:18
#> 21    4.26 2017-11-28 12:47:50
#> 22    4.64 2017-11-28 12:45:22
#> 23    5.17 2017-11-28 12:42:54
#> 24    5.10 2017-11-28 12:40:26
#> 25    5.07 2017-11-28 12:37:58
#> 26    5.02 2017-11-28 12:35:30
#> 27    4.52 2017-11-28 12:33:03
#> 28    4.94 2017-11-28 12:30:35
#> 29    5.05 2017-11-28 12:28:07
#> 30    4.98 2017-11-28 12:25:39
#> 31    4.70 2017-11-28 12:23:11
#> 32    4.50 2017-11-28 12:20:43
#> 33    4.58 2017-11-28 12:18:16
#> 34    4.07 2017-11-28 12:15:48
#> 35    3.96 2017-11-28 12:13:20
#> 36    3.79 2017-11-28 12:10:52
#> 37    4.03 2017-11-28 12:08:24
#> 38    4.02 2017-11-28 12:05:57
#> 39    4.32 2017-11-28 12:03:29
#> 40    4.98 2017-11-28 12:01:01
#> 41    5.14 2017-11-28 11:58:33
#> 42    5.09 2017-11-28 11:56:06
#> 43    4.58 2017-11-28 11:53:38
#> 44    4.44 2017-11-28 11:51:10
#> 45    5.45 2017-11-28 11:48:42
#> 46    4.93 2017-11-28 11:46:14
#> 47    4.49 2017-11-28 11:43:47
#> 48    4.23 2017-11-28 11:41:19
#> 49    4.35 2017-11-28 11:38:51
#> 50    4.70 2017-11-28 11:36:23
#> 51    5.28 2017-11-28 11:33:56
#> 52    4.94 2017-11-28 11:31:28
#> 53    4.09 2017-11-28 11:29:00
#> 54    3.64 2017-11-28 11:26:31
#> 55    3.80 2017-11-28 11:24:03
#> 56    3.77 2017-11-28 11:21:36
#> 57    3.81 2017-11-28 11:19:08
#> 58    4.38 2017-11-28 11:16:40
#> 59    5.31 2017-11-28 11:14:12
#> 60    6.09 2017-11-28 11:11:44
#> 61    5.87 2017-11-28 11:09:17
#> 62    5.47 2017-11-28 11:06:49
#> 63    5.28 2017-11-28 11:04:21
#> 64    5.09 2017-11-28 11:01:53
#> 65    4.70 2017-11-28 10:59:25
#> 66    3.94 2017-11-28 10:56:57
#> 67    3.90 2017-11-28 10:54:30
#> 68    4.03 2017-11-28 10:52:02
#> 69    3.93 2017-11-28 10:49:34
#> 70    4.34 2017-11-28 10:47:06
#> 71    4.61 2017-11-28 10:42:11
#> 72    4.14 2017-11-28 10:39:44
#> 73    2.88 2017-11-28 10:37:16
#> 74    3.45 2017-11-28 10:34:48
#> 75    3.33 2017-11-28 10:32:20
#> 76    3.46 2017-11-28 10:29:52
#> 77    3.49 2017-11-28 10:27:24
#> 78    3.00 2017-11-28 10:24:56
#> 79    3.31 2017-11-28 10:22:29
#> 80    4.41 2017-11-28 10:20:01
#> 81    4.15 2017-11-28 10:17:33
#> 82    4.45 2017-11-28 10:15:05
#> 83    3.98 2017-11-28 10:12:37
#> 84    3.20 2017-11-28 10:10:10
#> 85    3.20 2017-11-28 10:07:42
#> 86    2.96 2017-11-28 10:05:14
#> 87    3.24 2017-11-28 10:02:46
#> 88    4.13 2017-11-28 10:00:19
#> 89    3.68 2017-11-28 09:57:50
#> 90    3.50 2017-11-28 09:55:22
#> 91    3.46 2017-11-28 09:52:54
#> 92    3.32 2017-11-28 09:50:27
#> 93    3.16 2017-11-28 09:47:59
#> 94    2.77 2017-11-28 09:45:31
#> 95    2.53 2017-11-28 09:43:03
#> 96    2.14 2017-11-28 09:40:36
#> 97    2.26 2017-11-28 09:38:08
#> 98    2.42 2017-11-28 09:35:40
#> 99    2.47 2017-11-28 09:33:12
#> 100   2.11 2017-11-28 09:30:44
#> 101   1.69 2017-11-28 09:28:17
#> 102   1.76 2017-11-28 09:25:49
#> 103   1.77 2017-11-28 09:23:21
#> 104   1.87 2017-11-28 09:20:53
#> 105   1.77 2017-11-28 09:18:25
#> 106   1.84 2017-11-28 09:15:57
#> 107   1.81 2017-11-28 09:13:30
#> 108   1.43 2017-11-28 09:11:02
#> 109   1.07 2017-11-28 09:08:34
#> 110   0.79 2017-11-28 09:06:06
#> 111   0.47 2017-11-28 09:03:39
#> 112   0.26 2017-11-28 09:01:11
#> 113  -0.06 2017-11-28 08:58:43
#> 114  -0.30 2017-11-28 08:56:15
#> 115  -0.20 2017-11-28 08:53:47
#> 116  -0.49 2017-11-28 08:51:19
#> 117  -0.64 2017-11-28 08:48:51
#> 118  -0.76 2017-11-28 08:46:24
#> 119  -0.87 2017-11-28 08:43:56
#> 120  -1.20 2017-11-28 08:41:28
#> 121  -1.32 2017-11-28 08:39:00
#> 122  -1.35 2017-11-28 08:36:32
#> 123  -1.29 2017-11-28 08:34:04
#> 124  -1.15 2017-11-28 08:31:36
#> 125  -0.83 2017-11-28 08:29:08
#> 126  -1.12 2017-11-28 08:26:41
#> 127  -1.32 2017-11-28 08:24:13
#> 128  -1.07 2017-11-28 08:21:45
#> 129  -0.85 2017-11-28 08:19:17
#> 130  -1.09 2017-11-28 08:16:49
#> 131  -1.32 2017-11-28 08:14:21
#> 132  -1.26 2017-11-28 08:11:54
#> 133  -0.75 2017-11-28 08:09:26
#> 134  -0.66 2017-11-28 08:06:58
#> 135  -0.82 2017-11-28 08:04:30
#> 136  -0.97 2017-11-28 08:02:02
#> 137  -1.11 2017-11-28 07:59:34
#> 138  -0.98 2017-11-28 07:57:06
#> 139  -0.89 2017-11-28 07:54:39
#> 140  -1.19 2017-11-28 07:52:11
#> 141  -1.31 2017-11-28 07:49:43
#> 142  -1.28 2017-11-28 07:47:15
#> 143  -1.37 2017-11-28 07:44:47
#> 144  -1.47 2017-11-28 07:42:20
#> 145  -1.62 2017-11-28 07:39:52
#> 146  -1.59 2017-11-28 07:37:24
#> 147  -1.54 2017-11-28 07:34:56
#> 148  -1.54 2017-11-28 07:32:28
#> 149  -1.58 2017-11-28 07:30:00
#> 150  -1.59 2017-11-28 07:27:32
#> 151  -1.64 2017-11-28 07:25:04
#> 152  -1.68 2017-11-28 07:22:35
#> 153  -1.69 2017-11-28 07:20:07
#> 154  -1.72 2017-11-28 07:17:39
#> 155  -1.71 2017-11-28 07:15:10
#> 156  -1.80 2017-11-28 07:12:42
#> 157  -1.84 2017-11-28 07:10:15
#> 158  -1.83 2017-11-28 07:07:47
#> 159  -1.93 2017-11-28 07:05:19
#> 160  -1.96 2017-11-28 07:02:51
#> 161  -1.95 2017-11-28 07:00:23
#> 162  -2.00 2017-11-28 06:57:55
#> 163  -2.16 2017-11-28 06:55:28
#> 164  -2.35 2017-11-28 06:53:00
#> 165  -2.23 2017-11-28 06:50:32
#> 166  -2.11 2017-11-28 06:48:04
#> 167  -2.22 2017-11-28 06:45:36
#> 168  -2.24 2017-11-28 06:43:07
#> 169  -2.24 2017-11-28 06:40:39
#> 170  -2.16 2017-11-28 06:38:11
#> 171  -2.18 2017-11-28 06:35:44
#> 172  -2.21 2017-11-28 06:33:16
#> 173  -2.13 2017-11-28 06:30:48
#> 174  -2.14 2017-11-28 06:28:20
#> 175  -2.11 2017-11-28 06:25:52
#> 176  -2.21 2017-11-28 06:23:24
#> 177  -2.28 2017-11-28 06:20:56
#> 178  -2.22 2017-11-28 06:18:28
#> 179  -2.28 2017-11-28 06:16:00
#> 180  -2.30 2017-11-28 06:13:32
#> 181  -2.39 2017-11-28 06:11:04
#> 182  -2.39 2017-11-28 06:08:36
#> 183  -2.39 2017-11-28 06:06:08
#> 184  -2.33 2017-11-28 06:03:41
#> 185  -2.41 2017-11-28 06:01:13
#> 186  -2.44 2017-11-28 05:58:44
#> 187  -2.35 2017-11-28 05:56:16
#> 188  -2.37 2017-11-28 05:53:48
#> 189  -2.45 2017-11-28 05:51:21
#> 190  -2.30 2017-11-28 05:48:53
#> 191  -2.23 2017-11-28 05:46:25
#> 192  -2.30 2017-11-28 05:43:57
#> 193  -2.25 2017-11-28 05:41:29
#> 194  -2.19 2017-11-28 05:39:01
#> 195  -2.23 2017-11-28 05:36:33
#> 196  -2.27 2017-11-28 05:34:06
#> 197  -2.28 2017-11-28 05:31:38
#> 198  -2.27 2017-11-28 05:29:10
#> 199  -2.30 2017-11-28 05:26:41
#> 200  -2.44 2017-11-28 05:24:13
#> 201  -2.43 2017-11-28 05:21:45
#> 202  -2.28 2017-11-28 05:19:17
#> 203  -2.14 2017-11-28 05:16:49
#> 204  -2.20 2017-11-28 05:14:21
#> 205  -2.28 2017-11-28 05:11:53
#> 206  -2.17 2017-11-28 05:09:26
#> 207  -2.13 2017-11-28 05:06:58
#> 208  -2.16 2017-11-28 05:04:30
#> 209  -2.20 2017-11-28 05:02:02
#> 210  -2.22 2017-11-28 04:59:34
#> 211  -2.20 2017-11-28 04:57:06
#> 212  -2.20 2017-11-28 04:54:39
#> 213  -2.13 2017-11-28 04:52:11
#> 214  -2.07 2017-11-28 04:49:43
#> 215  -2.13 2017-11-28 04:47:15
#> 216  -2.20 2017-11-28 04:44:47
#> 217  -2.11 2017-11-28 04:42:19
#> 218  -2.14 2017-11-28 04:39:51
#> 219  -2.15 2017-11-28 04:37:24
#> 220  -2.22 2017-11-28 04:34:56
#> 221  -2.18 2017-11-28 04:32:28
#> 222  -2.10 2017-11-28 04:30:00
#> 223  -2.03 2017-11-28 04:27:32
#> 224  -1.98 2017-11-28 04:25:04
#> 225  -1.99 2017-11-28 04:22:37
#> 226  -1.99 2017-11-28 04:20:09
#> 227  -2.04 2017-11-28 04:17:41
#> 228  -1.87 2017-11-28 04:15:13
#> 229  -1.95 2017-11-28 04:12:45
#> 230  -1.99 2017-11-28 04:10:17
#> 231  -2.04 2017-11-28 04:07:49
#> 232  -2.00 2017-11-28 04:05:22
#> 233  -1.99 2017-11-28 04:02:54
#> 234  -1.88 2017-11-28 04:00:26
#> 235  -1.85 2017-11-28 03:57:55
#> 236  -1.76 2017-11-28 03:55:28
#> 237  -1.77 2017-11-28 03:53:00
#> 238  -1.72 2017-11-28 03:50:32
#> 239  -1.68 2017-11-28 03:48:04
#> 240  -1.50 2017-11-28 03:42:38
#> 241  -1.49 2017-11-28 03:40:10
#> 242  -1.41 2017-11-28 03:37:42
#> 243  -1.30 2017-11-28 03:35:14
#> 244  -1.32 2017-11-28 03:32:47
#> 245  -1.37 2017-11-28 03:30:19
#> 246  -1.35 2017-11-28 03:27:40
#> 247  -1.30 2017-11-28 03:21:59
#> 248  -1.38 2017-11-28 03:19:28
#> 249  -1.38 2017-11-28 03:16:58
#> 250  -1.42 2017-11-28 03:11:33
#> 251  -1.39 2017-11-28 03:09:05
#> 252  -1.37 2017-11-28 03:06:36
#> 253  -1.38 2017-11-28 03:04:09
#> 254  -1.36 2017-11-28 03:01:41
#> 255  -1.35 2017-11-28 02:59:06
#> 256  -1.34 2017-11-28 02:56:23
#> 257  -1.38 2017-11-28 02:53:55
#> 258  -1.48 2017-11-28 02:51:27
#> 259  -1.49 2017-11-28 02:48:59
#> 260  -1.47 2017-11-28 02:46:19
#> 261  -1.55 2017-11-28 02:43:50
#> 262  -1.71 2017-11-28 02:41:22
#> 263  -1.66 2017-11-28 02:38:45
#> 264  -1.57 2017-11-28 02:36:18
#> 265  -1.44 2017-11-28 02:33:50
#> 266  -1.48 2017-11-28 02:31:22
#> 267  -1.49 2017-11-28 02:28:54
#> 268  -1.46 2017-11-28 02:26:26
#> 269  -1.51 2017-11-28 02:23:58
#> 270  -1.53 2017-11-28 02:21:31
#> 271  -1.49 2017-11-28 02:19:03
#> 272  -1.41 2017-11-28 02:16:35
#> 273  -1.44 2017-11-28 02:14:07
#> 274  -1.47 2017-11-28 02:11:39
#> 275  -1.49 2017-11-28 02:09:12
#> 276  -1.46 2017-11-28 02:06:44
#> 277  -1.52 2017-11-28 02:04:16
#> 278  -1.52 2017-11-28 02:01:48
#> 279  -1.53 2017-11-28 01:59:20
#> 280  -1.53 2017-11-28 01:56:52
#> 281  -1.57 2017-11-28 01:54:25
#> 282  -1.61 2017-11-28 01:51:57
#> 283  -1.67 2017-11-28 01:49:29
#> 284  -1.67 2017-11-28 01:47:01
#> 285  -1.71 2017-11-28 01:44:33
#> 286  -1.75 2017-11-28 01:42:06
#> 287  -1.72 2017-11-28 01:39:38
#> 288  -1.71 2017-11-28 01:37:10
#> 289  -1.62 2017-11-28 01:34:42
#> 290  -1.61 2017-11-28 01:32:15
#> 291  -1.61 2017-11-28 01:29:47
#> 292  -1.62 2017-11-28 01:27:19
#> 293  -1.52 2017-11-28 01:24:51
#> 294  -1.53 2017-11-28 01:22:23
#> 295  -1.53 2017-11-28 01:19:56
#> 296  -1.53 2017-11-28 01:17:28
#> 297  -1.58 2017-11-28 01:15:00
#> 298  -1.53 2017-11-28 01:12:32
#> 299  -1.57 2017-11-28 01:10:04
#> 300  -1.59 2017-11-28 01:07:37
#> 301  -1.57 2017-11-28 01:05:09
#> 302  -1.58 2017-11-28 01:02:41
#> 303  -1.59 2017-11-28 01:00:13
#> 304  -1.64 2017-11-28 00:57:46
#> 305  -1.65 2017-11-28 00:55:18
#> 306  -1.67 2017-11-28 00:52:50
#> 307  -1.69 2017-11-28 00:50:22
#> 308  -1.70 2017-11-28 00:47:55
#> 309  -1.74 2017-11-28 00:45:27
#> 310  -1.77 2017-11-28 00:42:59
#> 311  -1.85 2017-11-28 00:40:31
#> 312  -1.94 2017-11-28 00:38:03
#> 313  -1.94 2017-11-28 00:35:36
#> 314  -1.82 2017-11-28 00:33:08
#> 315  -1.70 2017-11-28 00:30:40
#> 316  -1.71 2017-11-28 00:28:12
#> 317  -1.67 2017-11-28 00:25:45
#> 318  -1.67 2017-11-28 00:23:17
#> 319  -1.71 2017-11-28 00:20:49
#> 320  -1.76 2017-11-28 00:18:21
#> 321  -1.79 2017-11-28 00:15:53
#> 322  -1.85 2017-11-28 00:13:26
#> 323  -1.92 2017-11-28 00:10:58
#> 324  -2.00 2017-11-28 00:08:30
#> 325  -2.03 2017-11-28 00:06:02
#> 326  -2.08 2017-11-28 00:03:34
#> 327  -2.08 2017-11-28 00:01:07
#> 328  -2.21 2017-11-27 23:58:38
#> 329  -2.36 2017-11-27 23:56:10
#> 330  -2.36 2017-11-27 23:53:42
#> 331  -2.41 2017-11-27 23:51:14
#> 332  -2.40 2017-11-27 23:48:47
#> 333  -2.48 2017-11-27 23:46:19
#> 334  -2.50 2017-11-27 23:43:51
#> 335  -2.60 2017-11-27 23:41:23
#> 336  -2.70 2017-11-27 23:38:55
#> 337  -2.81 2017-11-27 23:36:28
#> 338  -2.84 2017-11-27 23:34:00
#> 339  -2.90 2017-11-27 23:31:32
#> 340  -2.97 2017-11-27 23:29:04
#> 341  -2.92 2017-11-27 23:26:37
#> 342  -2.81 2017-11-27 23:24:09
#> 343  -2.54 2017-11-27 23:21:41
#> 344  -2.34 2017-11-27 23:19:13
#> 345  -2.27 2017-11-27 23:16:45
#> 346  -2.43 2017-11-27 23:14:18
#> 347  -2.55 2017-11-27 23:11:50
#> 348  -2.60 2017-11-27 23:09:22
#> 349  -2.60 2017-11-27 23:06:54
#> 350  -2.59 2017-11-27 23:04:26
#> 351  -2.70 2017-11-27 23:01:59
#> 352  -2.54 2017-11-27 22:59:31
#> 353  -2.55 2017-11-27 22:57:03
#> 354  -2.61 2017-11-27 22:54:35
#> 355  -2.71 2017-11-27 22:52:07
#> 356  -2.80 2017-11-27 22:49:40
#> 357  -2.83 2017-11-27 22:47:12
#> 358  -2.90 2017-11-27 22:44:44
#> 359  -2.90 2017-11-27 22:42:16
#> 360  -2.93 2017-11-27 22:39:48
#> 361  -2.91 2017-11-27 22:37:21
#> 362  -2.93 2017-11-27 22:34:52
#> 363  -2.93 2017-11-27 22:32:25
#> 364  -2.94 2017-11-27 22:29:57
#> 365  -2.85 2017-11-27 22:27:29
#> 366  -2.90 2017-11-27 22:25:01
#> 367  -2.88 2017-11-27 22:22:33
#> 368  -2.96 2017-11-27 22:20:06
#> 369  -3.04 2017-11-27 22:17:38
#> 370  -3.20 2017-11-27 22:15:10
#> 371  -3.12 2017-11-27 22:12:42
#> 372  -2.86 2017-11-27 22:10:14
#> 373  -2.96 2017-11-27 22:07:46
#> 374  -2.96 2017-11-27 22:05:19
#> 375  -2.99 2017-11-27 22:02:51
#> 376  -2.98 2017-11-27 22:00:23
#> 377  -2.88 2017-11-27 21:57:55
#> 378  -2.80 2017-11-27 21:55:28
#> 379  -2.72 2017-11-27 21:53:00
#> 380  -2.72 2017-11-27 21:50:32
#> 381  -2.67 2017-11-27 21:48:04
#> 382  -2.68 2017-11-27 21:45:36
#> 383  -2.74 2017-11-27 21:43:08
#> 384  -2.69 2017-11-27 21:40:40
#> 385  -2.73 2017-11-27 21:38:12
#> 386  -2.77 2017-11-27 21:35:44
#> 387  -2.79 2017-11-27 21:33:16
#> 388  -2.76 2017-11-27 21:30:49
#> 389  -2.65 2017-11-27 21:28:21
#> 390  -2.59 2017-11-27 21:25:53
#> 391  -2.64 2017-11-27 21:23:25
#> 392  -2.62 2017-11-27 21:20:57
#> 393  -2.78 2017-11-27 21:18:30
#> 394  -2.63 2017-11-27 21:16:02
#> 395  -2.66 2017-11-27 21:13:33
#> 396  -2.54 2017-11-27 21:11:05
#> 397  -2.49 2017-11-27 21:08:37
#> 398  -2.49 2017-11-27 21:06:09
#> 399  -2.44 2017-11-27 21:03:41
#> 400  -2.50 2017-11-27 21:01:14
#> 401  -2.42 2017-11-27 20:58:46
#> 402  -2.34 2017-11-27 20:56:18
#> 403  -2.37 2017-11-27 20:53:50
#> 404  -2.39 2017-11-27 20:51:22
#> 405  -2.52 2017-11-27 20:48:54
#> 406  -2.48 2017-11-27 20:46:26
#> 407  -2.07 2017-11-27 20:43:59
#> 408  -2.01 2017-11-27 20:41:31
#> 409  -2.00 2017-11-27 20:39:02
#> 410  -1.97 2017-11-27 20:36:34
#> 411  -1.96 2017-11-27 20:34:06
#> 412  -2.01 2017-11-27 20:31:38
#> 413  -1.87 2017-11-27 20:29:10
#> 414  -1.84 2017-11-27 20:26:41
#> 415  -1.84 2017-11-27 20:24:13
#> 416  -1.81 2017-11-27 20:21:45
#> 417  -1.90 2017-11-27 20:19:16
#> 418  -1.97 2017-11-27 20:16:48
#> 419  -2.01 2017-11-27 20:14:20
#> 420  -1.93 2017-11-27 20:11:52
#> 421  -1.83 2017-11-27 20:09:24
#> 422  -1.93 2017-11-27 20:06:56
#> 423  -1.98 2017-11-27 20:04:27
#> 424  -1.89 2017-11-27 20:01:59
#> 425  -1.81 2017-11-27 19:59:31
#> 426  -1.80 2017-11-27 19:57:03
#> 427  -1.74 2017-11-27 19:54:35
#> 428  -1.70 2017-11-27 19:52:07
#> 429  -1.72 2017-11-27 19:49:39
#> 430  -1.59 2017-11-27 19:47:11
#> 431  -1.56 2017-11-27 19:44:42
#> 432  -1.60 2017-11-27 19:42:12
#> 433  -1.54 2017-11-27 19:39:42
#> 434  -1.38 2017-11-27 19:37:10
#> 435  -1.26 2017-11-27 19:34:41
#> 436  -1.22 2017-11-27 19:32:11
#> 437  -1.18 2017-11-27 19:29:42
#> 438  -1.14 2017-11-27 19:27:13
#> 439  -1.11 2017-11-27 19:24:43
#> 440  -1.06 2017-11-27 19:22:14
#> 441  -1.03 2017-11-27 19:19:45
#> 442  -1.03 2017-11-27 19:17:15
#> 443  -1.00 2017-11-27 19:14:46
#> 444  -0.93 2017-11-27 19:12:18
#> 445  -0.94 2017-11-27 19:09:50
#> 446  -0.96 2017-11-27 19:07:21
#> 447  -0.94 2017-11-27 19:04:54
#> 448  -0.83 2017-11-27 19:02:26
#> 449  -0.89 2017-11-27 18:59:58
#> 450  -0.89 2017-11-27 18:57:30
#> 451  -0.84 2017-11-27 18:55:02
#> 452  -0.78 2017-11-27 18:52:34
#> 453  -0.73 2017-11-27 18:50:05
#> 454  -0.71 2017-11-27 18:47:37
#> 455  -0.72 2017-11-27 18:45:09
#> 456  -0.84 2017-11-27 18:42:40
#> 457  -0.77 2017-11-27 18:40:13
#> 458  -0.78 2017-11-27 18:37:45
#> 459  -0.75 2017-11-27 18:35:17
#> 460  -0.66 2017-11-27 18:32:49
#> 461  -0.57 2017-11-27 18:30:21
#> 462  -0.37 2017-11-27 18:27:53
#> 463  -0.39 2017-11-27 18:25:25
#> 464  -0.47 2017-11-27 18:22:57
#> 465  -0.56 2017-11-27 18:20:29
#> 466  -0.40 2017-11-27 18:18:01
#> 467  -0.35 2017-11-27 18:15:33
#> 468  -0.30 2017-11-27 18:13:05
#> 469  -0.46 2017-11-27 18:10:37
#> 470  -0.42 2017-11-27 18:08:09
#> 471  -0.48 2017-11-27 18:05:41
#> 472  -0.39 2017-11-27 18:03:13
#> 473  -0.50 2017-11-27 18:00:45
#> 474  -0.47 2017-11-27 17:58:18
#> 475  -0.46 2017-11-27 17:55:50
#> 476  -0.35 2017-11-27 17:53:22
#> 477  -0.43 2017-11-27 17:50:54
#> 478  -0.24 2017-11-27 17:48:26
#> 479  -0.24 2017-11-27 17:45:58
#> 480  -0.17 2017-11-27 17:43:30
#> 481  -0.07 2017-11-27 17:41:03
#> 482  -0.21 2017-11-27 17:38:32
#> 483  -0.11 2017-11-27 17:36:05
#> 484  -0.18 2017-11-27 17:33:36
#> 485  -0.22 2017-11-27 17:31:08
#> 486  -0.09 2017-11-27 17:28:40
#> 487  -0.08 2017-11-27 17:26:12
#> 488   0.08 2017-11-27 17:23:44
#> 489   0.07 2017-11-27 17:21:17
#> 490   0.04 2017-11-27 17:18:49
#> 491   0.25 2017-11-27 17:16:21
#> 492   0.25 2017-11-27 17:13:53
#> 493   0.26 2017-11-27 17:11:25
#> 494   0.37 2017-11-27 17:08:58
#> 495   0.44 2017-11-27 17:06:30
#> 496   0.36 2017-11-27 17:04:02
#> 497   0.40 2017-11-27 17:01:34
#> 498   0.62 2017-11-27 16:59:06
#> 499   0.68 2017-11-27 16:56:37
#> 500   0.61 2017-11-27 16:54:09
#> 501   0.68 2017-11-27 16:51:42
#> 502   0.81 2017-11-27 16:49:14
#> 503   0.86 2017-11-27 16:46:46
#> 504   0.80 2017-11-27 16:44:18
#> 505   0.86 2017-11-27 16:41:50
#> 506   0.84 2017-11-27 16:39:21
#> 507   0.94 2017-11-27 16:36:53
#> 508   1.20 2017-11-27 16:34:25
#> 509   1.23 2017-11-27 16:31:57
#> 510   1.27 2017-11-27 16:29:30
#> 511   1.27 2017-11-27 16:27:02
#> 512   1.37 2017-11-27 16:24:34
#> 513   1.39 2017-11-27 16:22:06
#> 514   1.47 2017-11-27 16:19:38
#> 515   1.47 2017-11-27 16:17:11
#> 516   1.51 2017-11-27 16:14:43
#> 517   1.51 2017-11-27 16:12:15
#> 518   1.54 2017-11-27 16:09:47
#> 519   1.57 2017-11-27 16:07:19
#> 520   1.55 2017-11-27 16:04:52
#> 521   1.53 2017-11-27 16:02:24
#> 522   1.55 2017-11-27 15:59:56
#> 523   1.49 2017-11-27 15:57:27
#> 524   1.45 2017-11-27 15:54:59
#> 525   1.46 2017-11-27 15:52:31
#> 526   1.49 2017-11-27 15:50:03
#> 527   1.54 2017-11-27 15:47:35
#> 528   1.54 2017-11-27 15:45:07
#> 529   1.59 2017-11-27 15:42:39
#> 530   1.67 2017-11-27 15:40:10
#> 531   1.49 2017-11-27 15:37:42
#> 532   1.85 2017-11-27 15:35:14
#> 533   1.98 2017-11-27 15:32:46
#> 534   1.97 2017-11-27 15:30:18
#> 535   2.09 2017-11-27 15:27:50
#> 536   2.13 2017-11-27 15:25:21
#> 537   2.11 2017-11-27 15:22:53
#> 538   2.05 2017-11-27 15:20:26
#> 539   2.08 2017-11-27 15:17:58
#> 540   2.12 2017-11-27 15:15:30
#> 541   2.08 2017-11-27 15:13:02
#> 542   1.98 2017-11-27 15:10:34
#> 543   2.07 2017-11-27 15:08:06
#> 544   1.98 2017-11-27 15:05:38
#> 545   2.18 2017-11-27 15:03:11
#> 546   2.28 2017-11-27 15:00:43
#> 547   2.32 2017-11-27 14:58:15
#> 548   2.43 2017-11-27 14:55:47
#> 549   2.51 2017-11-27 14:53:19
#> 550   2.69 2017-11-27 14:50:52
#> 551   2.58 2017-11-27 14:48:24
#> 552   2.64 2017-11-27 14:45:56
#> 553   2.78 2017-11-27 14:43:28
#> 554   2.83 2017-11-27 14:41:00
#> 555   2.73 2017-11-27 14:38:32
#> 556   2.79 2017-11-27 14:36:05
#> 557   2.83 2017-11-27 14:33:37
#> 558   3.18 2017-11-27 14:31:08
#> 559   3.26 2017-11-27 14:28:40
#> 560   3.55 2017-11-27 14:26:13
#> 561   3.53 2017-11-27 14:23:44
#> 562   3.61 2017-11-27 14:21:16
#> 563   3.43 2017-11-27 14:18:48
#> 564   3.53 2017-11-27 14:16:20
#> 565   3.70 2017-11-27 14:13:53
#> 566   3.89 2017-11-27 14:11:25
#> 567   4.14 2017-11-27 14:08:57
#> 568   3.93 2017-11-27 14:06:29
#> 569   4.02 2017-11-27 14:04:01
#> 570   3.93 2017-11-27 14:01:34
#> 571   4.03 2017-11-27 13:59:06
#> 572   4.14 2017-11-27 13:56:38
#> 573   4.26 2017-11-27 13:54:10
#> 574   4.25 2017-11-27 13:51:42
#> 575   4.39 2017-11-27 13:49:14
#> 576   4.23 2017-11-27 13:46:46
#> 577   4.17 2017-11-27 13:44:18
#> 578   4.28 2017-11-27 13:41:50
#> 579   4.17 2017-11-27 13:39:22
#> 580   4.13 2017-11-27 13:36:55
#> 581   4.24 2017-11-27 13:34:27
#> 582   4.16 2017-11-27 13:31:59
#> 583   4.38 2017-11-27 13:29:31
#> 584   4.69 2017-11-27 13:27:03
#> 585   4.64 2017-11-27 13:24:36
#> 586   4.72 2017-11-27 13:22:08
#> 587   4.70 2017-11-27 13:19:40
#> 588   4.88 2017-11-27 13:17:12
#> 589   4.85 2017-11-27 13:14:44
#> 590   5.20 2017-11-27 13:12:16
#> 591   5.02 2017-11-27 13:09:49
#> 592   4.81 2017-11-27 13:07:20
#> 593   4.72 2017-11-27 13:04:52
#> 594   4.63 2017-11-27 13:02:24
#> 595   4.53 2017-11-27 12:59:56
#> 596   4.57 2017-11-27 12:57:29
#> 597   4.53 2017-11-27 12:55:01
#> 598   4.38 2017-11-27 12:52:33
#> 599   4.16 2017-11-27 12:50:03
#> 600   4.40 2017-11-27 12:47:36
#> 601   4.39 2017-11-27 12:45:08
#> 602   4.34 2017-11-27 12:42:40
#> 603   4.58 2017-11-27 12:40:12
#> 604   4.49 2017-11-27 12:37:44
#> 605   4.46 2017-11-27 12:35:17
#> 606   4.53 2017-11-27 12:32:49
#> 607   4.38 2017-11-27 12:30:21
#> 608   4.26 2017-11-27 12:27:53
#> 609   4.24 2017-11-27 12:25:26
#> 610   4.91 2017-11-27 12:22:58
#> 611   4.91 2017-11-27 12:20:30
#> 612   4.99 2017-11-27 12:18:02
#> 613   4.85 2017-11-27 12:15:34
#> 614   4.91 2017-11-27 12:13:07
#> 615   4.82 2017-11-27 12:10:39
#> 616   4.67 2017-11-27 12:08:11
#> 617   4.93 2017-11-27 12:05:43
#> 618   4.68 2017-11-27 12:03:16
#> 619   4.41 2017-11-27 12:00:48
#> 620   4.41 2017-11-27 11:58:20
#> 621   3.93 2017-11-27 11:55:52
#> 622   3.98 2017-11-27 11:53:23
#> 623   3.79 2017-11-27 11:50:55
#> 624   4.32 2017-11-27 11:48:28
#> 625   4.65 2017-11-27 11:46:00
#> 626   4.42 2017-11-27 11:43:32
#> 627   4.58 2017-11-27 11:41:04
#> 628   4.53 2017-11-27 11:38:37
#> 629   4.97 2017-11-27 11:36:09
#> 630   4.74 2017-11-27 11:33:41
#> 631   4.89 2017-11-27 11:31:13
#> 632   4.63 2017-11-27 11:28:46
#> 633   4.22 2017-11-27 11:26:18
#> 634   4.57 2017-11-27 11:23:50
#> 635   4.61 2017-11-27 11:21:23
#> 636   4.74 2017-11-27 11:18:55
#> 637   5.29 2017-11-27 11:16:27
#> 638   5.59 2017-11-27 11:13:59
#> 639   6.38 2017-11-27 11:11:31
#> 640   5.41 2017-11-27 11:09:04
#> 641   6.11 2017-11-27 11:06:36
#> 642   6.13 2017-11-27 11:04:08
#> 643   7.41 2017-11-27 11:01:40
#> 644   7.52 2017-11-27 10:59:13
#> 645   7.04 2017-11-27 10:56:45
#> 646   7.62 2017-11-27 10:54:17
#> 647   7.52 2017-11-27 10:51:49
#> 648   7.30 2017-11-27 10:49:22
#> 649   7.00 2017-11-27 10:46:54
#> 650   6.31 2017-11-27 10:44:26
#> 651   5.88 2017-11-27 10:41:58
#> 652   5.14 2017-11-27 10:39:31
#> 653   4.67 2017-11-27 10:37:03
#> 654   4.35 2017-11-27 10:34:35
#> 655   4.35 2017-11-27 10:32:07
#> 656   4.17 2017-11-27 10:29:39
#> 657   4.22 2017-11-27 10:27:12
#> 658   4.33 2017-11-27 10:24:44
#> 659   4.25 2017-11-27 10:22:16
#> 660   4.35 2017-11-27 10:19:48
#> 661   4.33 2017-11-27 10:17:21
#> 662   4.27 2017-11-27 10:14:53
#> 663   4.28 2017-11-27 10:12:25
#> 664   4.33 2017-11-27 10:09:57
#> 665   4.57 2017-11-27 10:07:29
#> 666   4.45 2017-11-27 10:05:01
#> 667   4.20 2017-11-27 10:02:33
#> 668   3.97 2017-11-27 10:00:05
#> 669   4.03 2017-11-27 09:57:38
#> 670   4.59 2017-11-27 09:55:10
#> 671   5.04 2017-11-27 09:52:40
#> 672   4.63 2017-11-27 09:50:12
#> 673   4.25 2017-11-27 09:47:44
#> 674   3.44 2017-11-27 09:45:16
#> 675   3.31 2017-11-27 09:42:48
#> 676   3.59 2017-11-27 09:40:21
#> 677   3.90 2017-11-27 09:37:53
#> 678   3.52 2017-11-27 09:35:25
#> 679   3.18 2017-11-27 09:32:57
#> 680   3.07 2017-11-27 09:30:30
#> 681   2.83 2017-11-27 09:28:02
#> 682   2.79 2017-11-27 09:25:34
#> 683   2.85 2017-11-27 09:23:06
#> 684   2.54 2017-11-27 09:20:39
#> 685   2.70 2017-11-27 09:18:11
#> 686   2.33 2017-11-27 09:15:43
#> 687   2.70 2017-11-27 09:13:15
#> 688   2.46 2017-11-27 09:10:47
#> 689   2.70 2017-11-27 09:08:20
#> 690   2.83 2017-11-27 09:05:52
#> 691   2.82 2017-11-27 09:03:24
#> 692   2.82 2017-11-27 09:00:56
#> 693   2.75 2017-11-27 08:58:29
#> 694   3.10 2017-11-27 08:56:01
#> 695   2.86 2017-11-27 08:53:32
#> 696   2.69 2017-11-27 08:51:04
#> 697   2.41 2017-11-27 08:48:36
#> 698   1.87 2017-11-27 08:46:09
#> 699   1.64 2017-11-27 08:43:41
#> 700   1.52 2017-11-27 08:41:13
#> 701   1.72 2017-11-27 08:38:45
#> 702   2.18 2017-11-27 08:36:16
#> 703   2.43 2017-11-27 08:33:49
#> 704   2.36 2017-11-27 08:31:21
#> 705   2.07 2017-11-27 08:28:53
#> 706   2.12 2017-11-27 08:26:25
#> 707   2.05 2017-11-27 08:23:58
#> 708   2.00 2017-11-27 08:21:30
#> 709   1.92 2017-11-27 08:19:02
#> 710   1.83 2017-11-27 08:16:34
#> 711   1.91 2017-11-27 08:14:06
#> 712   1.99 2017-11-27 08:11:38
#> 713   2.06 2017-11-27 08:09:11
#> 714   1.47 2017-11-27 08:06:43
#> 715   1.35 2017-11-27 08:04:15
#> 716   1.20 2017-11-27 08:01:47
#> 717   1.16 2017-11-27 07:59:20
#> 718   1.12 2017-11-27 07:56:52
#> 719   1.09 2017-11-27 07:54:24
#> 720   1.21 2017-11-27 07:51:56
#> 721   0.85 2017-11-27 07:49:29
#> 722   0.72 2017-11-27 07:47:01
#> 723   0.87 2017-11-27 07:44:33
#> 724   0.96 2017-11-27 07:42:05
#> 725   1.17 2017-11-27 07:39:38
#> 726   1.06 2017-11-27 07:37:04
#> 727   1.01 2017-11-27 07:34:36
#> 728   0.89 2017-11-27 07:32:08
#> 729   0.86 2017-11-27 07:29:40
#> 730   0.80 2017-11-27 07:27:13
#> 731   0.75 2017-11-27 07:24:45
#> 732   0.73 2017-11-27 07:22:17
#> 733   0.64 2017-11-27 07:19:50
#> 734   0.58 2017-11-27 07:17:22
#> 735   0.61 2017-11-27 07:14:54
#> 736   0.60 2017-11-27 07:12:26
#> 737   0.51 2017-11-27 07:09:58
#> 738   0.47 2017-11-27 07:07:31
#> 739   0.42 2017-11-27 07:05:02
#> 740   0.34 2017-11-27 07:02:34
#> 741   0.37 2017-11-27 07:00:06
#> 742   0.25 2017-11-27 06:57:39
#> 743   0.19 2017-11-27 06:55:10
#> 744   0.21 2017-11-27 06:52:42
#> 745   0.30 2017-11-27 06:50:14
#> 746   0.28 2017-11-27 06:47:46
#> 747   0.22 2017-11-27 06:45:18
#> 748   0.25 2017-11-27 06:42:50
#> 749   0.34 2017-11-27 06:40:22
#> 750   0.26 2017-11-27 06:37:55
#> 751   0.24 2017-11-27 06:35:27
#> 752   0.24 2017-11-27 06:32:59
#> 753   0.24 2017-11-27 06:30:31
#> 754   0.27 2017-11-27 06:28:03
#> 755   0.26 2017-11-27 06:25:35
#> 756   0.20 2017-11-27 06:23:08
#> 757   0.17 2017-11-27 06:20:40
#> 758   0.20 2017-11-27 06:18:12
#> 759   0.19 2017-11-27 06:15:44
#> 760   0.21 2017-11-27 06:13:17
#> 761   0.20 2017-11-27 06:10:49
#> 762   0.32 2017-11-27 06:08:21
#> 763   0.30 2017-11-27 06:05:54
#> 764   0.24 2017-11-27 06:03:26
#> 765   0.30 2017-11-27 06:00:58
#> 766   0.21 2017-11-27 05:58:30
#> 767   0.20 2017-11-27 05:56:03
#> 768   0.30 2017-11-27 05:53:35
#> 769   0.27 2017-11-27 05:51:07
#> 770   0.29 2017-11-27 05:48:39
#> 771   0.27 2017-11-27 05:46:12
#> 772   0.30 2017-11-27 05:43:44
#> 773   0.23 2017-11-27 05:41:16
#> 774   0.39 2017-11-27 05:38:49
#> 775   0.30 2017-11-27 05:36:21
#> 776   0.31 2017-11-27 05:33:53
#> 777   0.31 2017-11-27 05:31:25
#> 778   0.36 2017-11-27 05:28:58
#> 779   0.40 2017-11-27 05:26:30
#> 780   0.34 2017-11-27 05:24:02
#> 781   0.32 2017-11-27 05:21:34
#> 782   0.35 2017-11-27 05:19:07
#> 783   0.39 2017-11-27 05:16:39
#> 784   0.36 2017-11-27 05:14:11
#> 785   0.27 2017-11-27 05:11:43
#> 786   0.27 2017-11-27 05:09:16
#> 787   0.28 2017-11-27 05:06:48
#> 788   0.26 2017-11-27 05:04:20
#> 789   0.23 2017-11-27 05:01:52
#> 790   0.21 2017-11-27 04:59:25
#> 791   0.20 2017-11-27 04:56:57
#> 792   0.28 2017-11-27 04:54:29
#> 793   0.26 2017-11-27 04:52:01
#> 794   0.24 2017-11-27 04:49:34
#> 795   0.30 2017-11-27 04:47:06
#> 796   0.31 2017-11-27 04:44:38
#> 797   0.32 2017-11-27 04:42:10
#> 798   0.35 2017-11-27 04:39:43
#> 799   0.36 2017-11-27 04:37:15
#> 800   0.34 2017-11-27 04:34:47
#> 801   0.34 2017-11-27 04:32:19
#> 802   0.29 2017-11-27 04:29:52
#> 803   0.30 2017-11-27 04:27:24
#> 804   0.31 2017-11-27 04:24:56
#> 805   0.30 2017-11-27 04:22:28
#> 806   0.29 2017-11-27 04:20:01
#> 807   0.22 2017-11-27 04:17:33
#> 808   0.22 2017-11-27 04:15:05
#> 809   0.30 2017-11-27 04:12:37
#> 810   0.34 2017-11-27 04:10:10
#> 811   0.44 2017-11-27 04:07:42
#> 812   0.45 2017-11-27 04:05:14
#> 813   0.41 2017-11-27 04:02:46
#> 814   0.40 2017-11-27 04:00:19
#> 815   0.41 2017-11-27 03:57:51
#> 816   0.41 2017-11-27 03:55:23
#> 817   0.33 2017-11-27 03:52:55
#> 818   0.27 2017-11-27 03:50:28
#> 819   0.21 2017-11-27 03:48:00
#> 820   0.18 2017-11-27 03:45:32
#> 821   0.23 2017-11-27 03:43:04
#> 822   0.14 2017-11-27 03:40:37
#> 823   0.10 2017-11-27 03:38:09
#> 824  -0.02 2017-11-27 03:35:41
#> 825   0.09 2017-11-27 03:33:13
#> 826   0.06 2017-11-27 03:30:46
#> 827   0.09 2017-11-27 03:28:18
#> 828   0.19 2017-11-27 03:25:50
#> 829   0.19 2017-11-27 03:23:22
#> 830   0.18 2017-11-27 03:20:55
#> 831   0.30 2017-11-27 03:18:27
#> 832   0.32 2017-11-27 03:15:59
#> 833   0.31 2017-11-27 03:13:31
#> 834   0.34 2017-11-27 03:11:04
#> 835   0.39 2017-11-27 03:08:36
#> 836   0.48 2017-11-27 03:06:08
#> 837   0.25 2017-11-27 03:03:40
#> 838   0.24 2017-11-27 03:01:13
#> 839   0.22 2017-11-27 02:58:45
#> 840   0.23 2017-11-27 02:56:16
#> 841   0.27 2017-11-27 02:53:48
#> 842   0.24 2017-11-27 02:51:20
#> 843   0.20 2017-11-27 02:48:53
#> 844   0.22 2017-11-27 02:46:25
#> 845   0.29 2017-11-27 02:43:57
#> 846   0.26 2017-11-27 02:41:29
#> 847   0.29 2017-11-27 02:39:01
#> 848   0.27 2017-11-27 02:36:34
#> 849   0.21 2017-11-27 02:34:06
#> 850   0.26 2017-11-27 02:31:38
#> 851   0.26 2017-11-27 02:29:10
#> 852   0.25 2017-11-27 02:26:42
#> 853   0.24 2017-11-27 02:24:14
#> 854   0.28 2017-11-27 02:21:47
#> 855   0.27 2017-11-27 02:19:19
#> 856   0.28 2017-11-27 02:16:51
#> 857   0.27 2017-11-27 02:14:23
#> 858   0.24 2017-11-27 02:11:56
#> 859   0.22 2017-11-27 02:09:28
#> 860   0.24 2017-11-27 02:07:00
#> 861   0.20 2017-11-27 02:04:32
#> 862   0.17 2017-11-27 02:02:05
#> 863   0.20 2017-11-27 01:59:37
#> 864   0.09 2017-11-27 01:57:09
#> 865   0.06 2017-11-27 01:54:41
#> 866   0.09 2017-11-27 01:52:13
#> 867   0.17 2017-11-27 01:49:45
#> 868   0.23 2017-11-27 01:47:18
#> 869   0.31 2017-11-27 01:44:50
#> 870   0.26 2017-11-27 01:42:22
#> 871   0.26 2017-11-27 01:39:54
#> 872   0.21 2017-11-27 01:37:27
#> 873   0.14 2017-11-27 01:34:59
#> 874   0.28 2017-11-27 01:32:31
#> 875   0.33 2017-11-27 01:30:03
#> 876   0.28 2017-11-27 01:27:36
#> 877   0.23 2017-11-27 01:25:08
#> 878   0.18 2017-11-27 01:22:40
#> 879   0.31 2017-11-27 01:20:13
#> 880   0.34 2017-11-27 01:17:45
#> 881   0.40 2017-11-27 01:15:17
#> 882   0.41 2017-11-27 01:12:49
#> 883   0.36 2017-11-27 01:10:21
#> 884   0.44 2017-11-27 01:07:54
#> 885   0.42 2017-11-27 01:05:26
#> 886   0.45 2017-11-27 01:02:58
#> 887   0.37 2017-11-27 01:00:31
#> 888   0.23 2017-11-27 00:58:03
#> 889   0.09 2017-11-27 00:55:35
#> 890   0.10 2017-11-27 00:53:07
#> 891   0.16 2017-11-27 00:50:39
#> 892   0.23 2017-11-27 00:48:11
#> 893   0.23 2017-11-27 00:45:43
#> 894   0.12 2017-11-27 00:43:16
#> 895   0.09 2017-11-27 00:40:48
#> 896   0.04 2017-11-27 00:38:20
#> 897   0.08 2017-11-27 00:35:52
#> 898   0.26 2017-11-27 00:33:25
#> 899   0.21 2017-11-27 00:30:57
#> 900   0.25 2017-11-27 00:28:29
#> 901   0.24 2017-11-27 00:26:01
#> 902   0.22 2017-11-27 00:23:33
#> 903   0.21 2017-11-27 00:21:06
#> 904   0.25 2017-11-27 00:18:38
#> 905   0.25 2017-11-27 00:16:10
#> 906   0.20 2017-11-27 00:13:42
#> 907   0.18 2017-11-27 00:11:15
#> 908   0.19 2017-11-27 00:08:47
#> 909   0.23 2017-11-27 00:06:19
#> 910   0.11 2017-11-27 00:03:51
#> 911   0.19 2017-11-27 00:01:23
#> 912   0.20 2017-11-26 23:58:56
#> 913   0.24 2017-11-26 23:56:28
#> 914   0.28 2017-11-26 23:54:00
#> 915   0.27 2017-11-26 23:51:32
#> 916   0.40 2017-11-26 23:49:05
#> 917   0.36 2017-11-26 23:46:37
#> 918   0.40 2017-11-26 23:44:09
#> 919   0.36 2017-11-26 23:41:41
#> 920   0.37 2017-11-26 23:39:13
#> 921   0.37 2017-11-26 23:36:46
#> 922   0.35 2017-11-26 23:34:18
#> 923   0.26 2017-11-26 23:31:50
#> 924   0.30 2017-11-26 23:29:22
#> 925   0.34 2017-11-26 23:26:55
#> 926   0.37 2017-11-26 23:24:27
#> 927   0.36 2017-11-26 23:21:59
#> 928   0.30 2017-11-26 23:19:31
#> 929   0.29 2017-11-26 23:17:03
#> 930   0.35 2017-11-26 23:14:36
#> 931   0.38 2017-11-26 23:12:08
#> 932   0.44 2017-11-26 23:09:40
#> 933   0.63 2017-11-26 23:07:12
#> 934   0.64 2017-11-26 23:04:45
#> 935   0.58 2017-11-26 23:02:17
#> 936   0.46 2017-11-26 22:59:49
#> 937   0.48 2017-11-26 22:57:21
#> 938   0.52 2017-11-26 22:54:54
#> 939   0.51 2017-11-26 22:52:26
#> 940   0.51 2017-11-26 22:49:57
#> 941   0.50 2017-11-26 22:47:29
#> 942   0.56 2017-11-26 22:45:01
#> 943   0.50 2017-11-26 22:42:32
#> 944   0.51 2017-11-26 22:40:04
#> 945   0.49 2017-11-26 22:37:37
#> 946   0.48 2017-11-26 22:35:09
#> 947   0.42 2017-11-26 22:32:41
#> 948   0.39 2017-11-26 22:30:13
#> 949   0.45 2017-11-26 22:27:46
#> 950   0.45 2017-11-26 22:25:18
#> 951   0.48 2017-11-26 22:22:50
#> 952   0.32 2017-11-26 22:20:22
#> 953   0.30 2017-11-26 22:17:54
#> 954   0.30 2017-11-26 22:15:27
#> 955   0.42 2017-11-26 22:12:59
#> 956   0.24 2017-11-26 22:10:31
#> 957   0.43 2017-11-26 22:08:03
#> 958   0.35 2017-11-26 22:05:36
#> 959   0.31 2017-11-26 22:03:08
#> 960   0.25 2017-11-26 22:00:40
#> 961   0.51 2017-11-26 21:58:12
#> 962   0.53 2017-11-26 21:55:45
#> 963   0.50 2017-11-26 21:53:17
#> 964   0.47 2017-11-26 21:50:49
#> 965   0.57 2017-11-26 21:48:21
#> 966   0.52 2017-11-26 21:45:54
#> 967   0.59 2017-11-26 21:43:26
#> 968   0.57 2017-11-26 21:40:57
#> 969   0.50 2017-11-26 21:38:30
#> 970   0.56 2017-11-26 21:36:02
#> 971   0.54 2017-11-26 21:33:34
#> 972   0.55 2017-11-26 21:31:06
#> 973   0.47 2017-11-26 21:28:39
#> 974   0.47 2017-11-26 21:26:11
#> 975   0.55 2017-11-26 21:23:43
#> 976   0.66 2017-11-26 21:21:15
#> 977   0.66 2017-11-26 21:18:46
#> 978   0.68 2017-11-26 21:16:18
#> 979   0.69 2017-11-26 21:13:50
#> 980   0.62 2017-11-26 21:11:21
#> 981   0.64 2017-11-26 21:08:51
#> 982   0.61 2017-11-26 21:06:24
#> 983   0.59 2017-11-26 21:03:56
#> 984   0.49 2017-11-26 21:01:28
#> 985   0.44 2017-11-26 20:59:00
#> 986   0.57 2017-11-26 20:56:32
#> 987   0.40 2017-11-26 20:54:05
#> 988   0.49 2017-11-26 20:51:36
#> 989   0.64 2017-11-26 20:49:05
#> 990   0.55 2017-11-26 20:46:38
#> 991   0.54 2017-11-26 20:44:10
#> 992   0.57 2017-11-26 20:41:41
#> 993   0.56 2017-11-26 20:39:12
#> 994   0.50 2017-11-26 20:36:44
#> 995   0.51 2017-11-26 20:34:17
#> 996   0.59 2017-11-26 20:31:49
#> 997   0.62 2017-11-26 20:29:21
#> 998   0.58 2017-11-26 20:26:53
#> 999   0.58 2017-11-26 20:24:25
#> 1000  0.68 2017-11-26 20:21:57
#> 1001  0.66 2017-11-26 20:19:29
#> 1002  0.61 2017-11-26 20:17:01
#> 1003  0.70 2017-11-26 20:14:34
#> 1004  0.70 2017-11-26 20:12:06
#> 1005  0.74 2017-11-26 20:09:38
#> 1006  0.69 2017-11-26 20:07:10
#> 1007  0.68 2017-11-26 20:04:41
#> 1008  0.72 2017-11-26 20:02:14
#> 1009  0.67 2017-11-26 19:59:46
#> 1010  0.72 2017-11-26 19:57:17
#> 1011  0.69 2017-11-26 19:54:50
#> 1012  0.69 2017-11-26 19:52:22
#> 1013  0.66 2017-11-26 19:49:54
#> 1014  0.65 2017-11-26 19:47:26
#> 1015  0.73 2017-11-26 19:44:58
#> 1016  0.65 2017-11-26 19:42:30
#> 1017  0.66 2017-11-26 19:40:02
#> 1018  0.68 2017-11-26 19:37:35
#> 1019  0.66 2017-11-26 19:35:07
#> 1020  0.64 2017-11-26 19:32:39
#> 1021  0.75 2017-11-26 19:30:11
#> 1022  0.73 2017-11-26 19:27:43
#> 1023  0.66 2017-11-26 19:25:15
#> 1024  0.66 2017-11-26 19:22:47
#> 1025  0.65 2017-11-26 19:20:19
#> 1026  0.70 2017-11-26 19:17:51
#> 1027  0.69 2017-11-26 19:15:23
#> 1028  0.72 2017-11-26 19:12:55
#> 1029  0.68 2017-11-26 19:10:27
#> 1030  0.74 2017-11-26 19:07:59
#> 1031  0.79 2017-11-26 19:05:32
#> 1032  0.78 2017-11-26 19:03:03
#> 1033  0.78 2017-11-26 19:00:36
#> 1034  0.83 2017-11-26 18:58:08
#> 1035  0.78 2017-11-26 18:55:40
#> 1036  0.80 2017-11-26 18:53:12
#> 1037  0.82 2017-11-26 18:50:45
#> 1038  0.88 2017-11-26 18:48:17
#> 1039  0.80 2017-11-26 18:45:49
#> 1040  0.79 2017-11-26 18:43:21
#> 1041  0.76 2017-11-26 18:40:53
#> 1042  0.81 2017-11-26 18:38:26
#> 1043  0.79 2017-11-26 18:35:58
#> 1044  0.74 2017-11-26 18:33:30
#> 1045  0.77 2017-11-26 18:31:02
#> 1046  0.83 2017-11-26 18:28:34
#> 1047  0.84 2017-11-26 18:26:07
#> 1048  0.88 2017-11-26 18:23:37
#> 1049  0.95 2017-11-26 18:21:10
#> 1050  0.94 2017-11-26 18:18:42
#> 1051  0.94 2017-11-26 18:16:14
#> 1052  0.86 2017-11-26 18:13:46
#> 1053  0.80 2017-11-26 18:11:18
#> 1054  0.85 2017-11-26 18:08:50
#> 1055  0.92 2017-11-26 18:06:21
#> 1056  0.98 2017-11-26 18:03:53
#> 1057  0.92 2017-11-26 18:01:25
#> 1058  0.93 2017-11-26 17:58:57
#> 1059  0.89 2017-11-26 17:56:29
#> 1060  0.90 2017-11-26 17:54:00
#> 1061  0.88 2017-11-26 17:51:32
#> 1062  0.98 2017-11-26 17:49:04
#> 1063  0.88 2017-11-26 17:46:36
#> 1064  0.99 2017-11-26 17:44:09
#> 1065  0.96 2017-11-26 17:41:41
#> 1066  0.87 2017-11-26 17:39:13
#> 1067  0.85 2017-11-26 17:36:45
#> 1068  0.80 2017-11-26 17:34:17
#> 1069  0.69 2017-11-26 17:31:49
#> 1070  0.81 2017-11-26 17:29:21
#> 1071  0.83 2017-11-26 17:26:53
#> 1072  0.74 2017-11-26 17:24:25
#> 1073  0.68 2017-11-26 17:21:57
#> 1074  0.64 2017-11-26 17:19:29
#> 1075  0.69 2017-11-26 17:17:01
#> 1076  0.71 2017-11-26 17:14:34
#> 1077  0.75 2017-11-26 17:12:06
#> 1078  0.88 2017-11-26 17:09:38
#> 1079  0.81 2017-11-26 17:07:10
#> 1080  0.78 2017-11-26 17:04:42
#> 1081  0.75 2017-11-26 17:02:14
#> 1082  0.66 2017-11-26 16:59:46
#> 1083  0.62 2017-11-26 16:57:18
#> 1084  0.82 2017-11-26 16:54:50
#> 1085  0.83 2017-11-26 16:52:22
#> 1086  0.87 2017-11-26 16:49:54
#> 1087  0.88 2017-11-26 16:47:26
#> 1088  0.90 2017-11-26 16:44:58
#> 1089  1.00 2017-11-26 16:42:31
#> 1090  1.11 2017-11-26 16:40:03
#> 1091  1.09 2017-11-26 16:37:35
#> 1092  1.10 2017-11-26 16:35:07
#> 1093  1.11 2017-11-26 16:32:39
#> 1094  1.09 2017-11-26 16:30:11
#> 1095  1.10 2017-11-26 16:27:44
#> 1096  1.07 2017-11-26 16:25:16
#> 1097  1.10 2017-11-26 16:22:48
#> 1098  1.10 2017-11-26 16:20:20
#> 1099  1.04 2017-11-26 16:17:52
#> 1100  1.03 2017-11-26 16:15:25
#> 1101  1.07 2017-11-26 16:12:57
#> 1102  1.09 2017-11-26 16:10:29
#> 1103  1.16 2017-11-26 16:08:01
#> 1104  1.29 2017-11-26 16:05:33
#> 1105  1.43 2017-11-26 16:03:05
#> 1106  1.41 2017-11-26 16:00:38
#> 1107  1.30 2017-11-26 15:58:10
#> 1108  1.14 2017-11-26 15:55:42
#> 1109  1.23 2017-11-26 15:53:14
#> 1110  1.24 2017-11-26 15:50:46
#> 1111  1.28 2017-11-26 15:48:18
#> 1112  1.41 2017-11-26 15:45:50
#> 1113  1.56 2017-11-26 15:43:22
#> 1114  1.70 2017-11-26 15:40:55
#> 1115  1.42 2017-11-26 15:38:27
#> 1116  1.45 2017-11-26 15:35:59
#> 1117  1.49 2017-11-26 15:33:31
#> 1118  1.45 2017-11-26 15:31:03
#> 1119  1.45 2017-11-26 15:28:35
#> 1120  1.61 2017-11-26 15:26:07
#> 1121  1.61 2017-11-26 15:23:39
#> 1122  1.63 2017-11-26 15:21:11
#> 1123  1.65 2017-11-26 15:18:43
#> 1124  1.69 2017-11-26 15:16:15
#> 1125  1.63 2017-11-26 15:13:47
#> 1126  1.64 2017-11-26 15:11:19
#> 1127  1.61 2017-11-26 15:08:51
#> 1128  1.63 2017-11-26 15:06:24
#> 1129  1.65 2017-11-26 15:03:56
#> 1130  1.78 2017-11-26 15:01:28
#> 1131  1.89 2017-11-26 14:59:00
#> 1132  1.71 2017-11-26 14:56:32
#> 1133  1.68 2017-11-26 14:54:04
#> 1134  1.83 2017-11-26 14:51:36
#> 1135  1.98 2017-11-26 14:49:07
#> 1136  1.93 2017-11-26 14:46:39
#> 1137  1.85 2017-11-26 14:44:11
#> 1138  1.71 2017-11-26 14:41:44
#> 1139  1.75 2017-11-26 14:39:16
#> 1140  1.80 2017-11-26 14:36:48
#> 1141  2.03 2017-11-26 14:34:20
#> 1142  1.90 2017-11-26 14:31:52
#> 1143  1.81 2017-11-26 14:29:24
#> 1144  1.82 2017-11-26 14:26:56
#> 1145  1.92 2017-11-26 14:24:28
#> 1146  2.01 2017-11-26 14:22:00
#> 1147  1.98 2017-11-26 14:19:33
#> 1148  1.93 2017-11-26 14:17:05
#> 1149  2.01 2017-11-26 14:14:37
#> 1150  1.85 2017-11-26 14:12:08
#> 1151  1.84 2017-11-26 14:09:40
#> 1152  1.93 2017-11-26 14:07:12
#> 1153  1.86 2017-11-26 14:04:44
#> 1154  1.79 2017-11-26 14:02:16
#> 1155  1.93 2017-11-26 13:59:48
#> 1156  1.88 2017-11-26 13:57:20
#> 1157  1.97 2017-11-26 13:54:53
#> 1158  2.01 2017-11-26 13:52:25
#> 1159  1.91 2017-11-26 13:49:57
#> 1160  1.83 2017-11-26 13:47:29
#> 1161  1.98 2017-11-26 13:45:01
#> 1162  1.87 2017-11-26 13:42:33
#> 1163  1.85 2017-11-26 13:40:05
#> 
#> 

# NOT RUN {
get_senseBox_data(
  senseBoxId = c("5957b67494f0520011304cc1","5957b67494f0520011304cc1") ,
  sensorId = "all",
  fromDate = "2017-11-25-12:00:00",
  toDate = "2017-11-26-12:00:00")

get_senseBox_data(
  senseBoxId = "5957b67494f0520011304cc1",
  sensorId = "5957b67494f0520011304cc4",
  fromDate = "2017-11-25-12:00:00",
  toDate = "2017-11-26-12:00:00")
# }