Search senseBoxes with special phenomenas, dates, exposures, ...

If the arguments `fromDate` and `toDate` are NOT given, the measurements from the last 48 h from now are downloaded. The maximum numbers of downloaded measurements are 10,000. The maximum time frame is one month.

search_senseBox(phenomenon = NA, exposure = NA, model = NA,
  grouptag = NA, date = NA, fromDate = NA, toDate = NA,
  tidy = FALSE)

Arguments

phenomenon
[character] (**optional**): Which phenomenon do you want to search for?

exposure
[character] (**optional**): Which exposure do you want to search for? "indoor", "outdoor", "mobile", "unknown"

model
[character] (**optional**): Which model do you want to search for? "homeEthernet", "homeWifi", "homeEthernetFeinstaub", "homeWifiFeinstaub", "luftdaten_sds011", " luftdaten_sds011_dht11", "luftdaten_sds011_dht22", "luftdaten_sds011_bmp180", "luftdaten_sds011_bme280"

grouptag
[character] (**optional**): Which grouptag do you want to search for?

date
[character] (**optional**): Which date do you want to search for? Not useable in combination with fromDate and toDate

fromDate
[character] (**optional**): Just show data with from date in format YYYY-mm-dd HH:MM:SS

toDate
[character] (**optional**): Just show data with up to date in format YYYY-mm-dd HH:MM:SS

tidy
[logical] (**optional**): Should the output be a tidy data frame?

Value

A [data.frame] with every entry is a sensBoxId.

Function version

0.0.1

Examples

indoor <- search_senseBox(exposure = "indoor")

Helligkeit<- search_senseBox(phenomenon = "Helligkeit",
                             date = "2018-01-01 00:00:00")

ifgi_outdoor <- search_senseBox(grouptag = 'ifgi', exposure = 'outdoor')