Get data from one specific senseBox sensor.

If the arguments fromDate and toDate are NOT given, the measurements from the last 48 h from now are downloaded. The maximum numbers of downloaded measurements are 10,000. The maximum time frame is one month. It is possible to download the data directly as a csv file, when argument CSV is TRUE.

get_senseBox_data(senseBoxId, sensorId = rep("all", times =
  length(unlist(senseBoxId))), fromDate = NULL, toDate = NULL,
  CSV = FALSE, POSIXct = TRUE)

Arguments

senseBoxId
characteror vector of character (required): senseBoxId or a vector of character of senseBoxIds

sensorId
character or list of character (required): sensorId or a list of senseBoxIds

fromDate
character (optional): Just show data with from date in format YYYY-mm-dd HH:MM:SS

toDate
character (optional): Just show data with up to date in format YYYY-mm-dd HH:MM:SS

CSV
logical (optional): Download data as .csv? NOT POSSIBLE UNTIL NOW!

POSIXct
logical (optional): Should the timestamp be translated into POSIXct?

Value

list

Function version

0.0.1

Examples

##=====================================
## Example: get data from sensorId
##=====================================

get_senseBox_data(
  senseBoxId = "5957b67494f0520011304cc1",
  sensorId = "5957b67494f0520011304cc4")

#> $`5957b67494f0520011304cc1`
#> $`5957b67494f0520011304cc1`$`Temperatur [°C]`
#>      value           createdAt
#> 1    -1.26 2017-11-27 19:34:41
#> 2    -1.22 2017-11-27 19:32:11
#> 3    -1.18 2017-11-27 19:29:42
#> 4    -1.14 2017-11-27 19:27:13
#> 5    -1.11 2017-11-27 19:24:43
#> 6    -1.06 2017-11-27 19:22:14
#> 7    -1.03 2017-11-27 19:19:45
#> 8    -1.03 2017-11-27 19:17:15
#> 9    -1.00 2017-11-27 19:14:46
#> 10   -0.93 2017-11-27 19:12:18
#> 11   -0.94 2017-11-27 19:09:50
#> 12   -0.96 2017-11-27 19:07:21
#> 13   -0.94 2017-11-27 19:04:54
#> 14   -0.83 2017-11-27 19:02:26
#> 15   -0.89 2017-11-27 18:59:58
#> 16   -0.89 2017-11-27 18:57:30
#> 17   -0.84 2017-11-27 18:55:02
#> 18   -0.78 2017-11-27 18:52:34
#> 19   -0.73 2017-11-27 18:50:05
#> 20   -0.71 2017-11-27 18:47:37
#> 21   -0.72 2017-11-27 18:45:09
#> 22   -0.84 2017-11-27 18:42:40
#> 23   -0.77 2017-11-27 18:40:13
#> 24   -0.78 2017-11-27 18:37:45
#> 25   -0.75 2017-11-27 18:35:17
#> 26   -0.66 2017-11-27 18:32:49
#> 27   -0.57 2017-11-27 18:30:21
#> 28   -0.37 2017-11-27 18:27:53
#> 29   -0.39 2017-11-27 18:25:25
#> 30   -0.47 2017-11-27 18:22:57
#> 31   -0.56 2017-11-27 18:20:29
#> 32   -0.40 2017-11-27 18:18:01
#> 33   -0.35 2017-11-27 18:15:33
#> 34   -0.30 2017-11-27 18:13:05
#> 35   -0.46 2017-11-27 18:10:37
#> 36   -0.42 2017-11-27 18:08:09
#> 37   -0.48 2017-11-27 18:05:41
#> 38   -0.39 2017-11-27 18:03:13
#> 39   -0.50 2017-11-27 18:00:45
#> 40   -0.47 2017-11-27 17:58:18
#> 41   -0.46 2017-11-27 17:55:50
#> 42   -0.35 2017-11-27 17:53:22
#> 43   -0.43 2017-11-27 17:50:54
#> 44   -0.24 2017-11-27 17:48:26
#> 45   -0.24 2017-11-27 17:45:58
#> 46   -0.17 2017-11-27 17:43:30
#> 47   -0.07 2017-11-27 17:41:03
#> 48   -0.21 2017-11-27 17:38:32
#> 49   -0.11 2017-11-27 17:36:05
#> 50   -0.18 2017-11-27 17:33:36
#> 51   -0.22 2017-11-27 17:31:08
#> 52   -0.09 2017-11-27 17:28:40
#> 53   -0.08 2017-11-27 17:26:12
#> 54    0.08 2017-11-27 17:23:44
#> 55    0.07 2017-11-27 17:21:17
#> 56    0.04 2017-11-27 17:18:49
#> 57    0.25 2017-11-27 17:16:21
#> 58    0.25 2017-11-27 17:13:53
#> 59    0.26 2017-11-27 17:11:25
#> 60    0.37 2017-11-27 17:08:58
#> 61    0.44 2017-11-27 17:06:30
#> 62    0.36 2017-11-27 17:04:02
#> 63    0.40 2017-11-27 17:01:34
#> 64    0.62 2017-11-27 16:59:06
#> 65    0.68 2017-11-27 16:56:37
#> 66    0.61 2017-11-27 16:54:09
#> 67    0.68 2017-11-27 16:51:42
#> 68    0.81 2017-11-27 16:49:14
#> 69    0.86 2017-11-27 16:46:46
#> 70    0.80 2017-11-27 16:44:18
#> 71    0.86 2017-11-27 16:41:50
#> 72    0.84 2017-11-27 16:39:21
#> 73    0.94 2017-11-27 16:36:53
#> 74    1.20 2017-11-27 16:34:25
#> 75    1.23 2017-11-27 16:31:57
#> 76    1.27 2017-11-27 16:29:30
#> 77    1.27 2017-11-27 16:27:02
#> 78    1.37 2017-11-27 16:24:34
#> 79    1.39 2017-11-27 16:22:06
#> 80    1.47 2017-11-27 16:19:38
#> 81    1.47 2017-11-27 16:17:11
#> 82    1.51 2017-11-27 16:14:43
#> 83    1.51 2017-11-27 16:12:15
#> 84    1.54 2017-11-27 16:09:47
#> 85    1.57 2017-11-27 16:07:19
#> 86    1.55 2017-11-27 16:04:52
#> 87    1.53 2017-11-27 16:02:24
#> 88    1.55 2017-11-27 15:59:56
#> 89    1.49 2017-11-27 15:57:27
#> 90    1.45 2017-11-27 15:54:59
#> 91    1.46 2017-11-27 15:52:31
#> 92    1.49 2017-11-27 15:50:03
#> 93    1.54 2017-11-27 15:47:35
#> 94    1.54 2017-11-27 15:45:07
#> 95    1.59 2017-11-27 15:42:39
#> 96    1.67 2017-11-27 15:40:10
#> 97    1.49 2017-11-27 15:37:42
#> 98    1.85 2017-11-27 15:35:14
#> 99    1.98 2017-11-27 15:32:46
#> 100   1.97 2017-11-27 15:30:18
#> 101   2.09 2017-11-27 15:27:50
#> 102   2.13 2017-11-27 15:25:21
#> 103   2.11 2017-11-27 15:22:53
#> 104   2.05 2017-11-27 15:20:26
#> 105   2.08 2017-11-27 15:17:58
#> 106   2.12 2017-11-27 15:15:30
#> 107   2.08 2017-11-27 15:13:02
#> 108   1.98 2017-11-27 15:10:34
#> 109   2.07 2017-11-27 15:08:06
#> 110   1.98 2017-11-27 15:05:38
#> 111   2.18 2017-11-27 15:03:11
#> 112   2.28 2017-11-27 15:00:43
#> 113   2.32 2017-11-27 14:58:15
#> 114   2.43 2017-11-27 14:55:47
#> 115   2.51 2017-11-27 14:53:19
#> 116   2.69 2017-11-27 14:50:52
#> 117   2.58 2017-11-27 14:48:24
#> 118   2.64 2017-11-27 14:45:56
#> 119   2.78 2017-11-27 14:43:28
#> 120   2.83 2017-11-27 14:41:00
#> 121   2.73 2017-11-27 14:38:32
#> 122   2.79 2017-11-27 14:36:05
#> 123   2.83 2017-11-27 14:33:37
#> 124   3.18 2017-11-27 14:31:08
#> 125   3.26 2017-11-27 14:28:40
#> 126   3.55 2017-11-27 14:26:13
#> 127   3.53 2017-11-27 14:23:44
#> 128   3.61 2017-11-27 14:21:16
#> 129   3.43 2017-11-27 14:18:48
#> 130   3.53 2017-11-27 14:16:20
#> 131   3.70 2017-11-27 14:13:53
#> 132   3.89 2017-11-27 14:11:25
#> 133   4.14 2017-11-27 14:08:57
#> 134   3.93 2017-11-27 14:06:29
#> 135   4.02 2017-11-27 14:04:01
#> 136   3.93 2017-11-27 14:01:34
#> 137   4.03 2017-11-27 13:59:06
#> 138   4.14 2017-11-27 13:56:38
#> 139   4.26 2017-11-27 13:54:10
#> 140   4.25 2017-11-27 13:51:42
#> 141   4.39 2017-11-27 13:49:14
#> 142   4.23 2017-11-27 13:46:46
#> 143   4.17 2017-11-27 13:44:18
#> 144   4.28 2017-11-27 13:41:50
#> 145   4.17 2017-11-27 13:39:22
#> 146   4.13 2017-11-27 13:36:55
#> 147   4.24 2017-11-27 13:34:27
#> 148   4.16 2017-11-27 13:31:59
#> 149   4.38 2017-11-27 13:29:31
#> 150   4.69 2017-11-27 13:27:03
#> 151   4.64 2017-11-27 13:24:36
#> 152   4.72 2017-11-27 13:22:08
#> 153   4.70 2017-11-27 13:19:40
#> 154   4.88 2017-11-27 13:17:12
#> 155   4.85 2017-11-27 13:14:44
#> 156   5.20 2017-11-27 13:12:16
#> 157   5.02 2017-11-27 13:09:49
#> 158   4.81 2017-11-27 13:07:20
#> 159   4.72 2017-11-27 13:04:52
#> 160   4.63 2017-11-27 13:02:24
#> 161   4.53 2017-11-27 12:59:56
#> 162   4.57 2017-11-27 12:57:29
#> 163   4.53 2017-11-27 12:55:01
#> 164   4.38 2017-11-27 12:52:33
#> 165   4.16 2017-11-27 12:50:03
#> 166   4.40 2017-11-27 12:47:36
#> 167   4.39 2017-11-27 12:45:08
#> 168   4.34 2017-11-27 12:42:40
#> 169   4.58 2017-11-27 12:40:12
#> 170   4.49 2017-11-27 12:37:44
#> 171   4.46 2017-11-27 12:35:17
#> 172   4.53 2017-11-27 12:32:49
#> 173   4.38 2017-11-27 12:30:21
#> 174   4.26 2017-11-27 12:27:53
#> 175   4.24 2017-11-27 12:25:26
#> 176   4.91 2017-11-27 12:22:58
#> 177   4.91 2017-11-27 12:20:30
#> 178   4.99 2017-11-27 12:18:02
#> 179   4.85 2017-11-27 12:15:34
#> 180   4.91 2017-11-27 12:13:07
#> 181   4.82 2017-11-27 12:10:39
#> 182   4.67 2017-11-27 12:08:11
#> 183   4.93 2017-11-27 12:05:43
#> 184   4.68 2017-11-27 12:03:16
#> 185   4.41 2017-11-27 12:00:48
#> 186   4.41 2017-11-27 11:58:20
#> 187   3.93 2017-11-27 11:55:52
#> 188   3.98 2017-11-27 11:53:23
#> 189   3.79 2017-11-27 11:50:55
#> 190   4.32 2017-11-27 11:48:28
#> 191   4.65 2017-11-27 11:46:00
#> 192   4.42 2017-11-27 11:43:32
#> 193   4.58 2017-11-27 11:41:04
#> 194   4.53 2017-11-27 11:38:37
#> 195   4.97 2017-11-27 11:36:09
#> 196   4.74 2017-11-27 11:33:41
#> 197   4.89 2017-11-27 11:31:13
#> 198   4.63 2017-11-27 11:28:46
#> 199   4.22 2017-11-27 11:26:18
#> 200   4.57 2017-11-27 11:23:50
#> 201   4.61 2017-11-27 11:21:23
#> 202   4.74 2017-11-27 11:18:55
#> 203   5.29 2017-11-27 11:16:27
#> 204   5.59 2017-11-27 11:13:59
#> 205   6.38 2017-11-27 11:11:31
#> 206   5.41 2017-11-27 11:09:04
#> 207   6.11 2017-11-27 11:06:36
#> 208   6.13 2017-11-27 11:04:08
#> 209   7.41 2017-11-27 11:01:40
#> 210   7.52 2017-11-27 10:59:13
#> 211   7.04 2017-11-27 10:56:45
#> 212   7.62 2017-11-27 10:54:17
#> 213   7.52 2017-11-27 10:51:49
#> 214   7.30 2017-11-27 10:49:22
#> 215   7.00 2017-11-27 10:46:54
#> 216   6.31 2017-11-27 10:44:26
#> 217   5.88 2017-11-27 10:41:58
#> 218   5.14 2017-11-27 10:39:31
#> 219   4.67 2017-11-27 10:37:03
#> 220   4.35 2017-11-27 10:34:35
#> 221   4.35 2017-11-27 10:32:07
#> 222   4.17 2017-11-27 10:29:39
#> 223   4.22 2017-11-27 10:27:12
#> 224   4.33 2017-11-27 10:24:44
#> 225   4.25 2017-11-27 10:22:16
#> 226   4.35 2017-11-27 10:19:48
#> 227   4.33 2017-11-27 10:17:21
#> 228   4.27 2017-11-27 10:14:53
#> 229   4.28 2017-11-27 10:12:25
#> 230   4.33 2017-11-27 10:09:57
#> 231   4.57 2017-11-27 10:07:29
#> 232   4.45 2017-11-27 10:05:01
#> 233   4.20 2017-11-27 10:02:33
#> 234   3.97 2017-11-27 10:00:05
#> 235   4.03 2017-11-27 09:57:38
#> 236   4.59 2017-11-27 09:55:10
#> 237   5.04 2017-11-27 09:52:40
#> 238   4.63 2017-11-27 09:50:12
#> 239   4.25 2017-11-27 09:47:44
#> 240   3.44 2017-11-27 09:45:16
#> 241   3.31 2017-11-27 09:42:48
#> 242   3.59 2017-11-27 09:40:21
#> 243   3.90 2017-11-27 09:37:53
#> 244   3.52 2017-11-27 09:35:25
#> 245   3.18 2017-11-27 09:32:57
#> 246   3.07 2017-11-27 09:30:30
#> 247   2.83 2017-11-27 09:28:02
#> 248   2.79 2017-11-27 09:25:34
#> 249   2.85 2017-11-27 09:23:06
#> 250   2.54 2017-11-27 09:20:39
#> 251   2.70 2017-11-27 09:18:11
#> 252   2.33 2017-11-27 09:15:43
#> 253   2.70 2017-11-27 09:13:15
#> 254   2.46 2017-11-27 09:10:47
#> 255   2.70 2017-11-27 09:08:20
#> 256   2.83 2017-11-27 09:05:52
#> 257   2.82 2017-11-27 09:03:24
#> 258   2.82 2017-11-27 09:00:56
#> 259   2.75 2017-11-27 08:58:29
#> 260   3.10 2017-11-27 08:56:01
#> 261   2.86 2017-11-27 08:53:32
#> 262   2.69 2017-11-27 08:51:04
#> 263   2.41 2017-11-27 08:48:36
#> 264   1.87 2017-11-27 08:46:09
#> 265   1.64 2017-11-27 08:43:41
#> 266   1.52 2017-11-27 08:41:13
#> 267   1.72 2017-11-27 08:38:45
#> 268   2.18 2017-11-27 08:36:16
#> 269   2.43 2017-11-27 08:33:49
#> 270   2.36 2017-11-27 08:31:21
#> 271   2.07 2017-11-27 08:28:53
#> 272   2.12 2017-11-27 08:26:25
#> 273   2.05 2017-11-27 08:23:58
#> 274   2.00 2017-11-27 08:21:30
#> 275   1.92 2017-11-27 08:19:02
#> 276   1.83 2017-11-27 08:16:34
#> 277   1.91 2017-11-27 08:14:06
#> 278   1.99 2017-11-27 08:11:38
#> 279   2.06 2017-11-27 08:09:11
#> 280   1.47 2017-11-27 08:06:43
#> 281   1.35 2017-11-27 08:04:15
#> 282   1.20 2017-11-27 08:01:47
#> 283   1.16 2017-11-27 07:59:20
#> 284   1.12 2017-11-27 07:56:52
#> 285   1.09 2017-11-27 07:54:24
#> 286   1.21 2017-11-27 07:51:56
#> 287   0.85 2017-11-27 07:49:29
#> 288   0.72 2017-11-27 07:47:01
#> 289   0.87 2017-11-27 07:44:33
#> 290   0.96 2017-11-27 07:42:05
#> 291   1.17 2017-11-27 07:39:38
#> 292   1.06 2017-11-27 07:37:04
#> 293   1.01 2017-11-27 07:34:36
#> 294   0.89 2017-11-27 07:32:08
#> 295   0.86 2017-11-27 07:29:40
#> 296   0.80 2017-11-27 07:27:13
#> 297   0.75 2017-11-27 07:24:45
#> 298   0.73 2017-11-27 07:22:17
#> 299   0.64 2017-11-27 07:19:50
#> 300   0.58 2017-11-27 07:17:22
#> 301   0.61 2017-11-27 07:14:54
#> 302   0.60 2017-11-27 07:12:26
#> 303   0.51 2017-11-27 07:09:58
#> 304   0.47 2017-11-27 07:07:31
#> 305   0.42 2017-11-27 07:05:02
#> 306   0.34 2017-11-27 07:02:34
#> 307   0.37 2017-11-27 07:00:06
#> 308   0.25 2017-11-27 06:57:39
#> 309   0.19 2017-11-27 06:55:10
#> 310   0.21 2017-11-27 06:52:42
#> 311   0.30 2017-11-27 06:50:14
#> 312   0.28 2017-11-27 06:47:46
#> 313   0.22 2017-11-27 06:45:18
#> 314   0.25 2017-11-27 06:42:50
#> 315   0.34 2017-11-27 06:40:22
#> 316   0.26 2017-11-27 06:37:55
#> 317   0.24 2017-11-27 06:35:27
#> 318   0.24 2017-11-27 06:32:59
#> 319   0.24 2017-11-27 06:30:31
#> 320   0.27 2017-11-27 06:28:03
#> 321   0.26 2017-11-27 06:25:35
#> 322   0.20 2017-11-27 06:23:08
#> 323   0.17 2017-11-27 06:20:40
#> 324   0.20 2017-11-27 06:18:12
#> 325   0.19 2017-11-27 06:15:44
#> 326   0.21 2017-11-27 06:13:17
#> 327   0.20 2017-11-27 06:10:49
#> 328   0.32 2017-11-27 06:08:21
#> 329   0.30 2017-11-27 06:05:54
#> 330   0.24 2017-11-27 06:03:26
#> 331   0.30 2017-11-27 06:00:58
#> 332   0.21 2017-11-27 05:58:30
#> 333   0.20 2017-11-27 05:56:03
#> 334   0.30 2017-11-27 05:53:35
#> 335   0.27 2017-11-27 05:51:07
#> 336   0.29 2017-11-27 05:48:39
#> 337   0.27 2017-11-27 05:46:12
#> 338   0.30 2017-11-27 05:43:44
#> 339   0.23 2017-11-27 05:41:16
#> 340   0.39 2017-11-27 05:38:49
#> 341   0.30 2017-11-27 05:36:21
#> 342   0.31 2017-11-27 05:33:53
#> 343   0.31 2017-11-27 05:31:25
#> 344   0.36 2017-11-27 05:28:58
#> 345   0.40 2017-11-27 05:26:30
#> 346   0.34 2017-11-27 05:24:02
#> 347   0.32 2017-11-27 05:21:34
#> 348   0.35 2017-11-27 05:19:07
#> 349   0.39 2017-11-27 05:16:39
#> 350   0.36 2017-11-27 05:14:11
#> 351   0.27 2017-11-27 05:11:43
#> 352   0.27 2017-11-27 05:09:16
#> 353   0.28 2017-11-27 05:06:48
#> 354   0.26 2017-11-27 05:04:20
#> 355   0.23 2017-11-27 05:01:52
#> 356   0.21 2017-11-27 04:59:25
#> 357   0.20 2017-11-27 04:56:57
#> 358   0.28 2017-11-27 04:54:29
#> 359   0.26 2017-11-27 04:52:01
#> 360   0.24 2017-11-27 04:49:34
#> 361   0.30 2017-11-27 04:47:06
#> 362   0.31 2017-11-27 04:44:38
#> 363   0.32 2017-11-27 04:42:10
#> 364   0.35 2017-11-27 04:39:43
#> 365   0.36 2017-11-27 04:37:15
#> 366   0.34 2017-11-27 04:34:47
#> 367   0.34 2017-11-27 04:32:19
#> 368   0.29 2017-11-27 04:29:52
#> 369   0.30 2017-11-27 04:27:24
#> 370   0.31 2017-11-27 04:24:56
#> 371   0.30 2017-11-27 04:22:28
#> 372   0.29 2017-11-27 04:20:01
#> 373   0.22 2017-11-27 04:17:33
#> 374   0.22 2017-11-27 04:15:05
#> 375   0.30 2017-11-27 04:12:37
#> 376   0.34 2017-11-27 04:10:10
#> 377   0.44 2017-11-27 04:07:42
#> 378   0.45 2017-11-27 04:05:14
#> 379   0.41 2017-11-27 04:02:46
#> 380   0.40 2017-11-27 04:00:19
#> 381   0.41 2017-11-27 03:57:51
#> 382   0.41 2017-11-27 03:55:23
#> 383   0.33 2017-11-27 03:52:55
#> 384   0.27 2017-11-27 03:50:28
#> 385   0.21 2017-11-27 03:48:00
#> 386   0.18 2017-11-27 03:45:32
#> 387   0.23 2017-11-27 03:43:04
#> 388   0.14 2017-11-27 03:40:37
#> 389   0.10 2017-11-27 03:38:09
#> 390  -0.02 2017-11-27 03:35:41
#> 391   0.09 2017-11-27 03:33:13
#> 392   0.06 2017-11-27 03:30:46
#> 393   0.09 2017-11-27 03:28:18
#> 394   0.19 2017-11-27 03:25:50
#> 395   0.19 2017-11-27 03:23:22
#> 396   0.18 2017-11-27 03:20:55
#> 397   0.30 2017-11-27 03:18:27
#> 398   0.32 2017-11-27 03:15:59
#> 399   0.31 2017-11-27 03:13:31
#> 400   0.34 2017-11-27 03:11:04
#> 401   0.39 2017-11-27 03:08:36
#> 402   0.48 2017-11-27 03:06:08
#> 403   0.25 2017-11-27 03:03:40
#> 404   0.24 2017-11-27 03:01:13
#> 405   0.22 2017-11-27 02:58:45
#> 406   0.23 2017-11-27 02:56:16
#> 407   0.27 2017-11-27 02:53:48
#> 408   0.24 2017-11-27 02:51:20
#> 409   0.20 2017-11-27 02:48:53
#> 410   0.22 2017-11-27 02:46:25
#> 411   0.29 2017-11-27 02:43:57
#> 412   0.26 2017-11-27 02:41:29
#> 413   0.29 2017-11-27 02:39:01
#> 414   0.27 2017-11-27 02:36:34
#> 415   0.21 2017-11-27 02:34:06
#> 416   0.26 2017-11-27 02:31:38
#> 417   0.26 2017-11-27 02:29:10
#> 418   0.25 2017-11-27 02:26:42
#> 419   0.24 2017-11-27 02:24:14
#> 420   0.28 2017-11-27 02:21:47
#> 421   0.27 2017-11-27 02:19:19
#> 422   0.28 2017-11-27 02:16:51
#> 423   0.27 2017-11-27 02:14:23
#> 424   0.24 2017-11-27 02:11:56
#> 425   0.22 2017-11-27 02:09:28
#> 426   0.24 2017-11-27 02:07:00
#> 427   0.20 2017-11-27 02:04:32
#> 428   0.17 2017-11-27 02:02:05
#> 429   0.20 2017-11-27 01:59:37
#> 430   0.09 2017-11-27 01:57:09
#> 431   0.06 2017-11-27 01:54:41
#> 432   0.09 2017-11-27 01:52:13
#> 433   0.17 2017-11-27 01:49:45
#> 434   0.23 2017-11-27 01:47:18
#> 435   0.31 2017-11-27 01:44:50
#> 436   0.26 2017-11-27 01:42:22
#> 437   0.26 2017-11-27 01:39:54
#> 438   0.21 2017-11-27 01:37:27
#> 439   0.14 2017-11-27 01:34:59
#> 440   0.28 2017-11-27 01:32:31
#> 441   0.33 2017-11-27 01:30:03
#> 442   0.28 2017-11-27 01:27:36
#> 443   0.23 2017-11-27 01:25:08
#> 444   0.18 2017-11-27 01:22:40
#> 445   0.31 2017-11-27 01:20:13
#> 446   0.34 2017-11-27 01:17:45
#> 447   0.40 2017-11-27 01:15:17
#> 448   0.41 2017-11-27 01:12:49
#> 449   0.36 2017-11-27 01:10:21
#> 450   0.44 2017-11-27 01:07:54
#> 451   0.42 2017-11-27 01:05:26
#> 452   0.45 2017-11-27 01:02:58
#> 453   0.37 2017-11-27 01:00:31
#> 454   0.23 2017-11-27 00:58:03
#> 455   0.09 2017-11-27 00:55:35
#> 456   0.10 2017-11-27 00:53:07
#> 457   0.16 2017-11-27 00:50:39
#> 458   0.23 2017-11-27 00:48:11
#> 459   0.23 2017-11-27 00:45:43
#> 460   0.12 2017-11-27 00:43:16
#> 461   0.09 2017-11-27 00:40:48
#> 462   0.04 2017-11-27 00:38:20
#> 463   0.08 2017-11-27 00:35:52
#> 464   0.26 2017-11-27 00:33:25
#> 465   0.21 2017-11-27 00:30:57
#> 466   0.25 2017-11-27 00:28:29
#> 467   0.24 2017-11-27 00:26:01
#> 468   0.22 2017-11-27 00:23:33
#> 469   0.21 2017-11-27 00:21:06
#> 470   0.25 2017-11-27 00:18:38
#> 471   0.25 2017-11-27 00:16:10
#> 472   0.20 2017-11-27 00:13:42
#> 473   0.18 2017-11-27 00:11:15
#> 474   0.19 2017-11-27 00:08:47
#> 475   0.23 2017-11-27 00:06:19
#> 476   0.11 2017-11-27 00:03:51
#> 477   0.19 2017-11-27 00:01:23
#> 478   0.20 2017-11-26 23:58:56
#> 479   0.24 2017-11-26 23:56:28
#> 480   0.28 2017-11-26 23:54:00
#> 481   0.27 2017-11-26 23:51:32
#> 482   0.40 2017-11-26 23:49:05
#> 483   0.36 2017-11-26 23:46:37
#> 484   0.40 2017-11-26 23:44:09
#> 485   0.36 2017-11-26 23:41:41
#> 486   0.37 2017-11-26 23:39:13
#> 487   0.37 2017-11-26 23:36:46
#> 488   0.35 2017-11-26 23:34:18
#> 489   0.26 2017-11-26 23:31:50
#> 490   0.30 2017-11-26 23:29:22
#> 491   0.34 2017-11-26 23:26:55
#> 492   0.37 2017-11-26 23:24:27
#> 493   0.36 2017-11-26 23:21:59
#> 494   0.30 2017-11-26 23:19:31
#> 495   0.29 2017-11-26 23:17:03
#> 496   0.35 2017-11-26 23:14:36
#> 497   0.38 2017-11-26 23:12:08
#> 498   0.44 2017-11-26 23:09:40
#> 499   0.63 2017-11-26 23:07:12
#> 500   0.64 2017-11-26 23:04:45
#> 501   0.58 2017-11-26 23:02:17
#> 502   0.46 2017-11-26 22:59:49
#> 503   0.48 2017-11-26 22:57:21
#> 504   0.52 2017-11-26 22:54:54
#> 505   0.51 2017-11-26 22:52:26
#> 506   0.51 2017-11-26 22:49:57
#> 507   0.50 2017-11-26 22:47:29
#> 508   0.56 2017-11-26 22:45:01
#> 509   0.50 2017-11-26 22:42:32
#> 510   0.51 2017-11-26 22:40:04
#> 511   0.49 2017-11-26 22:37:37
#> 512   0.48 2017-11-26 22:35:09
#> 513   0.42 2017-11-26 22:32:41
#> 514   0.39 2017-11-26 22:30:13
#> 515   0.45 2017-11-26 22:27:46
#> 516   0.45 2017-11-26 22:25:18
#> 517   0.48 2017-11-26 22:22:50
#> 518   0.32 2017-11-26 22:20:22
#> 519   0.30 2017-11-26 22:17:54
#> 520   0.30 2017-11-26 22:15:27
#> 521   0.42 2017-11-26 22:12:59
#> 522   0.24 2017-11-26 22:10:31
#> 523   0.43 2017-11-26 22:08:03
#> 524   0.35 2017-11-26 22:05:36
#> 525   0.31 2017-11-26 22:03:08
#> 526   0.25 2017-11-26 22:00:40
#> 527   0.51 2017-11-26 21:58:12
#> 528   0.53 2017-11-26 21:55:45
#> 529   0.50 2017-11-26 21:53:17
#> 530   0.47 2017-11-26 21:50:49
#> 531   0.57 2017-11-26 21:48:21
#> 532   0.52 2017-11-26 21:45:54
#> 533   0.59 2017-11-26 21:43:26
#> 534   0.57 2017-11-26 21:40:57
#> 535   0.50 2017-11-26 21:38:30
#> 536   0.56 2017-11-26 21:36:02
#> 537   0.54 2017-11-26 21:33:34
#> 538   0.55 2017-11-26 21:31:06
#> 539   0.47 2017-11-26 21:28:39
#> 540   0.47 2017-11-26 21:26:11
#> 541   0.55 2017-11-26 21:23:43
#> 542   0.66 2017-11-26 21:21:15
#> 543   0.66 2017-11-26 21:18:46
#> 544   0.68 2017-11-26 21:16:18
#> 545   0.69 2017-11-26 21:13:50
#> 546   0.62 2017-11-26 21:11:21
#> 547   0.64 2017-11-26 21:08:51
#> 548   0.61 2017-11-26 21:06:24
#> 549   0.59 2017-11-26 21:03:56
#> 550   0.49 2017-11-26 21:01:28
#> 551   0.44 2017-11-26 20:59:00
#> 552   0.57 2017-11-26 20:56:32
#> 553   0.40 2017-11-26 20:54:05
#> 554   0.49 2017-11-26 20:51:36
#> 555   0.64 2017-11-26 20:49:05
#> 556   0.55 2017-11-26 20:46:38
#> 557   0.54 2017-11-26 20:44:10
#> 558   0.57 2017-11-26 20:41:41
#> 559   0.56 2017-11-26 20:39:12
#> 560   0.50 2017-11-26 20:36:44
#> 561   0.51 2017-11-26 20:34:17
#> 562   0.59 2017-11-26 20:31:49
#> 563   0.62 2017-11-26 20:29:21
#> 564   0.58 2017-11-26 20:26:53
#> 565   0.58 2017-11-26 20:24:25
#> 566   0.68 2017-11-26 20:21:57
#> 567   0.66 2017-11-26 20:19:29
#> 568   0.61 2017-11-26 20:17:01
#> 569   0.70 2017-11-26 20:14:34
#> 570   0.70 2017-11-26 20:12:06
#> 571   0.74 2017-11-26 20:09:38
#> 572   0.69 2017-11-26 20:07:10
#> 573   0.68 2017-11-26 20:04:41
#> 574   0.72 2017-11-26 20:02:14
#> 575   0.67 2017-11-26 19:59:46
#> 576   0.72 2017-11-26 19:57:17
#> 577   0.69 2017-11-26 19:54:50
#> 578   0.69 2017-11-26 19:52:22
#> 579   0.66 2017-11-26 19:49:54
#> 580   0.65 2017-11-26 19:47:26
#> 581   0.73 2017-11-26 19:44:58
#> 582   0.65 2017-11-26 19:42:30
#> 583   0.66 2017-11-26 19:40:02
#> 584   0.68 2017-11-26 19:37:35
#> 585   0.66 2017-11-26 19:35:07
#> 586   0.64 2017-11-26 19:32:39
#> 587   0.75 2017-11-26 19:30:11
#> 588   0.73 2017-11-26 19:27:43
#> 589   0.66 2017-11-26 19:25:15
#> 590   0.66 2017-11-26 19:22:47
#> 591   0.65 2017-11-26 19:20:19
#> 592   0.70 2017-11-26 19:17:51
#> 593   0.69 2017-11-26 19:15:23
#> 594   0.72 2017-11-26 19:12:55
#> 595   0.68 2017-11-26 19:10:27
#> 596   0.74 2017-11-26 19:07:59
#> 597   0.79 2017-11-26 19:05:32
#> 598   0.78 2017-11-26 19:03:03
#> 599   0.78 2017-11-26 19:00:36
#> 600   0.83 2017-11-26 18:58:08
#> 601   0.78 2017-11-26 18:55:40
#> 602   0.80 2017-11-26 18:53:12
#> 603   0.82 2017-11-26 18:50:45
#> 604   0.88 2017-11-26 18:48:17
#> 605   0.80 2017-11-26 18:45:49
#> 606   0.79 2017-11-26 18:43:21
#> 607   0.76 2017-11-26 18:40:53
#> 608   0.81 2017-11-26 18:38:26
#> 609   0.79 2017-11-26 18:35:58
#> 610   0.74 2017-11-26 18:33:30
#> 611   0.77 2017-11-26 18:31:02
#> 612   0.83 2017-11-26 18:28:34
#> 613   0.84 2017-11-26 18:26:07
#> 614   0.88 2017-11-26 18:23:37
#> 615   0.95 2017-11-26 18:21:10
#> 616   0.94 2017-11-26 18:18:42
#> 617   0.94 2017-11-26 18:16:14
#> 618   0.86 2017-11-26 18:13:46
#> 619   0.80 2017-11-26 18:11:18
#> 620   0.85 2017-11-26 18:08:50
#> 621   0.92 2017-11-26 18:06:21
#> 622   0.98 2017-11-26 18:03:53
#> 623   0.92 2017-11-26 18:01:25
#> 624   0.93 2017-11-26 17:58:57
#> 625   0.89 2017-11-26 17:56:29
#> 626   0.90 2017-11-26 17:54:00
#> 627   0.88 2017-11-26 17:51:32
#> 628   0.98 2017-11-26 17:49:04
#> 629   0.88 2017-11-26 17:46:36
#> 630   0.99 2017-11-26 17:44:09
#> 631   0.96 2017-11-26 17:41:41
#> 632   0.87 2017-11-26 17:39:13
#> 633   0.85 2017-11-26 17:36:45
#> 634   0.80 2017-11-26 17:34:17
#> 635   0.69 2017-11-26 17:31:49
#> 636   0.81 2017-11-26 17:29:21
#> 637   0.83 2017-11-26 17:26:53
#> 638   0.74 2017-11-26 17:24:25
#> 639   0.68 2017-11-26 17:21:57
#> 640   0.64 2017-11-26 17:19:29
#> 641   0.69 2017-11-26 17:17:01
#> 642   0.71 2017-11-26 17:14:34
#> 643   0.75 2017-11-26 17:12:06
#> 644   0.88 2017-11-26 17:09:38
#> 645   0.81 2017-11-26 17:07:10
#> 646   0.78 2017-11-26 17:04:42
#> 647   0.75 2017-11-26 17:02:14
#> 648   0.66 2017-11-26 16:59:46
#> 649   0.62 2017-11-26 16:57:18
#> 650   0.82 2017-11-26 16:54:50
#> 651   0.83 2017-11-26 16:52:22
#> 652   0.87 2017-11-26 16:49:54
#> 653   0.88 2017-11-26 16:47:26
#> 654   0.90 2017-11-26 16:44:58
#> 655   1.00 2017-11-26 16:42:31
#> 656   1.11 2017-11-26 16:40:03
#> 657   1.09 2017-11-26 16:37:35
#> 658   1.10 2017-11-26 16:35:07
#> 659   1.11 2017-11-26 16:32:39
#> 660   1.09 2017-11-26 16:30:11
#> 661   1.10 2017-11-26 16:27:44
#> 662   1.07 2017-11-26 16:25:16
#> 663   1.10 2017-11-26 16:22:48
#> 664   1.10 2017-11-26 16:20:20
#> 665   1.04 2017-11-26 16:17:52
#> 666   1.03 2017-11-26 16:15:25
#> 667   1.07 2017-11-26 16:12:57
#> 668   1.09 2017-11-26 16:10:29
#> 669   1.16 2017-11-26 16:08:01
#> 670   1.29 2017-11-26 16:05:33
#> 671   1.43 2017-11-26 16:03:05
#> 672   1.41 2017-11-26 16:00:38
#> 673   1.30 2017-11-26 15:58:10
#> 674   1.14 2017-11-26 15:55:42
#> 675   1.23 2017-11-26 15:53:14
#> 676   1.24 2017-11-26 15:50:46
#> 677   1.28 2017-11-26 15:48:18
#> 678   1.41 2017-11-26 15:45:50
#> 679   1.56 2017-11-26 15:43:22
#> 680   1.70 2017-11-26 15:40:55
#> 681   1.42 2017-11-26 15:38:27
#> 682   1.45 2017-11-26 15:35:59
#> 683   1.49 2017-11-26 15:33:31
#> 684   1.45 2017-11-26 15:31:03
#> 685   1.45 2017-11-26 15:28:35
#> 686   1.61 2017-11-26 15:26:07
#> 687   1.61 2017-11-26 15:23:39
#> 688   1.63 2017-11-26 15:21:11
#> 689   1.65 2017-11-26 15:18:43
#> 690   1.69 2017-11-26 15:16:15
#> 691   1.63 2017-11-26 15:13:47
#> 692   1.64 2017-11-26 15:11:19
#> 693   1.61 2017-11-26 15:08:51
#> 694   1.63 2017-11-26 15:06:24
#> 695   1.65 2017-11-26 15:03:56
#> 696   1.78 2017-11-26 15:01:28
#> 697   1.89 2017-11-26 14:59:00
#> 698   1.71 2017-11-26 14:56:32
#> 699   1.68 2017-11-26 14:54:04
#> 700   1.83 2017-11-26 14:51:36
#> 701   1.98 2017-11-26 14:49:07
#> 702   1.93 2017-11-26 14:46:39
#> 703   1.85 2017-11-26 14:44:11
#> 704   1.71 2017-11-26 14:41:44
#> 705   1.75 2017-11-26 14:39:16
#> 706   1.80 2017-11-26 14:36:48
#> 707   2.03 2017-11-26 14:34:20
#> 708   1.90 2017-11-26 14:31:52
#> 709   1.81 2017-11-26 14:29:24
#> 710   1.82 2017-11-26 14:26:56
#> 711   1.92 2017-11-26 14:24:28
#> 712   2.01 2017-11-26 14:22:00
#> 713   1.98 2017-11-26 14:19:33
#> 714   1.93 2017-11-26 14:17:05
#> 715   2.01 2017-11-26 14:14:37
#> 716   1.85 2017-11-26 14:12:08
#> 717   1.84 2017-11-26 14:09:40
#> 718   1.93 2017-11-26 14:07:12
#> 719   1.86 2017-11-26 14:04:44
#> 720   1.79 2017-11-26 14:02:16
#> 721   1.93 2017-11-26 13:59:48
#> 722   1.88 2017-11-26 13:57:20
#> 723   1.97 2017-11-26 13:54:53
#> 724   2.01 2017-11-26 13:52:25
#> 725   1.91 2017-11-26 13:49:57
#> 726   1.83 2017-11-26 13:47:29
#> 727   1.98 2017-11-26 13:45:01
#> 728   1.87 2017-11-26 13:42:33
#> 729   1.85 2017-11-26 13:40:05
#> 730   1.75 2017-11-26 13:37:36
#> 731   2.09 2017-11-26 13:35:07
#> 732   2.15 2017-11-26 13:32:38
#> 733   2.26 2017-11-26 13:30:10
#> 734   2.27 2017-11-26 13:27:42
#> 735   2.36 2017-11-26 13:25:14
#> 736   2.42 2017-11-26 13:22:46
#> 737   2.37 2017-11-26 13:20:18
#> 738   2.31 2017-11-26 13:17:50
#> 739   2.29 2017-11-26 13:15:23
#> 740   2.46 2017-11-26 13:12:55
#> 741   2.29 2017-11-26 13:10:27
#> 742   2.37 2017-11-26 13:07:59
#> 743   2.37 2017-11-26 13:05:31
#> 744   2.18 2017-11-26 13:03:03
#> 745   2.33 2017-11-26 13:00:36
#> 746   2.31 2017-11-26 12:58:08
#> 747   2.22 2017-11-26 12:55:40
#> 748   2.23 2017-11-26 12:53:11
#> 749   2.33 2017-11-26 12:50:43
#> 750   2.28 2017-11-26 12:48:15
#> 751   2.23 2017-11-26 12:45:48
#> 752   2.10 2017-11-26 12:43:20
#> 753   2.08 2017-11-26 12:40:51
#> 754   2.00 2017-11-26 12:38:24
#> 755   2.18 2017-11-26 12:35:56
#> 756   2.00 2017-11-26 12:33:28
#> 757   2.19 2017-11-26 12:31:00
#> 758   2.24 2017-11-26 12:28:32
#> 759   2.10 2017-11-26 12:26:04
#> 760   2.30 2017-11-26 12:23:37
#> 761   2.15 2017-11-26 12:21:08
#> 762   2.29 2017-11-26 12:18:39
#> 763   2.11 2017-11-26 12:13:37
#> 764   2.11 2017-11-26 12:11:08
#> 765   2.16 2017-11-26 12:08:40
#> 766   2.08 2017-11-26 12:06:13
#> 767   1.94 2017-11-26 12:03:45
#> 768   1.94 2017-11-26 12:01:17
#> 769   2.14 2017-11-26 11:58:49
#> 770   1.95 2017-11-26 11:56:21
#> 771   1.96 2017-11-26 11:53:53
#> 772   1.98 2017-11-26 11:51:24
#> 773   1.84 2017-11-26 11:48:56
#> 774   1.83 2017-11-26 11:46:29
#> 775   1.91 2017-11-26 11:44:01
#> 776   1.93 2017-11-26 11:41:33
#> 777   1.95 2017-11-26 11:39:05
#> 778   1.98 2017-11-26 11:36:38
#> 779   1.95 2017-11-26 11:34:10
#> 780   2.13 2017-11-26 11:31:42
#> 781   2.19 2017-11-26 11:29:14
#> 782   2.04 2017-11-26 11:26:47
#> 783   2.22 2017-11-26 11:24:19
#> 784   2.13 2017-11-26 11:21:51
#> 785   2.22 2017-11-26 11:19:23
#> 786   1.79 2017-11-26 11:16:55
#> 787   1.95 2017-11-26 11:14:28
#> 788   2.09 2017-11-26 11:12:00
#> 789   2.31 2017-11-26 11:09:32
#> 790   2.30 2017-11-26 11:07:04
#> 791   2.29 2017-11-26 11:04:36
#> 792   2.12 2017-11-26 11:02:09
#> 793   2.22 2017-11-26 10:59:41
#> 794   2.21 2017-11-26 10:57:12
#> 795   2.22 2017-11-26 10:54:44
#> 796   2.30 2017-11-26 10:52:16
#> 797   2.48 2017-11-26 10:49:48
#> 798   2.53 2017-11-26 10:47:20
#> 799   2.54 2017-11-26 10:44:53
#> 800   2.70 2017-11-26 10:42:25
#> 801   2.81 2017-11-26 10:39:57
#> 802   2.93 2017-11-26 10:37:29
#> 803   3.04 2017-11-26 10:35:01
#> 804   3.02 2017-11-26 10:32:34
#> 805   2.97 2017-11-26 10:30:06
#> 806   2.94 2017-11-26 10:27:38
#> 807   3.03 2017-11-26 10:25:10
#> 808   3.05 2017-11-26 10:22:43
#> 809   3.01 2017-11-26 10:20:15
#> 810   3.04 2017-11-26 10:17:47
#> 811   2.98 2017-11-26 10:15:19
#> 812   3.02 2017-11-26 10:12:51
#> 813   3.04 2017-11-26 10:10:23
#> 814   3.03 2017-11-26 10:07:55
#> 815   2.97 2017-11-26 10:05:28
#> 816   2.93 2017-11-26 10:03:00
#> 817   2.92 2017-11-26 10:00:32
#> 818   3.02 2017-11-26 09:58:04
#> 819   2.94 2017-11-26 09:55:37
#> 820   3.07 2017-11-26 09:53:09
#> 821   3.13 2017-11-26 09:50:41
#> 822   3.12 2017-11-26 09:48:13
#> 823   3.08 2017-11-26 09:45:46
#> 824   3.10 2017-11-26 09:43:18
#> 825   3.19 2017-11-26 09:40:50
#> 826   3.18 2017-11-26 09:38:22
#> 827   3.15 2017-11-26 09:35:54
#> 828   3.35 2017-11-26 09:33:26
#> 829   3.38 2017-11-26 09:30:58
#> 830   3.40 2017-11-26 09:28:30
#> 831   3.41 2017-11-26 09:26:03
#> 832   3.46 2017-11-26 09:23:35
#> 833   3.41 2017-11-26 09:21:07
#> 834   3.36 2017-11-26 09:18:39
#> 835   3.33 2017-11-26 09:16:11
#> 836   3.38 2017-11-26 09:13:43
#> 837   3.31 2017-11-26 09:11:15
#> 838   3.29 2017-11-26 09:08:48
#> 839   3.29 2017-11-26 09:06:03
#> 840   3.16 2017-11-26 09:03:35
#> 841   3.06 2017-11-26 09:01:06
#> 842   2.86 2017-11-26 08:58:37
#> 843   2.89 2017-11-26 08:56:09
#> 844   3.20 2017-11-26 08:53:41
#> 845   3.39 2017-11-26 08:51:12
#> 846   3.36 2017-11-26 08:48:45
#> 847   3.32 2017-11-26 08:46:17
#> 848   3.18 2017-11-26 08:43:49
#> 849   3.23 2017-11-26 08:41:21
#> 850   3.27 2017-11-26 08:38:53
#> 851   2.97 2017-11-26 08:36:26
#> 852   2.94 2017-11-26 08:33:57
#> 853   2.91 2017-11-26 08:31:30
#> 854   2.92 2017-11-26 08:29:02
#> 855   2.94 2017-11-26 08:26:34
#> 856   2.87 2017-11-26 08:24:06
#> 857   2.95 2017-11-26 08:21:38
#> 858   2.91 2017-11-26 08:19:08
#> 859   2.83 2017-11-26 08:16:40
#> 860   2.80 2017-11-26 08:14:13
#> 861   2.81 2017-11-26 08:11:42
#> 862   2.82 2017-11-26 08:09:14
#> 863   2.84 2017-11-26 08:06:46
#> 864   2.70 2017-11-26 08:04:17
#> 865   2.62 2017-11-26 08:01:49
#> 866   2.70 2017-11-26 07:59:22
#> 867   2.67 2017-11-26 07:56:52
#> 868   2.68 2017-11-26 07:54:22
#> 869   3.02 2017-11-26 07:51:52
#> 870   2.77 2017-11-26 07:49:24
#> 871   2.73 2017-11-26 07:46:57
#> 872   2.78 2017-11-26 07:44:29
#> 873   2.80 2017-11-26 07:42:00
#> 874   2.78 2017-11-26 07:39:32
#> 875   2.80 2017-11-26 07:37:02
#> 876   2.72 2017-11-26 07:34:34
#> 877   2.69 2017-11-26 07:32:06
#> 878   2.69 2017-11-26 07:29:38
#> 879   2.71 2017-11-26 07:27:10
#> 880   2.56 2017-11-26 07:24:43
#> 881   2.61 2017-11-26 07:22:15
#> 882   2.55 2017-11-26 07:19:47
#> 883   2.53 2017-11-26 07:17:19
#> 884   2.45 2017-11-26 07:14:51
#> 885   2.40 2017-11-26 07:12:23
#> 886   2.47 2017-11-26 07:09:55
#> 887   2.31 2017-11-26 07:07:28
#> 888   2.42 2017-11-26 07:05:00
#> 889   2.49 2017-11-26 07:02:32
#> 890   2.42 2017-11-26 07:00:04
#> 891   2.38 2017-11-26 06:57:36
#> 892   2.28 2017-11-26 06:55:09
#> 893   2.37 2017-11-26 06:52:41
#> 894   2.40 2017-11-26 06:50:13
#> 895   2.44 2017-11-26 06:47:44
#> 896   2.63 2017-11-26 06:45:16
#> 897   2.78 2017-11-26 06:42:49
#> 898   2.87 2017-11-26 06:40:21
#> 899   2.98 2017-11-26 06:37:53
#> 900   2.67 2017-11-26 06:35:25
#> 901   2.74 2017-11-26 06:32:57
#> 902   2.83 2017-11-26 06:30:29
#> 903   2.71 2017-11-26 06:28:02
#> 904   2.70 2017-11-26 06:25:34
#> 905   2.69 2017-11-26 06:23:06
#> 906   2.59 2017-11-26 06:20:38
#> 907   2.59 2017-11-26 06:18:11
#> 908   2.61 2017-11-26 06:15:43
#> 909   2.66 2017-11-26 06:13:15
#> 910   2.69 2017-11-26 06:10:47
#> 911   2.59 2017-11-26 06:08:20
#> 912   2.54 2017-11-26 06:05:52
#> 913   2.65 2017-11-26 06:03:24
#> 914   2.70 2017-11-26 06:00:56
#> 915   2.73 2017-11-26 05:58:28
#> 916   2.70 2017-11-26 05:56:00
#> 917   2.71 2017-11-26 05:53:33
#> 918   2.79 2017-11-26 05:51:05
#> 919   2.85 2017-11-26 05:48:37
#> 920   2.79 2017-11-26 05:46:09
#> 921   2.75 2017-11-26 05:43:42
#> 922   2.82 2017-11-26 05:41:13
#> 923   2.84 2017-11-26 05:38:45
#> 924   2.91 2017-11-26 05:36:17
#> 925   2.96 2017-11-26 05:33:49
#> 926   3.01 2017-11-26 05:31:22
#> 927   2.94 2017-11-26 05:28:53
#> 928   2.98 2017-11-26 05:26:26
#> 929   2.92 2017-11-26 05:23:58
#> 930   3.03 2017-11-26 05:21:30
#> 931   3.09 2017-11-26 05:19:02
#> 932   2.98 2017-11-26 05:16:35
#> 933   2.91 2017-11-26 05:14:07
#> 934   2.95 2017-11-26 05:11:39
#> 935   2.99 2017-11-26 05:09:12
#> 936   3.03 2017-11-26 05:06:44
#> 937   3.09 2017-11-26 05:04:16
#> 938   3.02 2017-11-26 05:01:48
#> 939   2.93 2017-11-26 04:59:20
#> 940   2.91 2017-11-26 04:56:52
#> 941   2.80 2017-11-26 04:54:25
#> 942   2.80 2017-11-26 04:51:57
#> 943   2.89 2017-11-26 04:49:29
#> 944   2.83 2017-11-26 04:47:02
#> 945   2.94 2017-11-26 04:44:34
#> 946   2.93 2017-11-26 04:42:06
#> 947   3.05 2017-11-26 04:39:38
#> 948   3.07 2017-11-26 04:37:11
#> 949   3.13 2017-11-26 04:34:43
#> 950   3.20 2017-11-26 04:32:15
#> 951   3.13 2017-11-26 04:29:47
#> 952   3.16 2017-11-26 04:27:20
#> 953   3.08 2017-11-26 04:24:52
#> 954   3.19 2017-11-26 04:22:24
#> 955   3.23 2017-11-26 04:19:56
#> 956   3.22 2017-11-26 04:17:28
#> 957   3.23 2017-11-26 04:15:00
#> 958   3.26 2017-11-26 04:12:33
#> 959   3.22 2017-11-26 04:10:05
#> 960   3.19 2017-11-26 04:07:37
#> 961   3.17 2017-11-26 04:05:09
#> 962   3.15 2017-11-26 04:02:42
#> 963   3.16 2017-11-26 04:00:14
#> 964   3.22 2017-11-26 03:57:46
#> 965   3.23 2017-11-26 03:55:18
#> 966   3.14 2017-11-26 03:52:51
#> 967   3.10 2017-11-26 03:50:23
#> 968   3.09 2017-11-26 03:47:55
#> 969   3.11 2017-11-26 03:45:27
#> 970   3.11 2017-11-26 03:43:00
#> 971   3.15 2017-11-26 03:40:32
#> 972   3.17 2017-11-26 03:38:04
#> 973   3.16 2017-11-26 03:35:36
#> 974   3.21 2017-11-26 03:33:08
#> 975   3.21 2017-11-26 03:30:41
#> 976   3.23 2017-11-26 03:28:13
#> 977   3.26 2017-11-26 03:25:45
#> 978   3.23 2017-11-26 03:23:17
#> 979   3.21 2017-11-26 03:20:49
#> 980   3.23 2017-11-26 03:18:22
#> 981   3.29 2017-11-26 03:15:54
#> 982   3.22 2017-11-26 03:13:26
#> 983   3.25 2017-11-26 03:10:59
#> 984   3.20 2017-11-26 03:08:31
#> 985   3.18 2017-11-26 03:06:03
#> 986   3.22 2017-11-26 03:03:35
#> 987   3.15 2017-11-26 03:01:08
#> 988   3.17 2017-11-26 02:58:40
#> 989   3.17 2017-11-26 02:56:12
#> 990   3.17 2017-11-26 02:53:44
#> 991   3.16 2017-11-26 02:51:17
#> 992   3.24 2017-11-26 02:48:49
#> 993   3.17 2017-11-26 02:46:21
#> 994   3.22 2017-11-26 02:43:53
#> 995   3.29 2017-11-26 02:41:26
#> 996   3.38 2017-11-26 02:38:58
#> 997   3.41 2017-11-26 02:36:30
#> 998   3.41 2017-11-26 02:34:02
#> 999   3.39 2017-11-26 02:31:23
#> 1000  3.40 2017-11-26 02:28:43
#> 1001  3.44 2017-11-26 02:26:16
#> 1002  3.50 2017-11-26 02:23:48
#> 1003  3.44 2017-11-26 02:21:20
#> 1004  3.50 2017-11-26 02:18:52
#> 1005  3.51 2017-11-26 02:16:25
#> 1006  3.47 2017-11-26 02:13:57
#> 1007  3.44 2017-11-26 02:11:29
#> 1008  3.48 2017-11-26 02:09:01
#> 1009  3.51 2017-11-26 02:06:33
#> 1010  3.53 2017-11-26 02:04:06
#> 1011  3.53 2017-11-26 02:01:38
#> 1012  3.55 2017-11-26 01:59:10
#> 1013  3.59 2017-11-26 01:56:42
#> 1014  3.63 2017-11-26 01:54:15
#> 1015  3.61 2017-11-26 01:51:47
#> 1016  3.59 2017-11-26 01:49:19
#> 1017  3.59 2017-11-26 01:46:51
#> 1018  3.62 2017-11-26 01:44:24
#> 1019  3.60 2017-11-26 01:41:56
#> 1020  3.58 2017-11-26 01:39:28
#> 1021  3.53 2017-11-26 01:37:00
#> 1022  3.52 2017-11-26 01:34:32
#> 1023  3.53 2017-11-26 01:32:05
#> 1024  3.58 2017-11-26 01:29:37
#> 1025  3.59 2017-11-26 01:27:09
#> 1026  3.62 2017-11-26 01:24:41
#> 1027  3.74 2017-11-26 01:22:14
#> 1028  3.70 2017-11-26 01:19:46
#> 1029  3.62 2017-11-26 01:17:18
#> 1030  3.65 2017-11-26 01:14:50
#> 1031  3.57 2017-11-26 01:12:23
#> 1032  3.58 2017-11-26 01:09:55
#> 1033  3.64 2017-11-26 01:07:27
#> 1034  3.67 2017-11-26 01:05:00
#> 1035  3.64 2017-11-26 01:02:32
#> 1036  3.62 2017-11-26 01:00:04
#> 1037  3.65 2017-11-26 00:57:36
#> 1038  3.62 2017-11-26 00:55:09
#> 1039  3.61 2017-11-26 00:52:41
#> 1040  3.61 2017-11-26 00:50:13
#> 1041  3.63 2017-11-26 00:47:45
#> 1042  3.65 2017-11-26 00:45:18
#> 1043  3.69 2017-11-26 00:42:50
#> 1044  3.69 2017-11-26 00:40:22
#> 1045  3.68 2017-11-26 00:37:54
#> 1046  3.70 2017-11-26 00:35:27
#> 1047  3.70 2017-11-26 00:32:58
#> 1048  3.70 2017-11-26 00:30:31
#> 1049  3.73 2017-11-26 00:28:03
#> 1050  3.67 2017-11-26 00:25:35
#> 1051  3.75 2017-11-26 00:23:07
#> 1052  3.69 2017-11-26 00:20:40
#> 1053  3.70 2017-11-26 00:18:12
#> 1054  3.72 2017-11-26 00:15:44
#> 1055  3.76 2017-11-26 00:13:16
#> 1056  3.75 2017-11-26 00:10:49
#> 1057  3.72 2017-11-26 00:08:21
#> 1058  3.73 2017-11-26 00:05:53
#> 1059  3.71 2017-11-26 00:03:25
#> 1060  3.74 2017-11-26 00:00:58
#> 1061  3.79 2017-11-25 23:58:30
#> 1062  3.79 2017-11-25 23:56:02
#> 1063  3.80 2017-11-25 23:53:34
#> 1064  3.81 2017-11-25 23:51:07
#> 1065  3.73 2017-11-25 23:48:39
#> 1066  3.74 2017-11-25 23:46:11
#> 1067  3.78 2017-11-25 23:43:43
#> 1068  3.75 2017-11-25 23:41:16
#> 1069  3.72 2017-11-25 23:38:48
#> 1070  3.70 2017-11-25 23:36:20
#> 1071  3.70 2017-11-25 23:33:52
#> 1072  3.65 2017-11-25 23:31:24
#> 1073  3.67 2017-11-25 23:28:57
#> 1074  3.64 2017-11-25 23:26:29
#> 1075  3.51 2017-11-25 23:24:01
#> 1076  3.64 2017-11-25 23:21:33
#> 1077  3.71 2017-11-25 23:19:05
#> 1078  3.73 2017-11-25 23:16:38
#> 1079  3.75 2017-11-25 23:14:10
#> 1080  3.74 2017-11-25 23:11:42
#> 1081  3.82 2017-11-25 23:09:14
#> 1082  3.72 2017-11-25 23:06:46
#> 1083  3.71 2017-11-25 23:04:18
#> 1084  3.75 2017-11-25 23:01:51
#> 1085  3.82 2017-11-25 22:59:23
#> 1086  3.81 2017-11-25 22:56:55
#> 1087  3.81 2017-11-25 22:54:27
#> 1088  3.81 2017-11-25 22:52:00
#> 1089  3.90 2017-11-25 22:49:31
#> 1090  3.97 2017-11-25 22:47:04
#> 1091  3.96 2017-11-25 22:44:36
#> 1092  3.91 2017-11-25 22:42:08
#> 1093  3.90 2017-11-25 22:39:40
#> 1094  3.88 2017-11-25 22:37:12
#> 1095  3.84 2017-11-25 22:34:45
#> 1096  3.90 2017-11-25 22:32:17
#> 1097  4.02 2017-11-25 22:29:48
#> 1098  3.97 2017-11-25 22:27:20
#> 1099  3.86 2017-11-25 22:24:52
#> 1100  3.84 2017-11-25 22:22:24
#> 1101  3.82 2017-11-25 22:19:56
#> 1102  3.79 2017-11-25 22:17:28
#> 1103  3.70 2017-11-25 22:15:00
#> 1104  3.63 2017-11-25 22:12:33
#> 1105  3.72 2017-11-25 22:10:05
#> 1106  3.79 2017-11-25 22:07:37
#> 1107  3.86 2017-11-25 22:05:09
#> 1108  3.91 2017-11-25 22:02:42
#> 1109  4.04 2017-11-25 22:00:14
#> 1110  4.11 2017-11-25 21:57:46
#> 1111  4.16 2017-11-25 21:55:18
#> 1112  4.21 2017-11-25 21:52:51
#> 1113  4.23 2017-11-25 21:50:23
#> 1114  4.11 2017-11-25 21:47:55
#> 1115  4.05 2017-11-25 21:45:27
#> 1116  4.03 2017-11-25 21:42:59
#> 1117  4.04 2017-11-25 21:40:32
#> 1118  4.02 2017-11-25 21:38:04
#> 1119  3.94 2017-11-25 21:35:36
#> 1120  3.96 2017-11-25 21:33:08
#> 1121  3.97 2017-11-25 21:30:41
#> 1122  3.97 2017-11-25 21:28:13
#> 1123  4.01 2017-11-25 21:25:45
#> 1124  4.04 2017-11-25 21:23:17
#> 1125  4.07 2017-11-25 21:20:50
#> 1126  4.04 2017-11-25 21:18:22
#> 1127  4.09 2017-11-25 21:15:54
#> 1128  4.12 2017-11-25 21:13:26
#> 1129  4.14 2017-11-25 21:10:59
#> 1130  4.14 2017-11-25 21:08:29
#> 1131  4.14 2017-11-25 21:06:02
#> 1132  4.17 2017-11-25 21:03:33
#> 1133  4.24 2017-11-25 21:01:05
#> 1134  4.38 2017-11-25 20:58:37
#> 1135  4.37 2017-11-25 20:56:09
#> 1136  4.40 2017-11-25 20:53:42
#> 1137  4.41 2017-11-25 20:51:14
#> 1138  4.36 2017-11-25 20:48:46
#> 1139  4.34 2017-11-25 20:46:18
#> 1140  4.34 2017-11-25 20:43:51
#> 1141  4.35 2017-11-25 20:41:23
#> 1142  4.38 2017-11-25 20:38:55
#> 1143  4.38 2017-11-25 20:36:28
#> 1144  4.48 2017-11-25 20:34:00
#> 1145  4.48 2017-11-25 20:31:32
#> 1146  4.57 2017-11-25 20:29:04
#> 1147  4.56 2017-11-25 20:26:37
#> 1148  4.56 2017-11-25 20:24:09
#> 1149  4.64 2017-11-25 20:21:41
#> 1150  4.65 2017-11-25 20:19:13
#> 1151  4.59 2017-11-25 20:16:45
#> 1152  4.35 2017-11-25 20:14:18
#> 1153  4.25 2017-11-25 20:11:50
#> 1154  4.33 2017-11-25 20:09:22
#> 1155  4.32 2017-11-25 20:06:54
#> 1156  4.20 2017-11-25 20:04:27
#> 1157  4.17 2017-11-25 20:01:59
#> 1158  4.27 2017-11-25 19:59:31
#> 1159  4.28 2017-11-25 19:57:03
#> 1160  4.33 2017-11-25 19:54:35
#> 1161  4.33 2017-11-25 19:52:07
#> 1162  4.33 2017-11-25 19:49:38
#> 1163  4.34 2017-11-25 19:47:10
#> 1164  4.38 2017-11-25 19:44:43
#> 1165  4.34 2017-11-25 19:42:15
#> 1166  4.32 2017-11-25 19:39:47
#> 1167  4.30 2017-11-25 19:37:19
#> 
#> 

# NOT RUN {
get_senseBox_data(
  senseBoxId = c("5957b67494f0520011304cc1","5957b67494f0520011304cc1") ,
  sensorId = "all",
  fromDate = "2017-11-25-12:00:00",
  toDate = "2017-11-26-12:00:00")

get_senseBox_data(
  senseBoxId = "5957b67494f0520011304cc1",
  sensorId = "5957b67494f0520011304cc4",
  fromDate = "2017-11-25-12:00:00",
  toDate = "2017-11-26-12:00:00")
# }