Get data from a senseBox

If the arguments fromDate and toDate are NOT given, the measurements from the last 48 h from now are downloaded. The maximum numbers of downloaded measurements are 10,000. The maximum time frame is one month. It is possible to download the data directly as a csv file, when argument CSV is TRUE.

get_senseBox_data(senseBoxId, sensorId = rep("all", times =
  length(unlist(senseBoxId))), fromDate = NULL, toDate = NULL,
  parallel = FALSE, CSV = FALSE, POSIXct = TRUE)

Arguments

senseBoxId
characteror vector of character (required): senseBoxId or a vector of character of senseBoxIds

sensorId
character or list of character (optional): sensorId or a list of senseBoxIds. When no sensorIds are submitted, all sensorIds from this sensebox are requested.

fromDate
character (optional): Just show data with from date in format YYYY-mm-dd HH:MM:SS

toDate
character (optional): Just show data with up to date in format YYYY-mm-dd HH:MM:SS

parallel
logical (optional): Should the calculations be executed on multiple cores? At least 4 cores are necessary to use this feature.

CSV
logical (optional): Download data as csv file? NOT SUPPORTED UNTIL NOW!

POSIXct
logical (optional): Should the timestamp be translated into POSIXct?

Value

A list with every entry is a sensBoxId. Every list entry inherits a data.frame with values and dates.

Function version

0.0.1

Examples

##=====================================
## Example: get data from sensorId
##=====================================

get_senseBox_data(
  senseBoxId = "5957b67494f0520011304cc1",
  sensorId = "5957b67494f0520011304cc4")

#> $`5957b67494f0520011304cc1`
#> $`5957b67494f0520011304cc1`$`Temperatur [°C]`
#>      value           createdAt
#> 1    17.76 2018-09-08 17:37:41
#> 2    18.00 2018-09-08 17:35:12
#> 3    17.95 2018-09-08 17:32:42
#> 4    17.99 2018-09-08 17:30:12
#> 5    18.12 2018-09-08 17:27:42
#> 6    18.11 2018-09-08 17:25:13
#> 7    18.26 2018-09-08 17:22:43
#> 8    18.29 2018-09-08 17:20:14
#> 9    18.41 2018-09-08 17:17:44
#> 10   18.39 2018-09-08 17:15:14
#> 11   18.40 2018-09-08 17:12:45
#> 12   18.42 2018-09-08 17:10:15
#> 13   18.45 2018-09-08 17:07:45
#> 14   18.64 2018-09-08 17:05:16
#> 15   18.89 2018-09-08 17:02:47
#> 16   18.81 2018-09-08 17:00:17
#> 17   18.85 2018-09-08 16:57:47
#> 18   19.18 2018-09-08 16:55:17
#> 19   19.03 2018-09-08 16:52:48
#> 20   18.87 2018-09-08 16:50:18
#> 21   19.01 2018-09-08 16:47:48
#> 22   19.12 2018-09-08 16:45:19
#> 23   19.05 2018-09-08 16:42:50
#> 24   19.13 2018-09-08 16:40:20
#> 25   19.32 2018-09-08 16:37:50
#> 26   19.27 2018-09-08 16:35:21
#> 27   19.23 2018-09-08 16:32:52
#> 28   19.51 2018-09-08 16:30:22
#> 29   19.61 2018-09-08 16:27:52
#> 30   19.72 2018-09-08 16:25:21
#> 31   19.82 2018-09-08 16:22:49
#> 32   19.99 2018-09-08 16:20:20
#> 33   20.18 2018-09-08 16:17:50
#> 34   20.20 2018-09-08 16:15:20
#> 35   20.16 2018-09-08 16:12:51
#> 36   20.19 2018-09-08 16:10:21
#> 37   20.19 2018-09-08 16:07:51
#> 38   20.49 2018-09-08 16:05:22
#> 39   20.53 2018-09-08 16:02:44
#> 40   20.81 2018-09-08 16:00:14
#> 41   20.90 2018-09-08 15:57:45
#> 42   20.95 2018-09-08 15:55:15
#> 43   21.18 2018-09-08 15:52:45
#> 44   21.46 2018-09-08 15:50:16
#> 45   21.60 2018-09-08 15:47:47
#> 46   21.52 2018-09-08 15:45:17
#> 47   21.84 2018-09-08 15:42:48
#> 48   21.80 2018-09-08 15:40:19
#> 49   22.22 2018-09-08 15:37:49
#> 50   22.55 2018-09-08 15:35:20
#> 51   22.58 2018-09-08 15:32:50
#> 52   22.52 2018-09-08 15:30:21
#> 53   22.58 2018-09-08 15:27:51
#> 54   22.60 2018-09-08 15:25:21
#> 55   22.78 2018-09-08 15:22:52
#> 56   22.68 2018-09-08 15:20:23
#> 57   22.91 2018-09-08 15:17:53
#> 58   22.61 2018-09-08 15:15:24
#> 59   22.88 2018-09-08 15:12:54
#> 60   22.82 2018-09-08 15:10:25
#> 61   22.78 2018-09-08 15:07:55
#> 62   22.57 2018-09-08 15:05:25
#> 63   22.50 2018-09-08 15:02:56
#> 64   22.73 2018-09-08 15:00:26
#> 65   22.82 2018-09-08 14:57:57
#> 66   22.67 2018-09-08 14:55:27
#> 67   22.85 2018-09-08 14:52:58
#> 68   22.80 2018-09-08 14:50:29
#> 69   22.84 2018-09-08 14:47:59
#> 70   23.16 2018-09-08 14:45:30
#> 71   23.11 2018-09-08 14:43:01
#> 72   23.05 2018-09-08 14:40:32
#> 73   23.12 2018-09-08 14:38:03
#> 74   22.76 2018-09-08 14:35:34
#> 75   22.99 2018-09-08 14:33:04
#> 76   23.16 2018-09-08 14:30:35
#> 77   23.19 2018-09-08 14:28:05
#> 78   22.93 2018-09-08 14:25:36
#> 79   23.47 2018-09-08 14:23:06
#> 80   23.29 2018-09-08 14:20:36
#> 81   23.38 2018-09-08 14:18:06
#> 82   23.49 2018-09-08 14:15:37
#> 83   23.52 2018-09-08 14:13:07
#> 84   22.91 2018-09-08 14:10:38
#> 85   23.25 2018-09-08 14:08:08
#> 86   23.41 2018-09-08 14:05:36
#> 87   23.54 2018-09-08 14:03:07
#> 88   23.49 2018-09-08 14:00:37
#> 89   22.92 2018-09-08 13:58:08
#> 90   22.93 2018-09-08 13:55:39
#> 91   23.06 2018-09-08 13:53:09
#> 92   23.06 2018-09-08 13:50:40
#> 93   23.17 2018-09-08 13:48:10
#> 94   23.32 2018-09-08 13:45:41
#> 95   23.49 2018-09-08 13:43:11
#> 96   23.33 2018-09-08 13:40:41
#> 97   23.06 2018-09-08 13:38:11
#> 98   23.50 2018-09-08 13:35:42
#> 99   23.24 2018-09-08 13:33:13
#> 100  23.26 2018-09-08 13:30:43
#> 101  23.07 2018-09-08 13:28:13
#> 102  22.66 2018-09-08 13:25:43
#> 103  23.05 2018-09-08 13:23:14
#> 104  23.00 2018-09-08 13:20:44
#> 105  23.38 2018-09-08 13:18:15
#> 106  23.51 2018-09-08 13:15:46
#> 107  23.62 2018-09-08 13:13:16
#> 108  23.52 2018-09-08 13:10:47
#> 109  23.66 2018-09-08 13:08:18
#> 110  23.43 2018-09-08 13:05:48
#> 111  23.67 2018-09-08 13:03:19
#> 112  23.79 2018-09-08 13:00:49
#> 113  23.24 2018-09-08 12:58:20
#> 114  23.28 2018-09-08 12:55:50
#> 115  23.24 2018-09-08 12:53:21
#> 116  22.82 2018-09-08 12:50:51
#> 117  23.03 2018-09-08 12:48:21
#> 118  22.97 2018-09-08 12:45:52
#> 119  23.49 2018-09-08 12:43:22
#> 120  23.34 2018-09-08 12:40:53
#> 121  23.49 2018-09-08 12:38:23
#> 122  23.28 2018-09-08 12:35:54
#> 123  23.38 2018-09-08 12:33:24
#> 124  23.17 2018-09-08 12:30:55
#> 125  23.21 2018-09-08 12:28:24
#> 126  23.26 2018-09-08 12:25:55
#> 127  23.10 2018-09-08 12:23:24
#> 128  23.45 2018-09-08 12:20:55
#> 129  23.81 2018-09-08 12:18:25
#> 130  23.50 2018-09-08 12:15:56
#> 131  23.13 2018-09-08 12:13:26
#> 132  23.22 2018-09-08 12:10:57
#> 133  23.39 2018-09-08 12:08:28
#> 134  23.19 2018-09-08 12:05:58
#> 135  22.71 2018-09-08 12:03:29
#> 136  22.88 2018-09-08 12:00:59
#> 137  22.97 2018-09-08 11:58:30
#> 138  23.08 2018-09-08 11:56:01
#> 139  22.49 2018-09-08 11:53:32
#> 140  23.32 2018-09-08 11:51:01
#> 141  23.81 2018-09-08 11:48:32
#> 142  23.88 2018-09-08 11:46:01
#> 143  23.60 2018-09-08 11:43:32
#> 144  22.97 2018-09-08 11:41:02
#> 145  22.81 2018-09-08 11:38:33
#> 146  23.27 2018-09-08 11:36:03
#> 147  22.63 2018-09-08 11:33:34
#> 148  22.48 2018-09-08 11:31:05
#> 149  22.46 2018-09-08 11:28:35
#> 150  22.81 2018-09-08 11:26:06
#> 151  23.23 2018-09-08 11:23:36
#> 152  22.62 2018-09-08 11:21:07
#> 153  22.98 2018-09-08 11:18:37
#> 154  23.30 2018-09-08 11:16:08
#> 155  22.54 2018-09-08 11:13:38
#> 156  22.53 2018-09-08 11:11:08
#> 157  22.36 2018-09-08 11:08:39
#> 158  22.29 2018-09-08 11:06:10
#> 159  22.61 2018-09-08 11:03:40
#> 160  22.15 2018-09-08 11:01:10
#> 161  21.37 2018-09-08 10:58:40
#> 162  21.48 2018-09-08 10:56:11
#> 163  22.07 2018-09-08 10:53:41
#> 164  22.05 2018-09-08 10:49:05
#> 165  22.60 2018-09-08 10:46:35
#> 166  22.08 2018-09-08 10:44:05
#> 167  22.38 2018-09-08 10:41:35
#> 168  22.24 2018-09-08 10:39:06
#> 169  21.69 2018-09-08 10:36:35
#> 170  22.09 2018-09-08 10:34:06
#> 171  22.71 2018-09-08 10:31:36
#> 172  22.29 2018-09-08 10:29:07
#> 173  22.00 2018-09-08 10:26:38
#> 174  21.99 2018-09-08 10:24:09
#> 175  21.38 2018-09-08 10:21:39
#> 176  21.20 2018-09-08 10:19:10
#> 177  21.27 2018-09-08 10:16:41
#> 178  21.03 2018-09-08 10:14:11
#> 179  20.98 2018-09-08 10:11:42
#> 180  21.00 2018-09-08 10:09:13
#> 181  20.90 2018-09-08 10:06:44
#> 182  21.15 2018-09-08 10:04:13
#> 183  21.29 2018-09-08 10:01:44
#> 184  21.28 2018-09-08 09:59:12
#> 185  21.40 2018-09-08 09:56:43
#> 186  21.61 2018-09-08 09:54:14
#> 187  21.69 2018-09-08 09:51:44
#> 188  21.44 2018-09-08 09:49:15
#> 189  21.37 2018-09-08 09:46:45
#> 190  21.66 2018-09-08 09:44:16
#> 191  21.67 2018-09-08 09:41:46
#> 192  21.44 2018-09-08 09:39:17
#> 193  21.34 2018-09-08 09:36:47
#> 194  21.20 2018-09-08 09:34:18
#> 195  20.93 2018-09-08 09:31:48
#> 196  20.25 2018-09-08 09:29:19
#> 197  20.13 2018-09-08 09:26:49
#> 198  20.12 2018-09-08 09:24:20
#> 199  20.09 2018-09-08 09:21:29
#> 200  20.01 2018-09-08 09:18:49
#> 201  20.22 2018-09-08 09:16:08
#> 202  20.19 2018-09-08 09:13:39
#> 203  19.97 2018-09-08 09:11:09
#> 204  20.09 2018-09-08 09:08:39
#> 205  20.47 2018-09-08 09:06:08
#> 206  20.38 2018-09-08 09:03:39
#> 207  20.81 2018-09-08 09:01:10
#> 208  20.90 2018-09-08 08:58:40
#> 209  20.56 2018-09-08 08:56:10
#> 210  19.88 2018-09-08 08:53:40
#> 211  20.02 2018-09-08 08:48:54
#> 212  20.16 2018-09-08 08:46:25
#> 213  20.31 2018-09-08 08:43:55
#> 214  20.01 2018-09-08 08:41:26
#> 215  19.62 2018-09-08 08:38:56
#> 216  19.70 2018-09-08 08:36:27
#> 217  19.28 2018-09-08 08:33:56
#> 218  18.79 2018-09-08 08:31:27
#> 219  18.14 2018-09-08 08:28:57
#> 220  18.27 2018-09-08 08:26:27
#> 221  18.39 2018-09-08 08:23:58
#> 222  18.31 2018-09-08 08:21:29
#> 223  18.04 2018-09-08 08:18:59
#> 224  18.03 2018-09-08 08:16:30
#> 225  18.13 2018-09-08 08:14:00
#> 226  18.11 2018-09-08 08:11:31
#> 227  17.86 2018-09-08 08:09:01
#> 228  17.37 2018-09-08 08:06:32
#> 229  17.18 2018-09-08 08:04:02
#> 230  17.14 2018-09-08 08:01:32
#> 231  17.12 2018-09-08 07:59:02
#> 232  17.01 2018-09-08 07:56:33
#> 233  16.91 2018-09-08 07:54:03
#> 234  16.74 2018-09-08 07:51:34
#> 235  16.59 2018-09-08 07:49:04
#> 236  16.44 2018-09-08 07:46:35
#> 237  16.34 2018-09-08 07:44:05
#> 238  16.33 2018-09-08 07:41:36
#> 239  16.36 2018-09-08 07:39:07
#> 240  16.44 2018-09-08 07:36:35
#> 241  16.27 2018-09-08 07:34:05
#> 242  16.31 2018-09-08 07:31:36
#> 243  16.20 2018-09-08 07:29:06
#> 244  16.17 2018-09-08 07:26:37
#> 245  16.13 2018-09-08 07:24:07
#> 246  16.09 2018-09-08 07:21:38
#> 247  16.05 2018-09-08 07:19:08
#> 248  15.98 2018-09-08 07:16:38
#> 249  15.89 2018-09-08 07:14:09
#> 250  15.86 2018-09-08 07:11:40
#> 251  15.89 2018-09-08 07:09:11
#> 252  15.83 2018-09-08 07:06:40
#> 253  15.87 2018-09-08 07:04:11
#> 254  15.80 2018-09-08 07:01:41
#> 255  15.76 2018-09-08 06:59:11
#> 256  15.70 2018-09-08 06:56:42
#> 257  15.71 2018-09-08 06:54:13
#> 258  15.77 2018-09-08 06:51:43
#> 259  15.56 2018-09-08 06:49:13
#> 260  15.49 2018-09-08 06:46:43
#> 261  15.31 2018-09-08 06:44:13
#> 262  15.36 2018-09-08 06:41:43
#> 263  15.24 2018-09-08 06:39:14
#> 264  15.14 2018-09-08 06:36:43
#> 265  15.20 2018-09-08 06:34:13
#> 266  15.18 2018-09-08 06:31:44
#> 267  15.09 2018-09-08 06:29:14
#> 268  15.01 2018-09-08 06:26:45
#> 269  15.01 2018-09-08 06:24:15
#> 270  15.07 2018-09-08 06:21:45
#> 271  14.99 2018-09-08 06:19:16
#> 272  14.82 2018-09-08 06:16:47
#> 273  14.84 2018-09-08 06:14:17
#> 274  14.97 2018-09-08 06:11:48
#> 275  14.97 2018-09-08 06:09:18
#> 276  14.90 2018-09-08 06:06:49
#> 277  14.85 2018-09-08 06:04:19
#> 278  14.75 2018-09-08 06:01:50
#> 279  14.71 2018-09-08 05:59:20
#> 280  14.65 2018-09-08 05:56:51
#> 281  14.62 2018-09-08 05:54:21
#> 282  14.57 2018-09-08 05:51:52
#> 283  14.46 2018-09-08 05:49:22
#> 284  14.40 2018-09-08 05:46:52
#> 285  14.36 2018-09-08 05:44:23
#> 286  14.38 2018-09-08 05:41:53
#> 287  14.36 2018-09-08 05:39:24
#> 288  14.30 2018-09-08 05:36:53
#> 289  14.34 2018-09-08 05:34:24
#> 290  14.29 2018-09-08 05:31:54
#> 291  14.28 2018-09-08 05:29:25
#> 292  14.27 2018-09-08 05:26:55
#> 293  14.31 2018-09-08 05:24:26
#> 294  14.33 2018-09-08 05:21:56
#> 295  14.36 2018-09-08 05:19:27
#> 296  14.35 2018-09-08 05:16:58
#> 297  14.35 2018-09-08 05:14:28
#> 298  14.35 2018-09-08 05:11:59
#> 299  14.29 2018-09-08 05:09:30
#> 300  14.29 2018-09-08 05:07:00
#> 301  14.25 2018-09-08 05:04:31
#> 302  14.20 2018-09-08 05:02:01
#> 303  14.23 2018-09-08 04:59:32
#> 304  14.22 2018-09-08 04:57:02
#> 305  14.22 2018-09-08 04:54:33
#> 306  14.23 2018-09-08 04:52:03
#> 307  14.20 2018-09-08 04:49:34
#> 308  14.16 2018-09-08 04:47:04
#> 309  14.16 2018-09-08 04:44:35
#> 310  14.21 2018-09-08 04:42:05
#> 311  14.20 2018-09-08 04:39:36
#> 312  14.10 2018-09-08 04:37:07
#> 313  14.13 2018-09-08 04:34:37
#> 314  14.13 2018-09-08 04:32:08
#> 315  14.14 2018-09-08 04:29:39
#> 316  14.13 2018-09-08 04:27:09
#> 317  14.15 2018-09-08 04:24:40
#> 318  14.20 2018-09-08 04:22:11
#> 319  14.22 2018-09-08 04:19:41
#> 320  14.20 2018-09-08 04:17:11
#> 321  14.21 2018-09-08 04:14:41
#> 322  14.18 2018-09-08 04:12:11
#> 323  14.25 2018-09-08 04:09:41
#> 324  14.26 2018-09-08 04:07:12
#> 325  14.20 2018-09-08 04:04:42
#> 326  14.20 2018-09-08 04:02:13
#> 327  14.17 2018-09-08 03:59:45
#> 328  14.15 2018-09-08 03:57:17
#> 329  14.12 2018-09-08 03:54:50
#> 330  14.11 2018-09-08 03:52:22
#> 331  14.12 2018-09-08 03:49:54
#> 332  14.08 2018-09-08 03:47:26
#> 333  14.06 2018-09-08 03:44:58
#> 334  14.10 2018-09-08 03:42:30
#> 335  14.10 2018-09-08 03:40:02
#> 336  14.09 2018-09-08 03:37:35
#> 337  14.06 2018-09-08 03:35:07
#> 338  14.01 2018-09-08 03:32:39
#> 339  14.02 2018-09-08 03:30:11
#> 340  14.07 2018-09-08 03:27:44
#> 341  14.04 2018-09-08 03:25:16
#> 342  13.96 2018-09-08 03:22:48
#> 343  14.00 2018-09-08 03:20:21
#> 344  14.07 2018-09-08 03:17:53
#> 345  14.08 2018-09-08 03:15:25
#> 346  14.08 2018-09-08 03:12:58
#> 347  14.20 2018-09-08 03:10:30
#> 348  14.23 2018-09-08 03:08:02
#> 349  14.21 2018-09-08 03:05:34
#> 350  14.16 2018-09-08 03:03:07
#> 351  14.19 2018-09-08 03:00:39
#> 352  14.16 2018-09-08 02:58:11
#> 353  14.13 2018-09-08 02:55:43
#> 354  14.10 2018-09-08 02:53:15
#> 355  14.09 2018-09-08 02:50:48
#> 356  14.16 2018-09-08 02:48:20
#> 357  14.16 2018-09-08 02:45:50
#> 358  14.20 2018-09-08 02:43:20
#> 359  14.26 2018-09-08 02:40:51
#> 360  14.32 2018-09-08 02:38:21
#> 361  14.37 2018-09-08 02:35:51
#> 362  14.38 2018-09-08 02:33:22
#> 363  14.40 2018-09-08 02:30:52
#> 364  14.37 2018-09-08 02:28:22
#> 365  14.37 2018-09-08 02:25:53
#> 366  14.36 2018-09-08 02:23:22
#> 367  14.38 2018-09-08 02:20:53
#> 368  14.36 2018-09-08 02:18:23
#> 369  14.39 2018-09-08 02:15:53
#> 370  14.39 2018-09-08 02:13:24
#> 371  14.40 2018-09-08 02:10:54
#> 372  14.41 2018-09-08 02:08:25
#> 373  14.45 2018-09-08 02:05:54
#> 374  14.39 2018-09-08 02:03:25
#> 375  14.40 2018-09-08 02:00:55
#> 376  14.41 2018-09-08 01:58:26
#> 377  14.40 2018-09-08 01:55:55
#> 378  14.40 2018-09-08 01:53:26
#> 379  14.41 2018-09-08 01:50:56
#> 380  14.42 2018-09-08 01:48:27
#> 381  14.47 2018-09-08 01:45:55
#> 382  14.48 2018-09-08 01:43:25
#> 383  14.48 2018-09-08 01:40:56
#> 384  14.46 2018-09-08 01:38:26
#> 385  14.41 2018-09-08 01:35:56
#> 386  14.42 2018-09-08 01:33:27
#> 387  14.43 2018-09-08 01:30:57
#> 388  14.42 2018-09-08 01:28:27
#> 389  14.51 2018-09-08 01:25:58
#> 390  14.50 2018-09-08 01:23:28
#> 391  14.50 2018-09-08 01:20:58
#> 392  14.46 2018-09-08 01:18:28
#> 393  14.44 2018-09-08 01:15:59
#> 394  14.49 2018-09-08 01:13:29
#> 395  14.53 2018-09-08 01:11:00
#> 396  14.50 2018-09-08 01:08:31
#> 397  14.53 2018-09-08 01:06:01
#> 398  14.55 2018-09-08 01:03:31
#> 399  14.54 2018-09-08 01:01:02
#> 400  14.57 2018-09-08 00:58:32
#> 401  14.55 2018-09-08 00:56:02
#> 402  14.54 2018-09-08 00:53:33
#> 403  14.61 2018-09-08 00:51:03
#> 404  14.59 2018-09-08 00:48:33
#> 405  14.66 2018-09-08 00:46:04
#> 406  14.60 2018-09-08 00:43:32
#> 407  14.65 2018-09-08 00:41:02
#> 408  14.63 2018-09-08 00:38:32
#> 409  14.64 2018-09-08 00:36:02
#> 410  14.60 2018-09-08 00:33:33
#> 411  14.59 2018-09-08 00:31:03
#> 412  14.59 2018-09-08 00:28:34
#> 413  14.65 2018-09-08 00:25:55
#> 414  14.65 2018-09-08 00:23:21
#> 415  14.70 2018-09-08 00:20:47
#> 416  14.75 2018-09-08 00:18:11
#> 417  14.74 2018-09-08 00:15:37
#> 418  14.75 2018-09-08 00:13:02
#> 419  14.68 2018-09-08 00:10:33
#> 420  14.70 2018-09-08 00:08:03
#> 421  14.74 2018-09-08 00:05:33
#> 422  14.75 2018-09-08 00:03:03
#> 423  14.72 2018-09-08 00:00:32
#> 424  14.79 2018-09-07 23:58:02
#> 425  14.74 2018-09-07 23:55:32
#> 426  14.71 2018-09-07 23:53:03
#> 427  14.82 2018-09-07 23:50:34
#> 428  14.83 2018-09-07 23:48:04
#> 429  14.82 2018-09-07 23:45:35
#> 430  14.82 2018-09-07 23:43:05
#> 431  14.81 2018-09-07 23:40:36
#> 432  14.79 2018-09-07 23:38:06
#> 433  14.76 2018-09-07 23:35:35
#> 434  14.74 2018-09-07 23:33:06
#> 435  14.76 2018-09-07 23:30:36
#> 436  14.84 2018-09-07 23:28:06
#> 437  14.93 2018-09-07 23:25:37
#> 438  14.95 2018-09-07 23:23:08
#> 439  14.91 2018-09-07 23:20:38
#> 440  14.92 2018-09-07 23:18:09
#> 441  14.89 2018-09-07 23:15:39
#> 442  14.86 2018-09-07 23:13:10
#> 443  14.83 2018-09-07 23:10:40
#> 444  14.81 2018-09-07 23:08:11
#> 445  14.80 2018-09-07 23:05:41
#> 446  14.76 2018-09-07 23:03:12
#> 447  14.74 2018-09-07 23:00:42
#> 448  14.80 2018-09-07 22:58:13
#> 449  14.89 2018-09-07 22:55:43
#> 450  14.80 2018-09-07 22:53:14
#> 451  14.90 2018-09-07 22:50:44
#> 452  14.84 2018-09-07 22:48:15
#> 453  14.86 2018-09-07 22:45:45
#> 454  14.84 2018-09-07 22:43:15
#> 455  14.81 2018-09-07 22:40:45
#> 456  14.83 2018-09-07 22:38:16
#> 457  14.90 2018-09-07 22:35:46
#> 458  14.95 2018-09-07 22:33:17
#> 459  14.92 2018-09-07 22:30:47
#> 460  14.89 2018-09-07 22:28:18
#> 461  14.88 2018-09-07 22:25:47
#> 462  14.98 2018-09-07 22:23:16
#> 463  15.01 2018-09-07 22:20:45
#> 464  15.03 2018-09-07 22:18:16
#> 465  15.00 2018-09-07 22:15:46
#> 466  14.99 2018-09-07 22:13:17
#> 467  15.03 2018-09-07 22:10:47
#> 468  15.05 2018-09-07 22:08:17
#> 469  15.06 2018-09-07 22:04:39
#> 470  15.06 2018-09-07 22:02:09
#> 471  15.09 2018-09-07 21:59:39
#> 472  15.11 2018-09-07 21:57:10
#> 473  15.14 2018-09-07 21:54:40
#> 474  15.10 2018-09-07 21:52:10
#> 475  15.10 2018-09-07 21:49:40
#> 476  15.06 2018-09-07 21:47:10
#> 477  15.08 2018-09-07 21:44:38
#> 478  15.02 2018-09-07 21:42:07
#> 479  15.10 2018-09-07 21:39:37
#> 480  15.12 2018-09-07 21:37:05
#> 481  15.20 2018-09-07 21:34:34
#> 482  15.20 2018-09-07 21:32:03
#> 483  15.18 2018-09-07 21:29:33
#> 484  15.12 2018-09-07 21:27:03
#> 485  15.14 2018-09-07 21:24:30
#> 486  15.31 2018-09-07 21:21:59
#> 487  15.15 2018-09-07 21:19:28
#> 488  15.13 2018-09-07 21:16:56
#> 489  15.20 2018-09-07 21:14:23
#> 490  15.21 2018-09-07 21:11:52
#> 491  15.16 2018-09-07 21:09:20
#> 492  15.14 2018-09-07 21:06:48
#> 493  15.11 2018-09-07 21:04:16
#> 494  15.14 2018-09-07 21:01:44
#> 495  15.19 2018-09-07 20:59:13
#> 496  15.22 2018-09-07 20:56:44
#> 497  15.22 2018-09-07 20:54:12
#> 498  15.17 2018-09-07 20:51:40
#> 499  15.18 2018-09-07 20:49:09
#> 500  15.20 2018-09-07 20:46:38
#> 501  15.22 2018-09-07 20:44:08
#> 502  15.18 2018-09-07 20:41:36
#> 503  15.12 2018-09-07 20:39:05
#> 504  15.21 2018-09-07 20:36:32
#> 505  15.18 2018-09-07 20:33:57
#> 506  15.25 2018-09-07 20:31:23
#> 507  15.30 2018-09-07 20:28:48
#> 508  15.25 2018-09-07 20:26:10
#> 509  15.25 2018-09-07 20:23:37
#> 510  15.21 2018-09-07 20:21:05
#> 511  15.22 2018-09-07 20:18:20
#> 512  15.14 2018-09-07 20:15:48
#> 513  15.15 2018-09-07 20:13:17
#> 514  15.18 2018-09-07 20:10:45
#> 515  15.20 2018-09-07 20:08:11
#> 516  15.21 2018-09-07 20:05:38
#> 517  15.21 2018-09-07 20:03:04
#> 518  15.20 2018-09-07 20:00:33
#> 519  15.31 2018-09-07 19:57:58
#> 520  15.34 2018-09-07 19:55:25
#> 521  15.29 2018-09-07 19:52:52
#> 522  15.28 2018-09-07 19:50:13
#> 523  15.28 2018-09-07 19:47:41
#> 524  15.26 2018-09-07 19:45:10
#> 525  15.42 2018-09-07 19:42:40
#> 526  15.50 2018-09-07 19:40:10
#> 527  15.53 2018-09-07 19:37:38
#> 528  15.55 2018-09-07 19:35:05
#> 529  15.49 2018-09-07 19:32:34
#> 530  15.51 2018-09-07 19:30:00
#> 531  15.49 2018-09-07 19:27:28
#> 532  15.49 2018-09-07 19:24:55
#> 533  15.47 2018-09-07 19:22:23
#> 534  15.51 2018-09-07 19:19:47
#> 535  15.59 2018-09-07 19:17:14
#> 536  15.53 2018-09-07 19:14:34
#> 537  15.66 2018-09-07 19:11:56
#> 538  15.67 2018-09-07 19:09:20
#> 539  15.67 2018-09-07 19:06:46
#> 540  15.69 2018-09-07 19:04:10
#> 541  15.67 2018-09-07 19:01:36
#> 542  15.73 2018-09-07 18:59:04
#> 543  15.79 2018-09-07 18:56:27
#> 544  15.83 2018-09-07 18:53:58
#> 545  15.90 2018-09-07 18:51:22
#> 546  15.90 2018-09-07 18:48:52
#> 547  15.98 2018-09-07 18:46:22
#> 548  15.97 2018-09-07 18:43:53
#> 549  15.96 2018-09-07 18:41:23
#> 550  15.99 2018-09-07 18:38:54
#> 551  15.94 2018-09-07 18:36:25
#> 552  15.98 2018-09-07 18:33:56
#> 553  15.96 2018-09-07 18:31:26
#> 554  16.01 2018-09-07 18:28:57
#> 555  16.15 2018-09-07 18:26:27
#> 556  16.16 2018-09-07 18:23:58
#> 557  16.19 2018-09-07 18:21:28
#> 558  16.20 2018-09-07 18:18:58
#> 559  16.19 2018-09-07 18:16:28
#> 560  16.33 2018-09-07 18:13:59
#> 561  16.39 2018-09-07 18:11:28
#> 562  16.37 2018-09-07 18:08:58
#> 563  16.44 2018-09-07 18:06:27
#> 564  16.46 2018-09-07 18:03:42
#> 565  16.51 2018-09-07 18:01:12
#> 566  16.46 2018-09-07 17:58:43
#> 567  16.45 2018-09-07 17:56:14
#> 568  16.44 2018-09-07 17:53:44
#> 569  16.47 2018-09-07 17:51:14
#> 570  16.51 2018-09-07 17:48:45
#> 571  16.58 2018-09-07 17:46:15
#> 572  16.62 2018-09-07 17:43:45
#> 573  16.66 2018-09-07 17:41:16
#> 574  16.71 2018-09-07 17:38:45
#> 575  16.80 2018-09-07 17:36:16
#> 576  16.83 2018-09-07 17:33:46
#> 577  16.78 2018-09-07 17:31:17
#> 578  16.80 2018-09-07 17:28:47
#> 579  16.85 2018-09-07 17:26:17
#> 580  16.86 2018-09-07 17:23:47
#> 581  16.91 2018-09-07 17:21:17
#> 582  17.00 2018-09-07 17:18:48
#> 583  16.97 2018-09-07 17:16:18
#> 584  17.02 2018-09-07 17:13:49
#> 585  17.08 2018-09-07 17:11:19
#> 586  17.14 2018-09-07 17:08:50
#> 587  17.12 2018-09-07 17:06:20
#> 588  17.14 2018-09-07 17:03:51
#> 589  17.08 2018-09-07 17:01:22
#> 590  17.03 2018-09-07 16:58:52
#> 591  17.02 2018-09-07 16:56:23
#> 592  17.05 2018-09-07 16:53:53
#> 593  17.22 2018-09-07 16:51:24
#> 594  17.24 2018-09-07 16:48:54
#> 595  17.21 2018-09-07 16:46:24
#> 596  17.29 2018-09-07 16:43:55
#> 597  17.28 2018-09-07 16:41:24
#> 598  17.30 2018-09-07 16:38:55
#> 599  17.25 2018-09-07 16:36:26
#> 600  17.31 2018-09-07 16:33:56
#> 601  17.43 2018-09-07 16:31:27
#> 602  17.45 2018-09-07 16:28:57
#> 603  17.35 2018-09-07 16:26:28
#> 604  17.31 2018-09-07 16:23:57
#> 605  17.33 2018-09-07 16:21:26
#> 606  17.30 2018-09-07 16:18:57
#> 607  17.49 2018-09-07 16:16:26
#> 608  17.40 2018-09-07 16:13:56
#> 609  17.46 2018-09-07 16:11:27
#> 610  17.60 2018-09-07 16:08:56
#> 611  17.66 2018-09-07 16:06:25
#> 612  17.62 2018-09-07 16:03:55
#> 613  17.58 2018-09-07 16:01:26
#> 614  17.61 2018-09-07 15:58:55
#> 615  17.57 2018-09-07 15:56:26
#> 616  17.70 2018-09-07 15:53:56
#> 617  17.72 2018-09-07 15:51:27
#> 618  17.71 2018-09-07 15:48:57
#> 619  17.72 2018-09-07 15:46:28
#> 620  17.69 2018-09-07 15:43:58
#> 621  17.73 2018-09-07 15:41:29
#> 622  17.69 2018-09-07 15:38:59
#> 623  17.75 2018-09-07 15:36:30
#> 624  17.71 2018-09-07 15:34:00
#> 625  17.73 2018-09-07 15:31:31
#> 626  17.68 2018-09-07 15:29:00
#> 627  17.63 2018-09-07 15:26:31
#> 628  17.62 2018-09-07 15:24:01
#> 629  17.74 2018-09-07 15:21:32
#> 630  17.73 2018-09-07 15:19:02
#> 631  17.68 2018-09-07 15:16:33
#> 632  17.75 2018-09-07 15:14:03
#> 633  17.72 2018-09-07 15:11:34
#> 634  17.67 2018-09-07 15:09:04
#> 635  17.72 2018-09-07 15:06:33
#> 636  17.65 2018-09-07 15:04:04
#> 637  17.54 2018-09-07 15:01:34
#> 638  17.67 2018-09-07 14:59:05
#> 639  17.70 2018-09-07 14:56:35
#> 640  17.64 2018-09-07 14:54:06
#> 641  17.46 2018-09-07 14:51:36
#> 642  17.35 2018-09-07 14:49:07
#> 643  17.41 2018-09-07 14:46:37
#> 644  17.49 2018-09-07 14:44:07
#> 645  17.51 2018-09-07 14:41:37
#> 646  17.66 2018-09-07 14:39:08
#> 647  17.71 2018-09-07 14:36:39
#> 648  17.68 2018-09-07 14:34:09
#> 649  17.69 2018-09-07 14:31:39
#> 650  17.70 2018-09-07 14:29:10
#> 651  17.78 2018-09-07 14:26:40
#> 652  17.97 2018-09-07 14:24:11
#> 653  17.99 2018-09-07 14:21:40
#> 654  18.10 2018-09-07 14:19:10
#> 655  18.07 2018-09-07 14:16:40
#> 656  18.05 2018-09-07 14:14:11
#> 657  17.97 2018-09-07 14:11:41
#> 658  18.01 2018-09-07 14:09:12
#> 659  18.09 2018-09-07 14:06:41
#> 660  17.97 2018-09-07 14:04:11
#> 661  18.05 2018-09-07 14:01:41
#> 662  18.23 2018-09-07 13:59:11
#> 663  18.28 2018-09-07 13:56:42
#> 664  18.46 2018-09-07 13:54:12
#> 665  18.57 2018-09-07 13:51:42
#> 666  18.58 2018-09-07 13:49:12
#> 667  18.61 2018-09-07 13:46:43
#> 668  18.50 2018-09-07 13:44:13
#> 669  18.65 2018-09-07 13:41:43
#> 670  18.64 2018-09-07 13:39:13
#> 671  18.66 2018-09-07 13:36:43
#> 672  18.65 2018-09-07 13:34:12
#> 673  18.64 2018-09-07 13:31:42
#> 674  18.63 2018-09-07 13:29:13
#> 675  18.53 2018-09-07 13:26:43
#> 676  18.66 2018-09-07 13:24:14
#> 677  18.48 2018-09-07 13:21:44
#> 678  18.43 2018-09-07 13:19:15
#> 679  18.38 2018-09-07 13:16:44
#> 680  18.56 2018-09-07 13:14:14
#> 681  18.46 2018-09-07 13:11:44
#> 682  18.34 2018-09-07 13:09:15
#> 683  18.40 2018-09-07 13:06:43
#> 684  18.40 2018-09-07 13:04:12
#> 685  18.44 2018-09-07 13:01:39
#> 686  18.39 2018-09-07 12:59:09
#> 687  18.37 2018-09-07 12:56:40
#> 688  18.43 2018-09-07 12:54:10
#> 689  18.44 2018-09-07 12:51:41
#> 690  18.31 2018-09-07 12:49:11
#> 691  18.36 2018-09-07 12:46:41
#> 692  18.43 2018-09-07 12:44:10
#> 693  18.59 2018-09-07 12:41:41
#> 694  18.44 2018-09-07 12:39:09
#> 695  18.60 2018-09-07 12:36:37
#> 696  18.59 2018-09-07 12:34:07
#> 697  18.52 2018-09-07 12:31:38
#> 698  18.37 2018-09-07 12:29:08
#> 699  18.38 2018-09-07 12:26:38
#> 700  18.56 2018-09-07 12:24:08
#> 701  18.73 2018-09-07 12:21:38
#> 702  18.90 2018-09-07 12:19:08
#> 703  18.86 2018-09-07 12:16:38
#> 704  18.82 2018-09-07 12:14:09
#> 705  18.70 2018-09-07 12:11:39
#> 706  18.76 2018-09-07 12:09:09
#> 707  18.60 2018-09-07 12:06:38
#> 708  18.58 2018-09-07 12:04:09
#> 709  18.73 2018-09-07 12:01:40
#> 710  18.76 2018-09-07 11:59:10
#> 711  18.73 2018-09-07 11:56:40
#> 712  18.88 2018-09-07 11:54:11
#> 713  18.74 2018-09-07 11:51:41
#> 714  18.75 2018-09-07 11:49:12
#> 715  18.78 2018-09-07 11:46:42
#> 716  18.89 2018-09-07 11:44:12
#> 717  18.96 2018-09-07 11:41:43
#> 718  19.15 2018-09-07 11:39:13
#> 719  19.15 2018-09-07 11:36:44
#> 720  19.05 2018-09-07 11:34:14
#> 721  19.29 2018-09-07 11:31:45
#> 722  19.25 2018-09-07 11:29:15
#> 723  19.20 2018-09-07 11:26:46
#> 724  19.46 2018-09-07 11:24:16
#> 725  19.46 2018-09-07 11:21:47
#> 726  19.44 2018-09-07 11:19:17
#> 727  19.69 2018-09-07 11:16:47
#> 728  19.58 2018-09-07 11:14:17
#> 729  19.52 2018-09-07 11:11:48
#> 730  19.55 2018-09-07 11:09:18
#> 731  19.39 2018-09-07 11:06:49
#> 732  19.41 2018-09-07 11:04:19
#> 733  19.15 2018-09-07 11:01:50
#> 734  19.13 2018-09-07 10:59:20
#> 735  18.95 2018-09-07 10:56:49
#> 736  19.00 2018-09-07 10:54:20
#> 737  18.93 2018-09-07 10:51:50
#> 738  18.84 2018-09-07 10:49:21
#> 739  18.72 2018-09-07 10:46:51
#> 740  18.64 2018-09-07 10:44:21
#> 741  18.62 2018-09-07 10:41:52
#> 742  18.67 2018-09-07 10:39:22
#> 743  18.76 2018-09-07 10:36:53
#> 744  18.63 2018-09-07 10:34:23
#> 745  18.63 2018-09-07 10:31:53
#> 746  18.57 2018-09-07 10:29:22
#> 747  18.71 2018-09-07 10:26:52
#> 748  18.67 2018-09-07 10:24:23
#> 749  18.78 2018-09-07 10:21:53
#> 750  18.94 2018-09-07 10:19:24
#> 751  18.92 2018-09-07 10:16:55
#> 752  18.88 2018-09-07 10:14:25
#> 753  18.83 2018-09-07 10:11:56
#> 754  18.94 2018-09-07 10:09:26
#> 755  19.01 2018-09-07 10:06:57
#> 756  19.02 2018-09-07 10:04:27
#> 757  19.07 2018-09-07 10:01:53
#> 758  19.09 2018-09-07 09:59:18
#> 759  19.09 2018-09-07 09:56:49
#> 760  18.98 2018-09-07 09:54:19
#> 761  18.93 2018-09-07 09:51:49
#> 762  18.89 2018-09-07 09:49:19
#> 763  18.80 2018-09-07 09:46:49
#> 764  18.75 2018-09-07 09:44:19
#> 765  18.65 2018-09-07 09:41:50
#> 766  18.56 2018-09-07 09:39:21
#> 767  18.45 2018-09-07 09:36:51
#> 768  18.37 2018-09-07 09:34:22
#> 769  18.29 2018-09-07 09:31:52
#> 770  18.36 2018-09-07 09:29:23
#> 771  18.09 2018-09-07 09:26:53
#> 772  18.14 2018-09-07 09:24:23
#> 773  18.21 2018-09-07 09:21:53
#> 774  18.08 2018-09-07 09:19:23
#> 775  17.97 2018-09-07 09:16:54
#> 776  17.74 2018-09-07 09:14:24
#> 777  17.83 2018-09-07 09:11:55
#> 778  17.97 2018-09-07 09:09:25
#> 779  17.73 2018-09-07 09:06:55
#> 780  17.68 2018-09-07 09:04:26
#> 781  17.61 2018-09-07 09:01:57
#> 782  17.58 2018-09-07 08:59:27
#> 783  17.50 2018-09-07 08:56:57
#> 784  17.42 2018-09-07 08:54:28
#> 785  17.29 2018-09-07 08:51:58
#> 786  17.30 2018-09-07 08:49:29
#> 787  17.40 2018-09-07 08:46:59
#> 788  17.37 2018-09-07 08:44:28
#> 789  17.13 2018-09-07 08:41:57
#> 790  17.07 2018-09-07 08:39:27
#> 791  17.12 2018-09-07 08:36:57
#> 792  16.93 2018-09-07 08:34:25
#> 793  17.00 2018-09-07 08:31:54
#> 794  16.87 2018-09-07 08:29:25
#> 795  16.73 2018-09-07 08:26:53
#> 796  16.64 2018-09-07 08:24:23
#> 797  16.68 2018-09-07 08:21:52
#> 798  16.56 2018-09-07 08:19:22
#> 799  16.47 2018-09-07 08:16:52
#> 800  16.35 2018-09-07 08:14:23
#> 801  16.27 2018-09-07 08:11:53
#> 802  16.30 2018-09-07 08:09:22
#> 803  16.23 2018-09-07 08:06:53
#> 804  16.02 2018-09-07 08:04:23
#> 805  15.90 2018-09-07 08:01:52
#> 806  15.98 2018-09-07 07:59:22
#> 807  16.07 2018-09-07 07:56:51
#> 808  16.06 2018-09-07 07:54:21
#> 809  15.95 2018-09-07 07:51:51
#> 810  15.95 2018-09-07 07:49:21
#> 811  15.98 2018-09-07 07:46:51
#> 812  16.02 2018-09-07 07:44:21
#> 813  16.02 2018-09-07 07:41:51
#> 814  15.90 2018-09-07 07:39:22
#> 815  15.97 2018-09-07 07:36:52
#> 816  15.91 2018-09-07 07:34:23
#> 817  15.94 2018-09-07 07:31:53
#> 818  16.05 2018-09-07 07:29:24
#> 819  16.12 2018-09-07 07:26:55
#> 820  16.39 2018-09-07 07:24:25
#> 821  16.49 2018-09-07 07:21:55
#> 822  16.59 2018-09-07 07:19:24
#> 823  16.57 2018-09-07 07:16:53
#> 824  16.54 2018-09-07 07:14:23
#> 825  16.42 2018-09-07 07:11:53
#> 826  16.28 2018-09-07 07:09:23
#> 827  15.96 2018-09-07 07:06:54
#> 828  15.69 2018-09-07 07:04:24
#> 829  15.71 2018-09-07 07:01:55
#> 830  15.81 2018-09-07 06:59:25
#> 831  16.20 2018-09-07 06:56:55
#> 832  16.01 2018-09-07 06:54:25
#> 833  15.70 2018-09-07 06:51:55
#> 834  15.69 2018-09-07 06:49:20
#> 835  15.67 2018-09-07 06:46:51
#> 836  15.79 2018-09-07 06:44:21
#> 837  15.71 2018-09-07 06:41:51
#> 838  15.86 2018-09-07 06:39:22
#> 839  16.05 2018-09-07 06:36:52
#> 840  15.56 2018-09-07 06:34:22
#> 841  15.39 2018-09-07 06:31:53
#> 842  14.75 2018-09-07 06:29:22
#> 843  14.84 2018-09-07 06:26:53
#> 844  14.75 2018-09-07 06:24:23
#> 845  14.46 2018-09-07 06:21:53
#> 846  14.10 2018-09-07 06:19:24
#> 847  13.83 2018-09-07 06:16:55
#> 848  13.79 2018-09-07 06:14:25
#> 849  13.77 2018-09-07 06:11:55
#> 850  13.73 2018-09-07 06:09:26
#> 851  13.57 2018-09-07 06:06:56
#> 852  13.82 2018-09-07 06:04:26
#> 853  13.85 2018-09-07 06:01:56
#> 854  13.49 2018-09-07 05:59:27
#> 855  13.50 2018-09-07 05:56:57
#> 856  13.61 2018-09-07 05:54:27
#> 857  13.37 2018-09-07 05:51:57
#> 858  13.22 2018-09-07 05:49:27
#> 859  13.02 2018-09-07 05:46:56
#> 860  12.83 2018-09-07 05:44:26
#> 861  12.72 2018-09-07 05:41:56
#> 862  12.60 2018-09-07 05:39:26
#> 863  12.55 2018-09-07 05:36:55
#> 864  12.43 2018-09-07 05:34:25
#> 865  12.41 2018-09-07 05:31:55
#> 866  12.33 2018-09-07 05:29:26
#> 867  12.39 2018-09-07 05:26:56
#> 868  12.33 2018-09-07 05:24:27
#> 869  12.30 2018-09-07 05:21:56
#> 870  12.25 2018-09-07 05:19:27
#> 871  12.22 2018-09-07 05:16:57
#> 872  12.19 2018-09-07 05:14:27
#> 873  12.13 2018-09-07 05:11:58
#> 874  12.05 2018-09-07 05:09:29
#> 875  12.03 2018-09-07 05:06:59
#> 876  11.98 2018-09-07 05:04:30
#> 877  11.96 2018-09-07 05:02:00
#> 878  11.89 2018-09-07 04:59:31
#> 879  11.90 2018-09-07 04:57:02
#> 880  11.88 2018-09-07 04:54:32
#> 881  11.77 2018-09-07 04:52:02
#> 882  11.77 2018-09-07 04:49:31
#> 883  11.63 2018-09-07 04:47:02
#> 884  11.61 2018-09-07 04:44:32
#> 885  11.77 2018-09-07 04:42:03
#> 886  11.82 2018-09-07 04:39:33
#> 887  11.77 2018-09-07 04:37:04
#> 888  11.76 2018-09-07 04:34:34
#> 889  11.86 2018-09-07 04:32:05
#> 890  11.78 2018-09-07 04:29:36
#> 891  11.81 2018-09-07 04:27:06
#> 892  11.75 2018-09-07 04:24:36
#> 893  11.74 2018-09-07 04:22:06
#> 894  11.75 2018-09-07 04:19:37
#> 895  11.75 2018-09-07 04:17:07
#> 896  11.78 2018-09-07 04:14:36
#> 897  11.76 2018-09-07 04:12:07
#> 898  11.70 2018-09-07 04:09:38
#> 899  11.63 2018-09-07 04:07:08
#> 900  11.72 2018-09-07 04:04:39
#> 901  11.86 2018-09-07 04:02:09
#> 902  11.82 2018-09-07 03:59:40
#> 903  11.90 2018-09-07 03:57:10
#> 904  11.91 2018-09-07 03:54:39
#> 905  11.92 2018-09-07 03:52:10
#> 906  11.95 2018-09-07 03:49:41
#> 907  11.90 2018-09-07 03:47:11
#> 908  11.95 2018-09-07 03:44:41
#> 909  11.93 2018-09-07 03:42:11
#> 910  11.95 2018-09-07 03:39:41
#> 911  12.04 2018-09-07 03:37:12
#> 912  11.95 2018-09-07 03:34:43
#> 913  12.00 2018-09-07 03:32:13
#> 914  12.07 2018-09-07 03:29:44
#> 915  12.08 2018-09-07 03:27:14
#> 916  12.05 2018-09-07 03:24:45
#> 917  12.08 2018-09-07 03:22:15
#> 918  12.09 2018-09-07 03:19:46
#> 919  12.09 2018-09-07 03:17:16
#> 920  12.15 2018-09-07 03:14:46
#> 921  12.14 2018-09-07 03:12:16
#> 922  12.07 2018-09-07 03:09:47
#> 923  11.99 2018-09-07 03:07:17
#> 924  12.23 2018-09-07 03:04:47
#> 925  12.32 2018-09-07 03:02:17
#> 926  12.34 2018-09-07 02:59:48
#> 927  12.28 2018-09-07 02:57:19
#> 928  12.37 2018-09-07 02:54:49
#> 929  12.31 2018-09-07 02:52:20
#> 930  12.26 2018-09-07 02:49:50
#> 931  12.14 2018-09-07 02:47:21
#> 932  12.35 2018-09-07 02:44:50
#> 933  12.45 2018-09-07 02:42:21
#> 934  12.58 2018-09-07 02:39:51
#> 935  12.46 2018-09-07 02:37:22
#> 936  12.48 2018-09-07 02:34:52
#> 937  12.44 2018-09-07 02:32:23
#> 938  12.50 2018-09-07 02:29:53
#> 939  12.51 2018-09-07 02:27:24
#> 940  12.50 2018-09-07 02:24:54
#> 941  12.55 2018-09-07 02:22:25
#> 942  12.62 2018-09-07 02:19:56
#> 943  12.51 2018-09-07 02:17:26
#> 944  12.43 2018-09-07 02:14:57
#> 945  12.63 2018-09-07 02:12:22
#> 946  12.64 2018-09-07 02:09:53
#> 947  12.58 2018-09-07 02:07:24
#> 948  12.64 2018-09-07 02:04:54
#> 949  12.71 2018-09-07 02:02:25
#> 950  12.72 2018-09-07 01:59:56
#> 951  12.70 2018-09-07 01:57:27
#> 952  12.77 2018-09-07 01:54:57
#> 953  12.73 2018-09-07 01:52:27
#> 954  12.61 2018-09-07 01:49:58
#> 955  12.57 2018-09-07 01:47:28
#> 956  12.76 2018-09-07 01:44:58
#> 957  12.76 2018-09-07 01:42:29
#> 958  12.78 2018-09-07 01:39:59
#> 959  12.75 2018-09-07 01:37:30
#> 960  12.72 2018-09-07 01:34:55
#> 961  12.61 2018-09-07 01:32:26
#> 962  12.55 2018-09-07 01:29:57
#> 963  12.79 2018-09-07 01:27:27
#> 964  13.00 2018-09-07 01:24:57
#> 965  12.94 2018-09-07 01:22:28
#> 966  12.95 2018-09-07 01:19:59
#> 967  12.88 2018-09-07 01:17:29
#> 968  12.99 2018-09-07 01:15:00
#> 969  13.04 2018-09-07 01:12:30
#> 970  13.05 2018-09-07 01:10:01
#> 971  12.98 2018-09-07 01:07:31
#> 972  13.04 2018-09-07 01:05:02
#> 973  12.95 2018-09-07 01:02:32
#> 974  12.88 2018-09-07 01:00:02
#> 975  13.09 2018-09-07 00:57:32
#> 976  13.11 2018-09-07 00:55:02
#> 977  13.07 2018-09-07 00:52:32
#> 978  12.99 2018-09-07 00:50:02
#> 979  13.00 2018-09-07 00:47:33
#> 980  12.92 2018-09-07 00:45:03
#> 981  13.16 2018-09-07 00:42:34
#> 982  13.21 2018-09-07 00:40:05
#> 983  13.25 2018-09-07 00:37:35
#> 984  13.22 2018-09-07 00:35:06
#> 985  13.30 2018-09-07 00:32:36
#> 986  13.27 2018-09-07 00:30:06
#> 987  13.31 2018-09-07 00:27:37
#> 988  13.30 2018-09-07 00:25:08
#> 989  13.38 2018-09-07 00:22:38
#> 990  13.44 2018-09-07 00:20:08
#> 991  13.45 2018-09-07 00:17:38
#> 992  13.48 2018-09-07 00:15:08
#> 993  13.45 2018-09-07 00:12:39
#> 994  13.49 2018-09-07 00:10:09
#> 995  13.54 2018-09-07 00:07:40
#> 996  13.60 2018-09-07 00:05:10
#> 997  13.59 2018-09-07 00:02:40
#> 998  13.61 2018-09-07 00:00:10
#> 999  13.53 2018-09-06 23:57:38
#> 1000 13.67 2018-09-06 23:55:08
#> 1001 13.77 2018-09-06 23:52:39
#> 1002 13.76 2018-09-06 23:50:09
#> 1003 13.73 2018-09-06 23:47:39
#> 1004 13.81 2018-09-06 23:45:09
#> 1005 13.82 2018-09-06 23:42:39
#> 1006 13.77 2018-09-06 23:40:09
#> 1007 13.67 2018-09-06 23:37:40
#> 1008 13.76 2018-09-06 23:35:10
#> 1009 13.69 2018-09-06 23:32:41
#> 1010 13.76 2018-09-06 23:30:10
#> 1011 13.78 2018-09-06 23:27:41
#> 1012 13.94 2018-09-06 23:25:10
#> 1013 13.82 2018-09-06 23:22:41
#> 1014 13.80 2018-09-06 23:20:11
#> 1015 13.88 2018-09-06 23:17:42
#> 1016 13.95 2018-09-06 23:15:12
#> 1017 13.94 2018-09-06 23:12:42
#> 1018 14.03 2018-09-06 23:10:12
#> 1019 14.04 2018-09-06 23:07:43
#> 1020 14.10 2018-09-06 23:05:13
#> 1021 14.06 2018-09-06 23:02:44
#> 1022 14.23 2018-09-06 23:00:09
#> 1023 14.15 2018-09-06 22:57:40
#> 1024 14.20 2018-09-06 22:55:11
#> 1025 14.18 2018-09-06 22:52:41
#> 1026 14.23 2018-09-06 22:50:12
#> 1027 14.19 2018-09-06 22:47:41
#> 1028 14.24 2018-09-06 22:44:54
#> 1029 14.35 2018-09-06 22:42:25
#> 1030 14.44 2018-09-06 22:39:55
#> 1031 14.33 2018-09-06 22:37:26
#> 1032 14.38 2018-09-06 22:34:56
#> 1033 14.39 2018-09-06 22:32:26
#> 1034 14.47 2018-09-06 22:29:57
#> 1035 14.50 2018-09-06 22:27:27
#> 1036 14.58 2018-09-06 22:24:58
#> 1037 14.59 2018-09-06 22:22:28
#> 1038 14.62 2018-09-06 22:19:58
#> 1039 14.53 2018-09-06 22:17:28
#> 1040 14.53 2018-09-06 22:14:58
#> 1041 14.59 2018-09-06 22:12:29
#> 1042 14.66 2018-09-06 22:10:00
#> 1043 14.76 2018-09-06 22:07:29
#> 1044 14.83 2018-09-06 22:04:59
#> 1045 14.90 2018-09-06 22:02:29
#> 1046 14.96 2018-09-06 21:59:54
#> 1047 14.93 2018-09-06 21:57:23
#> 1048 14.92 2018-09-06 21:54:54
#> 1049 15.14 2018-09-06 21:52:24
#> 1050 15.08 2018-09-06 21:49:53
#> 1051 15.01 2018-09-06 21:47:23
#> 1052 15.15 2018-09-06 21:44:53
#> 1053 15.19 2018-09-06 21:42:22
#> 1054 15.22 2018-09-06 21:39:53
#> 1055 15.30 2018-09-06 21:37:22
#> 1056 15.22 2018-09-06 21:34:52
#> 1057 15.11 2018-09-06 21:32:22
#> 1058 15.06 2018-09-06 21:29:51
#> 1059 15.14 2018-09-06 21:27:22
#> 1060 15.33 2018-09-06 21:24:52
#> 1061 15.39 2018-09-06 21:22:22
#> 1062 15.56 2018-09-06 21:19:51
#> 1063 15.65 2018-09-06 21:17:21
#> 1064 15.57 2018-09-06 21:14:52
#> 1065 15.63 2018-09-06 21:12:21
#> 1066 15.58 2018-09-06 21:09:51
#> 1067 15.71 2018-09-06 21:07:21
#> 1068 15.68 2018-09-06 21:04:51
#> 1069 15.85 2018-09-06 21:02:21
#> 1070 15.66 2018-09-06 20:59:45
#> 1071 15.61 2018-09-06 20:57:15
#> 1072 15.70 2018-09-06 20:54:45
#> 1073 15.91 2018-09-06 20:52:14
#> 1074 15.90 2018-09-06 20:49:44
#> 1075 15.96 2018-09-06 20:47:14
#> 1076 15.96 2018-09-06 20:44:43
#> 1077 16.09 2018-09-06 20:42:13
#> 1078 16.05 2018-09-06 20:39:43
#> 1079 16.07 2018-09-06 20:37:13
#> 1080 16.16 2018-09-06 20:34:43
#> 1081 16.11 2018-09-06 20:32:13
#> 1082 16.28 2018-09-06 20:29:43
#> 1083 16.45 2018-09-06 20:27:13
#> 1084 16.51 2018-09-06 20:24:42
#> 1085 16.42 2018-09-06 20:22:12
#> 1086 16.47 2018-09-06 20:19:42
#> 1087 16.54 2018-09-06 20:17:12
#> 1088 16.58 2018-09-06 20:14:42
#> 1089 16.55 2018-09-06 20:12:11
#> 1090 16.53 2018-09-06 20:09:42
#> 1091 16.50 2018-09-06 20:07:12
#> 1092 16.71 2018-09-06 20:04:40
#> 1093 16.84 2018-09-06 20:02:05
#> 1094 16.80 2018-09-06 19:59:33
#> 1095 16.86 2018-09-06 19:57:03
#> 1096 16.94 2018-09-06 19:54:33
#> 1097 16.82 2018-09-06 19:52:04
#> 1098 16.89 2018-09-06 19:49:34
#> 1099 17.04 2018-09-06 19:47:04
#> 1100 17.14 2018-09-06 19:44:34
#> 1101 16.99 2018-09-06 19:42:05
#> 1102 16.92 2018-09-06 19:39:35
#> 1103 17.18 2018-09-06 19:37:05
#> 1104 17.14 2018-09-06 19:34:35
#> 1105 17.31 2018-09-06 19:32:06
#> 1106 17.36 2018-09-06 19:29:34
#> 1107 17.45 2018-09-06 19:27:04
#> 1108 17.53 2018-09-06 19:24:34
#> 1109 17.64 2018-09-06 19:22:04
#> 1110 17.69 2018-09-06 19:19:34
#> 1111 17.70 2018-09-06 19:17:03
#> 1112 17.80 2018-09-06 19:14:31
#> 1113 17.79 2018-09-06 19:12:01
#> 1114 17.63 2018-09-06 19:09:30
#> 1115 17.59 2018-09-06 19:07:00
#> 1116 17.64 2018-09-06 19:04:28
#> 1117 17.66 2018-09-06 19:01:57
#> 1118 17.86 2018-09-06 18:59:28
#> 1119 17.92 2018-09-06 18:56:57
#> 1120 17.97 2018-09-06 18:54:25
#> 1121 17.80 2018-09-06 18:51:56
#> 1122 18.02 2018-09-06 18:49:26
#> 1123 18.09 2018-09-06 18:46:56
#> 1124 18.21 2018-09-06 18:44:25
#> 1125 18.24 2018-09-06 18:41:56
#> 1126 18.29 2018-09-06 18:39:26
#> 1127 18.28 2018-09-06 18:36:56
#> 1128 18.38 2018-09-06 18:34:25
#> 1129 18.48 2018-09-06 18:31:54
#> 1130 18.48 2018-09-06 18:29:25
#> 1131 18.46 2018-09-06 18:26:54
#> 1132 18.54 2018-09-06 18:24:24
#> 1133 18.48 2018-09-06 18:21:54
#> 1134 18.57 2018-09-06 18:19:23
#> 1135 18.74 2018-09-06 18:16:51
#> 1136 18.82 2018-09-06 18:14:22
#> 1137 18.80 2018-09-06 18:11:50
#> 1138 18.89 2018-09-06 18:09:20
#> 1139 18.95 2018-09-06 18:06:51
#> 1140 19.03 2018-09-06 18:04:18
#> 1141 19.02 2018-09-06 18:01:48
#> 1142 19.02 2018-09-06 17:58:42
#> 1143 19.01 2018-09-06 17:56:13
#> 1144 19.14 2018-09-06 17:53:44
#> 1145 19.36 2018-09-06 17:51:13
#> 1146 19.43 2018-09-06 17:48:42
#> 1147 19.58 2018-09-06 17:46:12
#> 1148 19.57 2018-09-06 17:43:42
#> 1149 19.64 2018-09-06 17:41:12
#> 1150 19.80 2018-09-06 17:38:42
#> 
#> 

# NOT RUN {
get_senseBox_data(
  senseBoxId = c("5957b67494f0520011304cc1","5957b67494f0520011304cc1") ,
  sensorId = "all",
  fromDate = "2017-11-25-12:00:00",
  toDate = "2017-11-26-12:00:00")

get_senseBox_data(
  senseBoxId = "5957b67494f0520011304cc1",
  sensorId = "5957b67494f0520011304cc4",
  fromDate = "2017-11-25-12:00:00",
  toDate = "2017-11-26-12:00:00")
# }