Get data from one specific senseBox sensor.

If the arguments fromDate and toDate are NOT given, the measurements from the last 48 h from now are downloaded. The maximum numbers of downloaded measurements are 10,000. The maximum time frame is one month. It is possible to download the data directly as a csv file, when argument CSV is TRUE.

get_senseBox_data(senseBoxId, sensorId, fromDate = NULL, toDate = NULL,
  CSV = FALSE, POSIXct = TRUE)

Arguments

senseBoxId
character (required): senseBoxId

sensorId
character (required): sensorId

fromDate
character (optional): Just show data with from date in format YYYY-mm-dd HH:MM:SS

toDate
character (optional): Just show data with up to date in format YYYY-mm-dd HH:MM:SS

CSV
logical (optional): Download data as .csv? NOT POSSIBLE UNTIL NOW!

POSIXct
logical (optional): Should the timestamp be translated into POSIXct?

Value

list

Function version

0.0.1

Examples

##=====================================
## Example: get data from sensorId
##=====================================

get_senseBox_data(
  senseBoxId = "5957b67494f0520011304cc1",
  sensorId = "5957b67494f0520011304cc4")

#> $`5957b67494f0520011304cc1`
#> $`5957b67494f0520011304cc1`$`5957b67494f0520011304cc4`
#>      value           createdAt
#> 1     0.69 2017-11-26 19:15:23
#> 2     0.72 2017-11-26 19:12:55
#> 3     0.68 2017-11-26 19:10:27
#> 4     0.74 2017-11-26 19:07:59
#> 5     0.79 2017-11-26 19:05:32
#> 6     0.78 2017-11-26 19:03:03
#> 7     0.78 2017-11-26 19:00:36
#> 8     0.83 2017-11-26 18:58:08
#> 9     0.78 2017-11-26 18:55:40
#> 10    0.80 2017-11-26 18:53:12
#> 11    0.82 2017-11-26 18:50:45
#> 12    0.88 2017-11-26 18:48:17
#> 13    0.80 2017-11-26 18:45:49
#> 14    0.79 2017-11-26 18:43:21
#> 15    0.76 2017-11-26 18:40:53
#> 16    0.81 2017-11-26 18:38:26
#> 17    0.79 2017-11-26 18:35:58
#> 18    0.74 2017-11-26 18:33:30
#> 19    0.77 2017-11-26 18:31:02
#> 20    0.83 2017-11-26 18:28:34
#> 21    0.84 2017-11-26 18:26:07
#> 22    0.88 2017-11-26 18:23:37
#> 23    0.95 2017-11-26 18:21:10
#> 24    0.94 2017-11-26 18:18:42
#> 25    0.94 2017-11-26 18:16:14
#> 26    0.86 2017-11-26 18:13:46
#> 27    0.80 2017-11-26 18:11:18
#> 28    0.85 2017-11-26 18:08:50
#> 29    0.92 2017-11-26 18:06:21
#> 30    0.98 2017-11-26 18:03:53
#> 31    0.92 2017-11-26 18:01:25
#> 32    0.93 2017-11-26 17:58:57
#> 33    0.89 2017-11-26 17:56:29
#> 34    0.90 2017-11-26 17:54:00
#> 35    0.88 2017-11-26 17:51:32
#> 36    0.98 2017-11-26 17:49:04
#> 37    0.88 2017-11-26 17:46:36
#> 38    0.99 2017-11-26 17:44:09
#> 39    0.96 2017-11-26 17:41:41
#> 40    0.87 2017-11-26 17:39:13
#> 41    0.85 2017-11-26 17:36:45
#> 42    0.80 2017-11-26 17:34:17
#> 43    0.69 2017-11-26 17:31:49
#> 44    0.81 2017-11-26 17:29:21
#> 45    0.83 2017-11-26 17:26:53
#> 46    0.74 2017-11-26 17:24:25
#> 47    0.68 2017-11-26 17:21:57
#> 48    0.64 2017-11-26 17:19:29
#> 49    0.69 2017-11-26 17:17:01
#> 50    0.71 2017-11-26 17:14:34
#> 51    0.75 2017-11-26 17:12:06
#> 52    0.88 2017-11-26 17:09:38
#> 53    0.81 2017-11-26 17:07:10
#> 54    0.78 2017-11-26 17:04:42
#> 55    0.75 2017-11-26 17:02:14
#> 56    0.66 2017-11-26 16:59:46
#> 57    0.62 2017-11-26 16:57:18
#> 58    0.82 2017-11-26 16:54:50
#> 59    0.83 2017-11-26 16:52:22
#> 60    0.87 2017-11-26 16:49:54
#> 61    0.88 2017-11-26 16:47:26
#> 62    0.90 2017-11-26 16:44:58
#> 63    1.00 2017-11-26 16:42:31
#> 64    1.11 2017-11-26 16:40:03
#> 65    1.09 2017-11-26 16:37:35
#> 66    1.10 2017-11-26 16:35:07
#> 67    1.11 2017-11-26 16:32:39
#> 68    1.09 2017-11-26 16:30:11
#> 69    1.10 2017-11-26 16:27:44
#> 70    1.07 2017-11-26 16:25:16
#> 71    1.10 2017-11-26 16:22:48
#> 72    1.10 2017-11-26 16:20:20
#> 73    1.04 2017-11-26 16:17:52
#> 74    1.03 2017-11-26 16:15:25
#> 75    1.07 2017-11-26 16:12:57
#> 76    1.09 2017-11-26 16:10:29
#> 77    1.16 2017-11-26 16:08:01
#> 78    1.29 2017-11-26 16:05:33
#> 79    1.43 2017-11-26 16:03:05
#> 80    1.41 2017-11-26 16:00:38
#> 81    1.30 2017-11-26 15:58:10
#> 82    1.14 2017-11-26 15:55:42
#> 83    1.23 2017-11-26 15:53:14
#> 84    1.24 2017-11-26 15:50:46
#> 85    1.28 2017-11-26 15:48:18
#> 86    1.41 2017-11-26 15:45:50
#> 87    1.56 2017-11-26 15:43:22
#> 88    1.70 2017-11-26 15:40:55
#> 89    1.42 2017-11-26 15:38:27
#> 90    1.45 2017-11-26 15:35:59
#> 91    1.49 2017-11-26 15:33:31
#> 92    1.45 2017-11-26 15:31:03
#> 93    1.45 2017-11-26 15:28:35
#> 94    1.61 2017-11-26 15:26:07
#> 95    1.61 2017-11-26 15:23:39
#> 96    1.63 2017-11-26 15:21:11
#> 97    1.65 2017-11-26 15:18:43
#> 98    1.69 2017-11-26 15:16:15
#> 99    1.63 2017-11-26 15:13:47
#> 100   1.64 2017-11-26 15:11:19
#> 101   1.61 2017-11-26 15:08:51
#> 102   1.63 2017-11-26 15:06:24
#> 103   1.65 2017-11-26 15:03:56
#> 104   1.78 2017-11-26 15:01:28
#> 105   1.89 2017-11-26 14:59:00
#> 106   1.71 2017-11-26 14:56:32
#> 107   1.68 2017-11-26 14:54:04
#> 108   1.83 2017-11-26 14:51:36
#> 109   1.98 2017-11-26 14:49:07
#> 110   1.93 2017-11-26 14:46:39
#> 111   1.85 2017-11-26 14:44:11
#> 112   1.71 2017-11-26 14:41:44
#> 113   1.75 2017-11-26 14:39:16
#> 114   1.80 2017-11-26 14:36:48
#> 115   2.03 2017-11-26 14:34:20
#> 116   1.90 2017-11-26 14:31:52
#> 117   1.81 2017-11-26 14:29:24
#> 118   1.82 2017-11-26 14:26:56
#> 119   1.92 2017-11-26 14:24:28
#> 120   2.01 2017-11-26 14:22:00
#> 121   1.98 2017-11-26 14:19:33
#> 122   1.93 2017-11-26 14:17:05
#> 123   2.01 2017-11-26 14:14:37
#> 124   1.85 2017-11-26 14:12:08
#> 125   1.84 2017-11-26 14:09:40
#> 126   1.93 2017-11-26 14:07:12
#> 127   1.86 2017-11-26 14:04:44
#> 128   1.79 2017-11-26 14:02:16
#> 129   1.93 2017-11-26 13:59:48
#> 130   1.88 2017-11-26 13:57:20
#> 131   1.97 2017-11-26 13:54:53
#> 132   2.01 2017-11-26 13:52:25
#> 133   1.91 2017-11-26 13:49:57
#> 134   1.83 2017-11-26 13:47:29
#> 135   1.98 2017-11-26 13:45:01
#> 136   1.87 2017-11-26 13:42:33
#> 137   1.85 2017-11-26 13:40:05
#> 138   1.75 2017-11-26 13:37:36
#> 139   2.09 2017-11-26 13:35:07
#> 140   2.15 2017-11-26 13:32:38
#> 141   2.26 2017-11-26 13:30:10
#> 142   2.27 2017-11-26 13:27:42
#> 143   2.36 2017-11-26 13:25:14
#> 144   2.42 2017-11-26 13:22:46
#> 145   2.37 2017-11-26 13:20:18
#> 146   2.31 2017-11-26 13:17:50
#> 147   2.29 2017-11-26 13:15:23
#> 148   2.46 2017-11-26 13:12:55
#> 149   2.29 2017-11-26 13:10:27
#> 150   2.37 2017-11-26 13:07:59
#> 151   2.37 2017-11-26 13:05:31
#> 152   2.18 2017-11-26 13:03:03
#> 153   2.33 2017-11-26 13:00:36
#> 154   2.31 2017-11-26 12:58:08
#> 155   2.22 2017-11-26 12:55:40
#> 156   2.23 2017-11-26 12:53:11
#> 157   2.33 2017-11-26 12:50:43
#> 158   2.28 2017-11-26 12:48:15
#> 159   2.23 2017-11-26 12:45:48
#> 160   2.10 2017-11-26 12:43:20
#> 161   2.08 2017-11-26 12:40:51
#> 162   2.00 2017-11-26 12:38:24
#> 163   2.18 2017-11-26 12:35:56
#> 164   2.00 2017-11-26 12:33:28
#> 165   2.19 2017-11-26 12:31:00
#> 166   2.24 2017-11-26 12:28:32
#> 167   2.10 2017-11-26 12:26:04
#> 168   2.30 2017-11-26 12:23:37
#> 169   2.15 2017-11-26 12:21:08
#> 170   2.29 2017-11-26 12:18:39
#> 171   2.11 2017-11-26 12:13:37
#> 172   2.11 2017-11-26 12:11:08
#> 173   2.16 2017-11-26 12:08:40
#> 174   2.08 2017-11-26 12:06:13
#> 175   1.94 2017-11-26 12:03:45
#> 176   1.94 2017-11-26 12:01:17
#> 177   2.14 2017-11-26 11:58:49
#> 178   1.95 2017-11-26 11:56:21
#> 179   1.96 2017-11-26 11:53:53
#> 180   1.98 2017-11-26 11:51:24
#> 181   1.84 2017-11-26 11:48:56
#> 182   1.83 2017-11-26 11:46:29
#> 183   1.91 2017-11-26 11:44:01
#> 184   1.93 2017-11-26 11:41:33
#> 185   1.95 2017-11-26 11:39:05
#> 186   1.98 2017-11-26 11:36:38
#> 187   1.95 2017-11-26 11:34:10
#> 188   2.13 2017-11-26 11:31:42
#> 189   2.19 2017-11-26 11:29:14
#> 190   2.04 2017-11-26 11:26:47
#> 191   2.22 2017-11-26 11:24:19
#> 192   2.13 2017-11-26 11:21:51
#> 193   2.22 2017-11-26 11:19:23
#> 194   1.79 2017-11-26 11:16:55
#> 195   1.95 2017-11-26 11:14:28
#> 196   2.09 2017-11-26 11:12:00
#> 197   2.31 2017-11-26 11:09:32
#> 198   2.30 2017-11-26 11:07:04
#> 199   2.29 2017-11-26 11:04:36
#> 200   2.12 2017-11-26 11:02:09
#> 201   2.22 2017-11-26 10:59:41
#> 202   2.21 2017-11-26 10:57:12
#> 203   2.22 2017-11-26 10:54:44
#> 204   2.30 2017-11-26 10:52:16
#> 205   2.48 2017-11-26 10:49:48
#> 206   2.53 2017-11-26 10:47:20
#> 207   2.54 2017-11-26 10:44:53
#> 208   2.70 2017-11-26 10:42:25
#> 209   2.81 2017-11-26 10:39:57
#> 210   2.93 2017-11-26 10:37:29
#> 211   3.04 2017-11-26 10:35:01
#> 212   3.02 2017-11-26 10:32:34
#> 213   2.97 2017-11-26 10:30:06
#> 214   2.94 2017-11-26 10:27:38
#> 215   3.03 2017-11-26 10:25:10
#> 216   3.05 2017-11-26 10:22:43
#> 217   3.01 2017-11-26 10:20:15
#> 218   3.04 2017-11-26 10:17:47
#> 219   2.98 2017-11-26 10:15:19
#> 220   3.02 2017-11-26 10:12:51
#> 221   3.04 2017-11-26 10:10:23
#> 222   3.03 2017-11-26 10:07:55
#> 223   2.97 2017-11-26 10:05:28
#> 224   2.93 2017-11-26 10:03:00
#> 225   2.92 2017-11-26 10:00:32
#> 226   3.02 2017-11-26 09:58:04
#> 227   2.94 2017-11-26 09:55:37
#> 228   3.07 2017-11-26 09:53:09
#> 229   3.13 2017-11-26 09:50:41
#> 230   3.12 2017-11-26 09:48:13
#> 231   3.08 2017-11-26 09:45:46
#> 232   3.10 2017-11-26 09:43:18
#> 233   3.19 2017-11-26 09:40:50
#> 234   3.18 2017-11-26 09:38:22
#> 235   3.15 2017-11-26 09:35:54
#> 236   3.35 2017-11-26 09:33:26
#> 237   3.38 2017-11-26 09:30:58
#> 238   3.40 2017-11-26 09:28:30
#> 239   3.41 2017-11-26 09:26:03
#> 240   3.46 2017-11-26 09:23:35
#> 241   3.41 2017-11-26 09:21:07
#> 242   3.36 2017-11-26 09:18:39
#> 243   3.33 2017-11-26 09:16:11
#> 244   3.38 2017-11-26 09:13:43
#> 245   3.31 2017-11-26 09:11:15
#> 246   3.29 2017-11-26 09:08:48
#> 247   3.29 2017-11-26 09:06:03
#> 248   3.16 2017-11-26 09:03:35
#> 249   3.06 2017-11-26 09:01:06
#> 250   2.86 2017-11-26 08:58:37
#> 251   2.89 2017-11-26 08:56:09
#> 252   3.20 2017-11-26 08:53:41
#> 253   3.39 2017-11-26 08:51:12
#> 254   3.36 2017-11-26 08:48:45
#> 255   3.32 2017-11-26 08:46:17
#> 256   3.18 2017-11-26 08:43:49
#> 257   3.23 2017-11-26 08:41:21
#> 258   3.27 2017-11-26 08:38:53
#> 259   2.97 2017-11-26 08:36:26
#> 260   2.94 2017-11-26 08:33:57
#> 261   2.91 2017-11-26 08:31:30
#> 262   2.92 2017-11-26 08:29:02
#> 263   2.94 2017-11-26 08:26:34
#> 264   2.87 2017-11-26 08:24:06
#> 265   2.95 2017-11-26 08:21:38
#> 266   2.91 2017-11-26 08:19:08
#> 267   2.83 2017-11-26 08:16:40
#> 268   2.80 2017-11-26 08:14:13
#> 269   2.81 2017-11-26 08:11:42
#> 270   2.82 2017-11-26 08:09:14
#> 271   2.84 2017-11-26 08:06:46
#> 272   2.70 2017-11-26 08:04:17
#> 273   2.62 2017-11-26 08:01:49
#> 274   2.70 2017-11-26 07:59:22
#> 275   2.67 2017-11-26 07:56:52
#> 276   2.68 2017-11-26 07:54:22
#> 277   3.02 2017-11-26 07:51:52
#> 278   2.77 2017-11-26 07:49:24
#> 279   2.73 2017-11-26 07:46:57
#> 280   2.78 2017-11-26 07:44:29
#> 281   2.80 2017-11-26 07:42:00
#> 282   2.78 2017-11-26 07:39:32
#> 283   2.80 2017-11-26 07:37:02
#> 284   2.72 2017-11-26 07:34:34
#> 285   2.69 2017-11-26 07:32:06
#> 286   2.69 2017-11-26 07:29:38
#> 287   2.71 2017-11-26 07:27:10
#> 288   2.56 2017-11-26 07:24:43
#> 289   2.61 2017-11-26 07:22:15
#> 290   2.55 2017-11-26 07:19:47
#> 291   2.53 2017-11-26 07:17:19
#> 292   2.45 2017-11-26 07:14:51
#> 293   2.40 2017-11-26 07:12:23
#> 294   2.47 2017-11-26 07:09:55
#> 295   2.31 2017-11-26 07:07:28
#> 296   2.42 2017-11-26 07:05:00
#> 297   2.49 2017-11-26 07:02:32
#> 298   2.42 2017-11-26 07:00:04
#> 299   2.38 2017-11-26 06:57:36
#> 300   2.28 2017-11-26 06:55:09
#> 301   2.37 2017-11-26 06:52:41
#> 302   2.40 2017-11-26 06:50:13
#> 303   2.44 2017-11-26 06:47:44
#> 304   2.63 2017-11-26 06:45:16
#> 305   2.78 2017-11-26 06:42:49
#> 306   2.87 2017-11-26 06:40:21
#> 307   2.98 2017-11-26 06:37:53
#> 308   2.67 2017-11-26 06:35:25
#> 309   2.74 2017-11-26 06:32:57
#> 310   2.83 2017-11-26 06:30:29
#> 311   2.71 2017-11-26 06:28:02
#> 312   2.70 2017-11-26 06:25:34
#> 313   2.69 2017-11-26 06:23:06
#> 314   2.59 2017-11-26 06:20:38
#> 315   2.59 2017-11-26 06:18:11
#> 316   2.61 2017-11-26 06:15:43
#> 317   2.66 2017-11-26 06:13:15
#> 318   2.69 2017-11-26 06:10:47
#> 319   2.59 2017-11-26 06:08:20
#> 320   2.54 2017-11-26 06:05:52
#> 321   2.65 2017-11-26 06:03:24
#> 322   2.70 2017-11-26 06:00:56
#> 323   2.73 2017-11-26 05:58:28
#> 324   2.70 2017-11-26 05:56:00
#> 325   2.71 2017-11-26 05:53:33
#> 326   2.79 2017-11-26 05:51:05
#> 327   2.85 2017-11-26 05:48:37
#> 328   2.79 2017-11-26 05:46:09
#> 329   2.75 2017-11-26 05:43:42
#> 330   2.82 2017-11-26 05:41:13
#> 331   2.84 2017-11-26 05:38:45
#> 332   2.91 2017-11-26 05:36:17
#> 333   2.96 2017-11-26 05:33:49
#> 334   3.01 2017-11-26 05:31:22
#> 335   2.94 2017-11-26 05:28:53
#> 336   2.98 2017-11-26 05:26:26
#> 337   2.92 2017-11-26 05:23:58
#> 338   3.03 2017-11-26 05:21:30
#> 339   3.09 2017-11-26 05:19:02
#> 340   2.98 2017-11-26 05:16:35
#> 341   2.91 2017-11-26 05:14:07
#> 342   2.95 2017-11-26 05:11:39
#> 343   2.99 2017-11-26 05:09:12
#> 344   3.03 2017-11-26 05:06:44
#> 345   3.09 2017-11-26 05:04:16
#> 346   3.02 2017-11-26 05:01:48
#> 347   2.93 2017-11-26 04:59:20
#> 348   2.91 2017-11-26 04:56:52
#> 349   2.80 2017-11-26 04:54:25
#> 350   2.80 2017-11-26 04:51:57
#> 351   2.89 2017-11-26 04:49:29
#> 352   2.83 2017-11-26 04:47:02
#> 353   2.94 2017-11-26 04:44:34
#> 354   2.93 2017-11-26 04:42:06
#> 355   3.05 2017-11-26 04:39:38
#> 356   3.07 2017-11-26 04:37:11
#> 357   3.13 2017-11-26 04:34:43
#> 358   3.20 2017-11-26 04:32:15
#> 359   3.13 2017-11-26 04:29:47
#> 360   3.16 2017-11-26 04:27:20
#> 361   3.08 2017-11-26 04:24:52
#> 362   3.19 2017-11-26 04:22:24
#> 363   3.23 2017-11-26 04:19:56
#> 364   3.22 2017-11-26 04:17:28
#> 365   3.23 2017-11-26 04:15:00
#> 366   3.26 2017-11-26 04:12:33
#> 367   3.22 2017-11-26 04:10:05
#> 368   3.19 2017-11-26 04:07:37
#> 369   3.17 2017-11-26 04:05:09
#> 370   3.15 2017-11-26 04:02:42
#> 371   3.16 2017-11-26 04:00:14
#> 372   3.22 2017-11-26 03:57:46
#> 373   3.23 2017-11-26 03:55:18
#> 374   3.14 2017-11-26 03:52:51
#> 375   3.10 2017-11-26 03:50:23
#> 376   3.09 2017-11-26 03:47:55
#> 377   3.11 2017-11-26 03:45:27
#> 378   3.11 2017-11-26 03:43:00
#> 379   3.15 2017-11-26 03:40:32
#> 380   3.17 2017-11-26 03:38:04
#> 381   3.16 2017-11-26 03:35:36
#> 382   3.21 2017-11-26 03:33:08
#> 383   3.21 2017-11-26 03:30:41
#> 384   3.23 2017-11-26 03:28:13
#> 385   3.26 2017-11-26 03:25:45
#> 386   3.23 2017-11-26 03:23:17
#> 387   3.21 2017-11-26 03:20:49
#> 388   3.23 2017-11-26 03:18:22
#> 389   3.29 2017-11-26 03:15:54
#> 390   3.22 2017-11-26 03:13:26
#> 391   3.25 2017-11-26 03:10:59
#> 392   3.20 2017-11-26 03:08:31
#> 393   3.18 2017-11-26 03:06:03
#> 394   3.22 2017-11-26 03:03:35
#> 395   3.15 2017-11-26 03:01:08
#> 396   3.17 2017-11-26 02:58:40
#> 397   3.17 2017-11-26 02:56:12
#> 398   3.17 2017-11-26 02:53:44
#> 399   3.16 2017-11-26 02:51:17
#> 400   3.24 2017-11-26 02:48:49
#> 401   3.17 2017-11-26 02:46:21
#> 402   3.22 2017-11-26 02:43:53
#> 403   3.29 2017-11-26 02:41:26
#> 404   3.38 2017-11-26 02:38:58
#> 405   3.41 2017-11-26 02:36:30
#> 406   3.41 2017-11-26 02:34:02
#> 407   3.39 2017-11-26 02:31:23
#> 408   3.40 2017-11-26 02:28:43
#> 409   3.44 2017-11-26 02:26:16
#> 410   3.50 2017-11-26 02:23:48
#> 411   3.44 2017-11-26 02:21:20
#> 412   3.50 2017-11-26 02:18:52
#> 413   3.51 2017-11-26 02:16:25
#> 414   3.47 2017-11-26 02:13:57
#> 415   3.44 2017-11-26 02:11:29
#> 416   3.48 2017-11-26 02:09:01
#> 417   3.51 2017-11-26 02:06:33
#> 418   3.53 2017-11-26 02:04:06
#> 419   3.53 2017-11-26 02:01:38
#> 420   3.55 2017-11-26 01:59:10
#> 421   3.59 2017-11-26 01:56:42
#> 422   3.63 2017-11-26 01:54:15
#> 423   3.61 2017-11-26 01:51:47
#> 424   3.59 2017-11-26 01:49:19
#> 425   3.59 2017-11-26 01:46:51
#> 426   3.62 2017-11-26 01:44:24
#> 427   3.60 2017-11-26 01:41:56
#> 428   3.58 2017-11-26 01:39:28
#> 429   3.53 2017-11-26 01:37:00
#> 430   3.52 2017-11-26 01:34:32
#> 431   3.53 2017-11-26 01:32:05
#> 432   3.58 2017-11-26 01:29:37
#> 433   3.59 2017-11-26 01:27:09
#> 434   3.62 2017-11-26 01:24:41
#> 435   3.74 2017-11-26 01:22:14
#> 436   3.70 2017-11-26 01:19:46
#> 437   3.62 2017-11-26 01:17:18
#> 438   3.65 2017-11-26 01:14:50
#> 439   3.57 2017-11-26 01:12:23
#> 440   3.58 2017-11-26 01:09:55
#> 441   3.64 2017-11-26 01:07:27
#> 442   3.67 2017-11-26 01:05:00
#> 443   3.64 2017-11-26 01:02:32
#> 444   3.62 2017-11-26 01:00:04
#> 445   3.65 2017-11-26 00:57:36
#> 446   3.62 2017-11-26 00:55:09
#> 447   3.61 2017-11-26 00:52:41
#> 448   3.61 2017-11-26 00:50:13
#> 449   3.63 2017-11-26 00:47:45
#> 450   3.65 2017-11-26 00:45:18
#> 451   3.69 2017-11-26 00:42:50
#> 452   3.69 2017-11-26 00:40:22
#> 453   3.68 2017-11-26 00:37:54
#> 454   3.70 2017-11-26 00:35:27
#> 455   3.70 2017-11-26 00:32:58
#> 456   3.70 2017-11-26 00:30:31
#> 457   3.73 2017-11-26 00:28:03
#> 458   3.67 2017-11-26 00:25:35
#> 459   3.75 2017-11-26 00:23:07
#> 460   3.69 2017-11-26 00:20:40
#> 461   3.70 2017-11-26 00:18:12
#> 462   3.72 2017-11-26 00:15:44
#> 463   3.76 2017-11-26 00:13:16
#> 464   3.75 2017-11-26 00:10:49
#> 465   3.72 2017-11-26 00:08:21
#> 466   3.73 2017-11-26 00:05:53
#> 467   3.71 2017-11-26 00:03:25
#> 468   3.74 2017-11-26 00:00:58
#> 469   3.79 2017-11-25 23:58:30
#> 470   3.79 2017-11-25 23:56:02
#> 471   3.80 2017-11-25 23:53:34
#> 472   3.81 2017-11-25 23:51:07
#> 473   3.73 2017-11-25 23:48:39
#> 474   3.74 2017-11-25 23:46:11
#> 475   3.78 2017-11-25 23:43:43
#> 476   3.75 2017-11-25 23:41:16
#> 477   3.72 2017-11-25 23:38:48
#> 478   3.70 2017-11-25 23:36:20
#> 479   3.70 2017-11-25 23:33:52
#> 480   3.65 2017-11-25 23:31:24
#> 481   3.67 2017-11-25 23:28:57
#> 482   3.64 2017-11-25 23:26:29
#> 483   3.51 2017-11-25 23:24:01
#> 484   3.64 2017-11-25 23:21:33
#> 485   3.71 2017-11-25 23:19:05
#> 486   3.73 2017-11-25 23:16:38
#> 487   3.75 2017-11-25 23:14:10
#> 488   3.74 2017-11-25 23:11:42
#> 489   3.82 2017-11-25 23:09:14
#> 490   3.72 2017-11-25 23:06:46
#> 491   3.71 2017-11-25 23:04:18
#> 492   3.75 2017-11-25 23:01:51
#> 493   3.82 2017-11-25 22:59:23
#> 494   3.81 2017-11-25 22:56:55
#> 495   3.81 2017-11-25 22:54:27
#> 496   3.81 2017-11-25 22:52:00
#> 497   3.90 2017-11-25 22:49:31
#> 498   3.97 2017-11-25 22:47:04
#> 499   3.96 2017-11-25 22:44:36
#> 500   3.91 2017-11-25 22:42:08
#> 501   3.90 2017-11-25 22:39:40
#> 502   3.88 2017-11-25 22:37:12
#> 503   3.84 2017-11-25 22:34:45
#> 504   3.90 2017-11-25 22:32:17
#> 505   4.02 2017-11-25 22:29:48
#> 506   3.97 2017-11-25 22:27:20
#> 507   3.86 2017-11-25 22:24:52
#> 508   3.84 2017-11-25 22:22:24
#> 509   3.82 2017-11-25 22:19:56
#> 510   3.79 2017-11-25 22:17:28
#> 511   3.70 2017-11-25 22:15:00
#> 512   3.63 2017-11-25 22:12:33
#> 513   3.72 2017-11-25 22:10:05
#> 514   3.79 2017-11-25 22:07:37
#> 515   3.86 2017-11-25 22:05:09
#> 516   3.91 2017-11-25 22:02:42
#> 517   4.04 2017-11-25 22:00:14
#> 518   4.11 2017-11-25 21:57:46
#> 519   4.16 2017-11-25 21:55:18
#> 520   4.21 2017-11-25 21:52:51
#> 521   4.23 2017-11-25 21:50:23
#> 522   4.11 2017-11-25 21:47:55
#> 523   4.05 2017-11-25 21:45:27
#> 524   4.03 2017-11-25 21:42:59
#> 525   4.04 2017-11-25 21:40:32
#> 526   4.02 2017-11-25 21:38:04
#> 527   3.94 2017-11-25 21:35:36
#> 528   3.96 2017-11-25 21:33:08
#> 529   3.97 2017-11-25 21:30:41
#> 530   3.97 2017-11-25 21:28:13
#> 531   4.01 2017-11-25 21:25:45
#> 532   4.04 2017-11-25 21:23:17
#> 533   4.07 2017-11-25 21:20:50
#> 534   4.04 2017-11-25 21:18:22
#> 535   4.09 2017-11-25 21:15:54
#> 536   4.12 2017-11-25 21:13:26
#> 537   4.14 2017-11-25 21:10:59
#> 538   4.14 2017-11-25 21:08:29
#> 539   4.14 2017-11-25 21:06:02
#> 540   4.17 2017-11-25 21:03:33
#> 541   4.24 2017-11-25 21:01:05
#> 542   4.38 2017-11-25 20:58:37
#> 543   4.37 2017-11-25 20:56:09
#> 544   4.40 2017-11-25 20:53:42
#> 545   4.41 2017-11-25 20:51:14
#> 546   4.36 2017-11-25 20:48:46
#> 547   4.34 2017-11-25 20:46:18
#> 548   4.34 2017-11-25 20:43:51
#> 549   4.35 2017-11-25 20:41:23
#> 550   4.38 2017-11-25 20:38:55
#> 551   4.38 2017-11-25 20:36:28
#> 552   4.48 2017-11-25 20:34:00
#> 553   4.48 2017-11-25 20:31:32
#> 554   4.57 2017-11-25 20:29:04
#> 555   4.56 2017-11-25 20:26:37
#> 556   4.56 2017-11-25 20:24:09
#> 557   4.64 2017-11-25 20:21:41
#> 558   4.65 2017-11-25 20:19:13
#> 559   4.59 2017-11-25 20:16:45
#> 560   4.35 2017-11-25 20:14:18
#> 561   4.25 2017-11-25 20:11:50
#> 562   4.33 2017-11-25 20:09:22
#> 563   4.32 2017-11-25 20:06:54
#> 564   4.20 2017-11-25 20:04:27
#> 565   4.17 2017-11-25 20:01:59
#> 566   4.27 2017-11-25 19:59:31
#> 567   4.28 2017-11-25 19:57:03
#> 568   4.33 2017-11-25 19:54:35
#> 569   4.33 2017-11-25 19:52:07
#> 570   4.33 2017-11-25 19:49:38
#> 571   4.34 2017-11-25 19:47:10
#> 572   4.38 2017-11-25 19:44:43
#> 573   4.34 2017-11-25 19:42:15
#> 574   4.32 2017-11-25 19:39:47
#> 575   4.30 2017-11-25 19:37:19
#> 576   4.28 2017-11-25 19:34:51
#> 577   4.15 2017-11-25 19:32:23
#> 578   4.13 2017-11-25 19:29:56
#> 579   4.14 2017-11-25 19:27:28
#> 580   4.16 2017-11-25 19:25:00
#> 581   4.23 2017-11-25 19:22:31
#> 582   4.29 2017-11-25 19:20:03
#> 583   4.21 2017-11-25 19:17:36
#> 584   4.14 2017-11-25 19:15:08
#> 585   4.16 2017-11-25 19:12:40
#> 586   4.20 2017-11-25 19:10:12
#> 587   4.22 2017-11-25 19:07:44
#> 588   4.42 2017-11-25 19:05:17
#> 589   4.60 2017-11-25 19:02:49
#> 590   4.50 2017-11-25 19:00:21
#> 591   4.37 2017-11-25 18:57:52
#> 592   4.30 2017-11-25 18:55:24
#> 593   4.30 2017-11-25 18:52:56
#> 594   4.27 2017-11-25 18:50:28
#> 595   4.33 2017-11-25 18:48:00
#> 596   4.45 2017-11-25 18:45:33
#> 597   4.46 2017-11-25 18:43:04
#> 598   4.47 2017-11-25 18:40:36
#> 599   4.53 2017-11-25 18:38:09
#> 600   4.53 2017-11-25 18:35:41
#> 601   4.41 2017-11-25 18:33:13
#> 602   4.39 2017-11-25 18:30:45
#> 603   4.39 2017-11-25 18:28:17
#> 604   4.44 2017-11-25 18:25:50
#> 605   4.50 2017-11-25 18:23:22
#> 606   4.50 2017-11-25 18:20:54
#> 607   4.51 2017-11-25 18:18:26
#> 608   4.50 2017-11-25 18:15:58
#> 609   4.41 2017-11-25 18:13:31
#> 610   4.38 2017-11-25 18:11:03
#> 611   4.47 2017-11-25 18:08:35
#> 612   4.49 2017-11-25 18:06:07
#> 613   4.44 2017-11-25 18:03:40
#> 614   4.52 2017-11-25 18:01:12
#> 615   4.57 2017-11-25 17:58:44
#> 616   4.74 2017-11-25 17:56:16
#> 617   4.73 2017-11-25 17:53:49
#> 618   4.88 2017-11-25 17:51:20
#> 619   4.64 2017-11-25 17:48:52
#> 620   4.69 2017-11-25 17:46:23
#> 621   4.71 2017-11-25 17:43:56
#> 622   4.74 2017-11-25 17:41:27
#> 623   4.94 2017-11-25 17:38:59
#> 624   4.97 2017-11-25 17:36:32
#> 625   4.95 2017-11-25 17:34:04
#> 626   4.87 2017-11-25 17:31:36
#> 627   4.85 2017-11-25 17:29:08
#> 628   4.93 2017-11-25 17:26:40
#> 629   4.94 2017-11-25 17:24:12
#> 630   4.89 2017-11-25 17:21:45
#> 631   4.88 2017-11-25 17:19:17
#> 632   4.90 2017-11-25 17:16:49
#> 633   4.87 2017-11-25 17:14:22
#> 634   4.91 2017-11-25 17:11:54
#> 635   4.88 2017-11-25 17:09:26
#> 636   4.87 2017-11-25 17:06:58
#> 637   5.03 2017-11-25 17:04:30
#> 638   4.96 2017-11-25 17:02:02
#> 639   4.99 2017-11-25 16:59:34
#> 640   4.96 2017-11-25 16:57:06
#> 641   5.03 2017-11-25 16:54:39
#> 642   5.02 2017-11-25 16:52:11
#> 643   5.00 2017-11-25 16:49:43
#> 644   5.02 2017-11-25 16:47:15
#> 645   5.05 2017-11-25 16:44:48
#> 646   5.09 2017-11-25 16:42:20
#> 647   4.96 2017-11-25 16:39:52
#> 648   4.98 2017-11-25 16:37:24
#> 649   4.92 2017-11-25 16:34:56
#> 650   4.84 2017-11-25 16:32:29
#> 651   4.94 2017-11-25 16:30:01
#> 652   4.88 2017-11-25 16:27:33
#> 653   4.94 2017-11-25 16:25:05
#> 654   4.93 2017-11-25 16:22:37
#> 655   4.92 2017-11-25 16:20:09
#> 656   5.02 2017-11-25 16:17:41
#> 657   5.01 2017-11-25 16:15:13
#> 658   4.96 2017-11-25 16:12:46
#> 659   5.08 2017-11-25 16:10:17
#> 660   5.12 2017-11-25 16:07:49
#> 661   5.20 2017-11-25 16:05:21
#> 662   5.10 2017-11-25 16:02:53
#> 663   5.04 2017-11-25 16:00:25
#> 664   5.05 2017-11-25 15:57:58
#> 665   5.07 2017-11-25 15:55:30
#> 666   5.12 2017-11-25 15:53:02
#> 667   5.13 2017-11-25 15:50:34
#> 668   5.18 2017-11-25 15:48:06
#> 669   5.03 2017-11-25 15:45:39
#> 670   5.08 2017-11-25 15:43:11
#> 671   5.13 2017-11-25 15:40:43
#> 672   5.26 2017-11-25 15:38:14
#> 673   5.25 2017-11-25 15:35:47
#> 674   5.18 2017-11-25 15:33:19
#> 675   5.08 2017-11-25 15:30:50
#> 676   4.96 2017-11-25 15:28:22
#> 677   5.13 2017-11-25 15:25:54
#> 678   5.20 2017-11-25 15:23:27
#> 679   5.25 2017-11-25 15:20:59
#> 680   5.22 2017-11-25 15:18:31
#> 681   5.38 2017-11-25 15:16:03
#> 682   5.46 2017-11-25 15:13:35
#> 683   5.47 2017-11-25 15:11:07
#> 684   5.46 2017-11-25 15:08:39
#> 685   5.44 2017-11-25 15:06:11
#> 686   5.40 2017-11-25 15:03:44
#> 687   5.53 2017-11-25 15:01:16
#> 688   5.50 2017-11-25 14:58:48
#> 689   5.58 2017-11-25 14:56:20
#> 690   5.62 2017-11-25 14:53:53
#> 691   5.64 2017-11-25 14:51:24
#> 692   5.69 2017-11-25 14:48:56
#> 693   5.68 2017-11-25 14:46:28
#> 694   5.67 2017-11-25 14:44:01
#> 695   5.65 2017-11-25 14:41:33
#> 696   5.64 2017-11-25 14:39:05
#> 697   5.79 2017-11-25 14:36:37
#> 698   6.00 2017-11-25 14:34:10
#> 699   6.21 2017-11-25 14:31:42
#> 700   6.19 2017-11-25 14:29:14
#> 701   6.04 2017-11-25 14:26:46
#> 702   6.08 2017-11-25 14:24:19
#> 703   6.07 2017-11-25 14:21:51
#> 704   6.15 2017-11-25 14:19:23
#> 705   6.20 2017-11-25 14:16:54
#> 706   6.15 2017-11-25 14:14:27
#> 707   6.18 2017-11-25 14:11:59
#> 708   6.21 2017-11-25 14:09:31
#> 709   6.38 2017-11-25 14:07:03
#> 710   6.57 2017-11-25 14:04:35
#> 711   6.46 2017-11-25 14:02:07
#> 712   6.61 2017-11-25 13:59:40
#> 713   6.57 2017-11-25 13:57:12
#> 714   6.49 2017-11-25 13:54:44
#> 715   6.53 2017-11-25 13:52:16
#> 716   6.48 2017-11-25 13:49:48
#> 717   6.43 2017-11-25 13:47:21
#> 718   6.33 2017-11-25 13:44:53
#> 719   6.28 2017-11-25 13:42:24
#> 720   6.53 2017-11-25 13:39:57
#> 721   6.64 2017-11-25 13:37:29
#> 722   6.72 2017-11-25 13:35:01
#> 723   6.66 2017-11-25 13:32:33
#> 724   6.55 2017-11-25 13:30:05
#> 725   6.54 2017-11-25 13:27:36
#> 726   6.69 2017-11-25 13:25:08
#> 727   6.66 2017-11-25 13:22:40
#> 728   6.77 2017-11-25 13:20:13
#> 729   6.55 2017-11-25 13:17:45
#> 730   6.70 2017-11-25 13:15:17
#> 731   7.05 2017-11-25 13:12:49
#> 732   6.93 2017-11-25 13:10:22
#> 733   6.83 2017-11-25 13:07:54
#> 734   6.58 2017-11-25 13:05:26
#> 735   6.78 2017-11-25 13:02:58
#> 736   6.95 2017-11-25 13:00:30
#> 737   7.07 2017-11-25 12:58:02
#> 738   7.36 2017-11-25 12:55:33
#> 739   7.12 2017-11-25 12:53:05
#> 740   7.51 2017-11-25 12:50:37
#> 741   7.78 2017-11-25 12:48:10
#> 742   8.39 2017-11-25 12:45:42
#> 743   8.78 2017-11-25 12:43:14
#> 744   8.89 2017-11-25 12:40:46
#> 745   8.81 2017-11-25 12:38:18
#> 746   8.73 2017-11-25 12:35:51
#> 747   8.59 2017-11-25 12:33:22
#> 748   8.44 2017-11-25 12:30:53
#> 749   8.85 2017-11-25 12:28:25
#> 750   8.99 2017-11-25 12:25:57
#> 751   8.55 2017-11-25 12:23:29
#> 752   8.29 2017-11-25 12:21:01
#> 753   8.37 2017-11-25 12:18:34
#> 754   8.23 2017-11-25 12:16:06
#> 755   8.01 2017-11-25 12:13:38
#> 756   8.17 2017-11-25 12:11:10
#> 757   8.00 2017-11-25 12:08:42
#> 758   7.74 2017-11-25 12:06:14
#> 759   7.68 2017-11-25 12:03:47
#> 760   8.60 2017-11-25 12:01:19
#> 761   8.26 2017-11-25 11:58:51
#> 762   7.93 2017-11-25 11:56:23
#> 763   7.52 2017-11-25 11:53:55
#> 764   7.13 2017-11-25 11:51:27
#> 765   7.39 2017-11-25 11:48:59
#> 766   6.97 2017-11-25 11:46:31
#> 767   7.59 2017-11-25 11:44:03
#> 768   7.27 2017-11-25 11:41:35
#> 769   6.86 2017-11-25 11:39:08
#> 770   6.45 2017-11-25 11:36:40
#> 771   5.66 2017-11-25 11:34:12
#> 772   6.66 2017-11-25 11:31:44
#> 773   6.14 2017-11-25 11:29:16
#> 774   5.89 2017-11-25 11:26:49
#> 775   5.81 2017-11-25 11:24:21
#> 776   5.44 2017-11-25 11:21:53
#> 777   5.78 2017-11-25 11:19:25
#> 778   5.32 2017-11-25 11:16:57
#> 779   4.94 2017-11-25 11:14:30
#> 780   4.87 2017-11-25 11:12:02
#> 781   4.16 2017-11-25 11:09:34
#> 782   4.30 2017-11-25 11:07:06
#> 783   4.52 2017-11-25 11:04:38
#> 784   4.09 2017-11-25 11:02:11
#> 785   3.71 2017-11-25 10:59:42
#> 786   3.74 2017-11-25 10:57:14
#> 787   3.73 2017-11-25 10:54:46
#> 788   3.29 2017-11-25 10:52:19
#> 789   2.99 2017-11-25 10:49:51
#> 790   2.74 2017-11-25 10:47:23
#> 791   2.34 2017-11-25 10:44:55
#> 792   2.41 2017-11-25 10:42:27
#> 793   2.70 2017-11-25 10:39:59
#> 794   2.86 2017-11-25 10:37:31
#> 795   2.71 2017-11-25 10:35:03
#> 796   2.79 2017-11-25 10:32:35
#> 797   2.68 2017-11-25 10:30:08
#> 798   2.42 2017-11-25 10:27:40
#> 799   1.96 2017-11-25 10:25:12
#> 800   1.98 2017-11-25 10:22:44
#> 801   1.84 2017-11-25 10:20:16
#> 802   1.87 2017-11-25 10:17:49
#> 803   2.07 2017-11-25 10:15:21
#> 804   2.22 2017-11-25 10:12:52
#> 805   2.10 2017-11-25 10:10:24
#> 806   1.81 2017-11-25 10:07:56
#> 807   1.90 2017-11-25 10:05:28
#> 808   1.96 2017-11-25 10:02:59
#> 809   2.02 2017-11-25 10:00:31
#> 810   1.97 2017-11-25 09:58:03
#> 811   1.70 2017-11-25 09:55:35
#> 812   1.60 2017-11-25 09:53:07
#> 813   1.52 2017-11-25 09:50:39
#> 814   1.73 2017-11-25 09:48:11
#> 815   1.83 2017-11-25 09:45:43
#> 816   1.38 2017-11-25 09:43:15
#> 817   1.64 2017-11-25 09:40:47
#> 818   1.99 2017-11-25 09:38:19
#> 819   1.82 2017-11-25 09:35:51
#> 820   1.48 2017-11-25 09:33:23
#> 821   1.04 2017-11-25 09:30:55
#> 822   0.85 2017-11-25 09:28:27
#> 823   0.74 2017-11-25 09:25:59
#> 824   0.75 2017-11-25 09:23:31
#> 825   0.68 2017-11-25 09:21:03
#> 826   0.54 2017-11-25 09:18:35
#> 827   0.37 2017-11-25 09:16:07
#> 828   0.42 2017-11-25 09:13:39
#> 829   0.85 2017-11-25 09:11:12
#> 830   0.67 2017-11-25 09:08:44
#> 831   0.65 2017-11-25 09:06:16
#> 832   0.74 2017-11-25 09:03:48
#> 833   0.47 2017-11-25 09:01:21
#> 834   0.48 2017-11-25 08:58:53
#> 835   0.48 2017-11-25 08:56:25
#> 836   0.59 2017-11-25 08:53:57
#> 837   0.54 2017-11-25 08:51:29
#> 838   0.09 2017-11-25 08:49:00
#> 839   0.13 2017-11-25 08:46:33
#> 840   0.42 2017-11-25 08:44:05
#> 841   0.16 2017-11-25 08:41:37
#> 842   0.32 2017-11-25 08:39:09
#> 843   0.34 2017-11-25 08:36:42
#> 844   0.13 2017-11-25 08:34:14
#> 845   0.03 2017-11-25 08:31:45
#> 846   0.28 2017-11-25 08:29:17
#> 847   0.06 2017-11-25 08:26:50
#> 848   0.12 2017-11-25 08:24:22
#> 849   0.07 2017-11-25 08:21:54
#> 850  -0.07 2017-11-25 08:19:26
#> 851  -0.21 2017-11-25 08:16:58
#> 852  -0.14 2017-11-25 08:14:30
#> 853   0.03 2017-11-25 08:12:02
#> 854   0.12 2017-11-25 08:09:34
#> 855   0.47 2017-11-25 08:07:07
#> 856   0.62 2017-11-25 08:04:39
#> 857   0.30 2017-11-25 08:02:11
#> 858   0.27 2017-11-25 07:59:43
#> 859   0.23 2017-11-25 07:57:16
#> 860   0.16 2017-11-25 07:54:48
#> 861  -0.05 2017-11-25 07:52:20
#> 862  -0.08 2017-11-25 07:49:52
#> 863   0.10 2017-11-25 07:47:24
#> 864  -0.13 2017-11-25 07:44:57
#> 865  -0.31 2017-11-25 07:42:29
#> 866  -0.42 2017-11-25 07:40:01
#> 867  -0.46 2017-11-25 07:37:33
#> 868  -0.64 2017-11-25 07:35:05
#> 869  -0.49 2017-11-25 07:32:37
#> 870  -0.42 2017-11-25 07:30:09
#> 871  -0.46 2017-11-25 07:27:41
#> 872  -0.37 2017-11-25 07:25:13
#> 873  -0.48 2017-11-25 07:22:46
#> 874  -0.50 2017-11-25 07:20:18
#> 875  -0.43 2017-11-25 07:17:50
#> 876  -0.37 2017-11-25 07:15:22
#> 877  -0.38 2017-11-25 07:12:54
#> 878  -0.36 2017-11-25 07:10:27
#> 879  -0.41 2017-11-25 07:07:59
#> 880  -0.32 2017-11-25 07:05:31
#> 881  -0.28 2017-11-25 07:03:03
#> 882  -0.31 2017-11-25 07:00:36
#> 883  -0.40 2017-11-25 06:58:08
#> 884  -0.25 2017-11-25 06:55:40
#> 885  -0.30 2017-11-25 06:53:12
#> 886  -0.29 2017-11-25 06:50:45
#> 887  -0.30 2017-11-25 06:48:17
#> 888  -0.19 2017-11-25 06:45:49
#> 889  -0.11 2017-11-25 06:43:21
#> 890  -0.26 2017-11-25 06:40:53
#> 891  -0.15 2017-11-25 06:38:26
#> 892  -0.08 2017-11-25 06:35:58
#> 893  -0.15 2017-11-25 06:33:30
#> 894  -0.15 2017-11-25 06:31:03
#> 895  -0.11 2017-11-25 06:28:35
#> 896  -0.17 2017-11-25 06:26:07
#> 897   0.01 2017-11-25 06:23:39
#> 898   0.02 2017-11-25 06:21:11
#> 899  -0.04 2017-11-25 06:18:44
#> 900  -0.07 2017-11-25 06:16:16
#> 901   0.05 2017-11-25 06:13:48
#> 902   0.04 2017-11-25 06:11:20
#> 903   0.02 2017-11-25 06:08:53
#> 904   0.01 2017-11-25 06:06:25
#> 905   0.03 2017-11-25 06:03:57
#> 906   0.10 2017-11-25 06:01:29
#> 907   0.12 2017-11-25 05:59:01
#> 908   0.22 2017-11-25 05:56:33
#> 909   0.20 2017-11-25 05:54:05
#> 910   0.22 2017-11-25 05:51:37
#> 911   0.36 2017-11-25 05:49:09
#> 912   0.24 2017-11-25 05:46:42
#> 913   0.26 2017-11-25 05:44:14
#> 914   0.27 2017-11-25 05:41:46
#> 915   0.39 2017-11-25 05:39:18
#> 916   0.36 2017-11-25 05:36:50
#> 917   0.28 2017-11-25 05:34:23
#> 918   0.30 2017-11-25 05:31:55
#> 919   0.15 2017-11-25 05:29:27
#> 920   0.13 2017-11-25 05:26:59
#> 921   0.14 2017-11-25 05:24:32
#> 922   0.24 2017-11-25 05:22:04
#> 923   0.21 2017-11-25 05:19:36
#> 924   0.40 2017-11-25 05:17:08
#> 925   0.48 2017-11-25 05:14:40
#> 926   0.57 2017-11-25 05:12:13
#> 927   0.53 2017-11-25 05:09:45
#> 928   0.51 2017-11-25 05:07:17
#> 929   0.49 2017-11-25 05:04:50
#> 930   0.44 2017-11-25 05:02:22
#> 931   0.41 2017-11-25 04:59:54
#> 932   0.36 2017-11-25 04:57:26
#> 933   0.27 2017-11-25 04:54:58
#> 934   0.45 2017-11-25 04:52:31
#> 935   0.44 2017-11-25 04:50:03
#> 936   0.52 2017-11-25 04:47:35
#> 937   0.56 2017-11-25 04:45:07
#> 938   0.61 2017-11-25 04:42:39
#> 939   0.56 2017-11-25 04:40:12
#> 940   0.45 2017-11-25 04:37:44
#> 941   0.46 2017-11-25 04:35:16
#> 942   0.49 2017-11-25 04:32:49
#> 943   0.48 2017-11-25 04:30:21
#> 944   0.62 2017-11-25 04:27:53
#> 945   0.62 2017-11-25 04:25:25
#> 946   0.63 2017-11-25 04:22:58
#> 947   0.66 2017-11-25 04:20:30
#> 948   0.57 2017-11-25 04:18:02
#> 949   0.47 2017-11-25 04:15:34
#> 950   0.77 2017-11-25 04:13:07
#> 951   0.85 2017-11-25 04:10:39
#> 952   0.89 2017-11-25 04:08:11
#> 953   0.82 2017-11-25 04:05:44
#> 954   0.77 2017-11-25 04:03:16
#> 955   0.73 2017-11-25 04:00:48
#> 956   0.66 2017-11-25 03:58:20
#> 957   0.67 2017-11-25 03:55:53
#> 958   0.69 2017-11-25 03:53:25
#> 959   0.65 2017-11-25 03:50:57
#> 960   0.68 2017-11-25 03:48:29
#> 961   0.55 2017-11-25 03:46:02
#> 962   0.45 2017-11-25 03:43:34
#> 963   0.45 2017-11-25 03:41:06
#> 964   0.42 2017-11-25 03:38:38
#> 965   0.52 2017-11-25 03:36:11
#> 966   0.55 2017-11-25 03:33:43
#> 967   0.65 2017-11-25 03:31:15
#> 968   0.74 2017-11-25 03:28:47
#> 969   0.72 2017-11-25 03:26:20
#> 970   0.77 2017-11-25 03:23:52
#> 971   0.77 2017-11-25 03:21:24
#> 972   0.77 2017-11-25 03:18:56
#> 973   0.73 2017-11-25 03:16:29
#> 974   0.76 2017-11-25 03:14:01
#> 975   0.72 2017-11-25 03:11:33
#> 976   0.80 2017-11-25 03:09:05
#> 977   0.66 2017-11-25 03:06:38
#> 978   0.71 2017-11-25 03:04:10
#> 979   0.81 2017-11-25 03:01:42
#> 980   0.89 2017-11-25 02:59:14
#> 981   0.91 2017-11-25 02:56:47
#> 982   0.85 2017-11-25 02:54:19
#> 983   0.88 2017-11-25 02:51:51
#> 984   0.84 2017-11-25 02:49:23
#> 985   0.85 2017-11-25 02:46:55
#> 986   0.85 2017-11-25 02:44:28
#> 987   0.83 2017-11-25 02:42:00
#> 988   0.73 2017-11-25 02:39:32
#> 989   0.73 2017-11-25 02:37:04
#> 990   0.59 2017-11-25 02:34:37
#> 991   0.70 2017-11-25 02:32:09
#> 992   0.66 2017-11-25 02:29:41
#> 993   0.73 2017-11-25 02:27:13
#> 994   0.72 2017-11-25 02:24:45
#> 995   0.76 2017-11-25 02:22:18
#> 996   0.90 2017-11-25 02:19:50
#> 997   0.94 2017-11-25 02:17:22
#> 998   0.91 2017-11-25 02:14:54
#> 999   0.79 2017-11-25 02:12:27
#> 1000  0.72 2017-11-25 02:09:59
#> 1001  0.85 2017-11-25 02:07:31
#> 1002  0.69 2017-11-25 02:04:59
#> 1003  0.76 2017-11-25 02:02:31
#> 1004  0.94 2017-11-25 02:00:03
#> 1005  0.94 2017-11-25 01:57:35
#> 1006  0.84 2017-11-25 01:55:08
#> 1007  0.74 2017-11-25 01:52:40
#> 1008  0.89 2017-11-25 01:50:12
#> 1009  0.77 2017-11-25 01:47:44
#> 1010  0.84 2017-11-25 01:45:16
#> 1011  0.69 2017-11-25 01:42:49
#> 1012  0.69 2017-11-25 01:40:21
#> 1013  0.71 2017-11-25 01:37:53
#> 1014  0.73 2017-11-25 01:35:25
#> 1015  0.74 2017-11-25 01:32:58
#> 1016  0.76 2017-11-25 01:30:30
#> 1017  0.71 2017-11-25 01:28:02
#> 1018  0.82 2017-11-25 01:25:34
#> 1019  0.81 2017-11-25 01:23:06
#> 1020  0.80 2017-11-25 01:20:39
#> 1021  0.79 2017-11-25 01:18:11
#> 1022  0.99 2017-11-25 01:15:43
#> 1023  0.84 2017-11-25 01:13:15
#> 1024  1.03 2017-11-25 01:10:48
#> 1025  0.92 2017-11-25 01:08:20
#> 1026  0.77 2017-11-25 01:05:52
#> 1027  0.81 2017-11-25 01:03:24
#> 1028  0.85 2017-11-25 01:00:57
#> 1029  0.92 2017-11-25 00:58:29
#> 1030  0.88 2017-11-25 00:56:01
#> 1031  0.86 2017-11-25 00:53:33
#> 1032  0.81 2017-11-25 00:51:06
#> 1033  0.85 2017-11-25 00:48:38
#> 1034  0.81 2017-11-25 00:46:10
#> 1035  0.79 2017-11-25 00:43:42
#> 1036  0.81 2017-11-25 00:41:15
#> 1037  0.87 2017-11-25 00:38:47
#> 1038  0.91 2017-11-25 00:36:19
#> 1039  1.00 2017-11-25 00:33:51
#> 1040  0.95 2017-11-25 00:31:24
#> 1041  0.95 2017-11-25 00:28:56
#> 1042  0.91 2017-11-25 00:26:28
#> 1043  1.01 2017-11-25 00:24:00
#> 1044  1.11 2017-11-25 00:21:33
#> 1045  1.21 2017-11-25 00:19:05
#> 1046  1.20 2017-11-25 00:16:37
#> 1047  1.16 2017-11-25 00:14:09
#> 1048  1.21 2017-11-25 00:11:42
#> 1049  1.16 2017-11-25 00:09:14
#> 1050  1.08 2017-11-25 00:06:46
#> 1051  1.04 2017-11-25 00:04:18
#> 1052  1.00 2017-11-25 00:01:51
#> 1053  1.08 2017-11-24 23:59:23
#> 1054  1.10 2017-11-24 23:56:55
#> 1055  1.18 2017-11-24 23:54:27
#> 1056  1.12 2017-11-24 23:51:59
#> 1057  0.95 2017-11-24 23:49:32
#> 1058  1.03 2017-11-24 23:47:04
#> 1059  1.08 2017-11-24 23:44:36
#> 1060  1.13 2017-11-24 23:42:08
#> 1061  1.19 2017-11-24 23:39:40
#> 1062  1.13 2017-11-24 23:37:13
#> 1063  1.23 2017-11-24 23:34:45
#> 1064  1.33 2017-11-24 23:32:17
#> 1065  1.29 2017-11-24 23:29:49
#> 1066  1.23 2017-11-24 23:27:21
#> 1067  1.14 2017-11-24 23:24:53
#> 1068  1.13 2017-11-24 23:22:24
#> 1069  1.14 2017-11-24 23:19:56
#> 1070  1.22 2017-11-24 23:17:28
#> 1071  1.23 2017-11-24 23:15:00
#> 1072  1.17 2017-11-24 23:12:32
#> 1073  1.19 2017-11-24 23:10:04
#> 1074  1.20 2017-11-24 23:07:36
#> 1075  1.13 2017-11-24 23:05:08
#> 1076  1.16 2017-11-24 23:02:41
#> 1077  1.26 2017-11-24 23:00:13
#> 1078  1.41 2017-11-24 22:57:45
#> 1079  1.44 2017-11-24 22:55:17
#> 1080  1.39 2017-11-24 22:52:50
#> 1081  1.38 2017-11-24 22:50:22
#> 1082  1.34 2017-11-24 22:47:54
#> 1083  1.29 2017-11-24 22:45:26
#> 1084  1.31 2017-11-24 22:42:58
#> 1085  1.29 2017-11-24 22:40:30
#> 1086  1.33 2017-11-24 22:38:02
#> 1087  1.37 2017-11-24 22:35:34
#> 1088  1.33 2017-11-24 22:33:07
#> 1089  1.30 2017-11-24 22:30:39
#> 1090  1.28 2017-11-24 22:28:11
#> 1091  1.27 2017-11-24 22:25:43
#> 1092  1.14 2017-11-24 22:23:15
#> 1093  1.18 2017-11-24 22:20:47
#> 1094  1.27 2017-11-24 22:18:20
#> 1095  1.24 2017-11-24 22:15:52
#> 1096  1.40 2017-11-24 22:13:24
#> 1097  1.24 2017-11-24 22:10:56
#> 1098  1.18 2017-11-24 22:08:29
#> 1099  1.25 2017-11-24 22:06:00
#> 1100  1.33 2017-11-24 22:03:32
#> 1101  1.44 2017-11-24 22:01:04
#> 1102  1.46 2017-11-24 21:58:36
#> 1103  1.35 2017-11-24 21:56:08
#> 1104  1.37 2017-11-24 21:53:40
#> 1105  1.37 2017-11-24 21:51:13
#> 1106  1.42 2017-11-24 21:48:44
#> 1107  1.38 2017-11-24 21:46:16
#> 1108  1.39 2017-11-24 21:43:48
#> 1109  1.33 2017-11-24 21:41:20
#> 1110  1.29 2017-11-24 21:38:53
#> 1111  1.26 2017-11-24 21:36:25
#> 1112  1.26 2017-11-24 21:33:57
#> 1113  1.25 2017-11-24 21:31:29
#> 1114  1.28 2017-11-24 21:29:01
#> 1115  1.27 2017-11-24 21:26:33
#> 1116  1.26 2017-11-24 21:24:06
#> 1117  1.25 2017-11-24 21:21:37
#> 1118  1.23 2017-11-24 21:19:09
#> 1119  1.17 2017-11-24 21:16:41
#> 1120  1.22 2017-11-24 21:14:13
#> 1121  1.31 2017-11-24 21:11:45
#> 1122  1.28 2017-11-24 21:09:17
#> 1123  1.33 2017-11-24 21:06:49
#> 1124  1.32 2017-11-24 21:04:21
#> 1125  1.31 2017-11-24 21:01:53
#> 1126  1.24 2017-11-24 20:59:25
#> 1127  1.31 2017-11-24 20:56:57
#> 1128  1.45 2017-11-24 20:54:29
#> 1129  1.38 2017-11-24 20:52:01
#> 1130  1.47 2017-11-24 20:49:32
#> 1131  1.49 2017-11-24 20:47:04
#> 1132  1.43 2017-11-24 20:44:36
#> 1133  1.45 2017-11-24 20:42:09
#> 1134  1.42 2017-11-24 20:39:41
#> 1135  1.54 2017-11-24 20:37:13
#> 1136  1.50 2017-11-24 20:34:45
#> 1137  1.58 2017-11-24 20:32:16
#> 1138  1.60 2017-11-24 20:29:48
#> 1139  1.57 2017-11-24 20:27:20
#> 1140  1.50 2017-11-24 20:24:52
#> 1141  1.56 2017-11-24 20:22:24
#> 1142  1.60 2017-11-24 20:19:56
#> 1143  1.54 2017-11-24 20:17:29
#> 1144  1.50 2017-11-24 20:15:00
#> 1145  1.62 2017-11-24 20:12:32
#> 1146  1.66 2017-11-24 20:10:04
#> 1147  1.61 2017-11-24 20:07:36
#> 1148  1.65 2017-11-24 20:05:08
#> 1149  1.67 2017-11-24 20:02:40
#> 1150  1.70 2017-11-24 20:00:12
#> 1151  1.58 2017-11-24 19:57:44
#> 1152  1.56 2017-11-24 19:55:16
#> 1153  1.62 2017-11-24 19:52:48
#> 1154  1.61 2017-11-24 19:50:20
#> 1155  1.59 2017-11-24 19:47:52
#> 1156  1.68 2017-11-24 19:45:24
#> 1157  1.66 2017-11-24 19:42:56
#> 1158  1.62 2017-11-24 19:40:28
#> 1159  1.62 2017-11-24 19:38:00
#> 1160  1.59 2017-11-24 19:35:32
#> 1161  1.57 2017-11-24 19:33:03
#> 1162  1.55 2017-11-24 19:30:35
#> 1163  1.60 2017-11-24 19:28:07
#> 1164  1.59 2017-11-24 19:25:39
#> 1165  1.53 2017-11-24 19:23:11
#> 1166  1.50 2017-11-24 19:20:44
#> 1167  1.51 2017-11-24 19:18:16
#> 
#> 

# NOT RUN {
get_senseBox_data(
  senseBoxId = "5957b67494f0520011304cc1",
  sensorId = "5957b67494f0520011304cc4",
  fromDate = "2017-11-25-12:00:00",
  toDate = "2017-11-26-12:00:00")
# }