Get data from a senseBox

If the arguments fromDate and toDate are NOT given, the measurements from the last 48 h from now are downloaded. The maximum numbers of downloaded measurements are 10,000. The maximum time frame is one month. It is possible to download the data directly as a csv file, when argument CSV is TRUE.

get_senseBox_data(senseBoxId, sensorId = rep("all", times =
  length(unlist(senseBoxId))), fromDate = NULL, toDate = NULL,
  parallel = FALSE, CSV = FALSE, POSIXct = TRUE)

Arguments

senseBoxId
characteror vector of character (required): senseBoxId or a vector of character of senseBoxIds

sensorId
character or list of character (required): sensorId or a list of senseBoxIds

fromDate
character (optional): Just show data with from date in format YYYY-mm-dd HH:MM:SS

toDate
character (optional): Just show data with up to date in format YYYY-mm-dd HH:MM:SS

parallel
logical (optional): Should the calculations be executed on multiple cores? At least 4 cores are necessary to use this feature.

CSV
logical (optional): Download data as csv file? NOT SUPPORTED UNTIL NOW!

POSIXct
logical (optional): Should the timestamp be translated into POSIXct?

Value

: A list with every entry is a sensBoxId. Every list entry inherits a data.frame with values and dates.

Function version

0.0.1

Examples

##=====================================
## Example: get data from sensorId
##=====================================

get_senseBox_data(
  senseBoxId = "5957b67494f0520011304cc1",
  sensorId = "5957b67494f0520011304cc4")

#> $`5957b67494f0520011304cc1`
#> $`5957b67494f0520011304cc1`$`Temperatur [°C]`
#>      value           createdAt
#> 1     1.97 2017-12-21 10:46:19
#> 2     2.03 2017-12-21 10:43:52
#> 3     1.95 2017-12-21 10:41:24
#> 4     1.93 2017-12-21 10:38:56
#> 5     1.92 2017-12-21 10:36:28
#> 6     1.99 2017-12-21 10:34:01
#> 7     1.91 2017-12-21 10:31:33
#> 8     1.90 2017-12-21 10:29:05
#> 9     1.86 2017-12-21 10:26:37
#> 10    1.89 2017-12-21 10:24:09
#> 11    1.87 2017-12-21 10:21:42
#> 12    1.81 2017-12-21 10:19:14
#> 13    1.90 2017-12-21 10:16:46
#> 14    1.69 2017-12-21 10:14:18
#> 15    1.64 2017-12-21 10:11:51
#> 16    1.39 2017-12-21 10:09:23
#> 17    1.43 2017-12-21 10:06:55
#> 18    1.32 2017-12-21 10:04:28
#> 19    1.25 2017-12-21 10:02:00
#> 20    1.14 2017-12-21 09:59:31
#> 21    1.40 2017-12-21 09:57:03
#> 22    1.40 2017-12-21 09:54:36
#> 23    1.63 2017-12-21 09:52:08
#> 24    1.51 2017-12-21 09:49:40
#> 25    1.47 2017-12-21 09:47:12
#> 26    1.76 2017-12-21 09:44:45
#> 27    1.66 2017-12-21 09:42:16
#> 28    1.78 2017-12-21 09:39:49
#> 29    1.71 2017-12-21 09:37:21
#> 30    1.73 2017-12-21 09:34:53
#> 31    1.79 2017-12-21 09:32:25
#> 32    2.02 2017-12-21 09:29:57
#> 33    1.81 2017-12-21 09:27:29
#> 34    1.78 2017-12-21 09:25:00
#> 35    1.44 2017-12-21 09:22:33
#> 36    1.68 2017-12-21 09:20:05
#> 37    1.20 2017-12-21 09:17:37
#> 38    0.76 2017-12-21 09:15:09
#> 39    0.32 2017-12-21 09:12:42
#> 40    0.12 2017-12-21 09:10:14
#> 41   -0.05 2017-12-21 09:07:46
#> 42   -0.08 2017-12-21 09:05:18
#> 43   -0.32 2017-12-21 09:02:51
#> 44   -0.42 2017-12-21 09:00:23
#> 45   -0.35 2017-12-21 08:57:55
#> 46   -0.33 2017-12-21 08:55:27
#> 47   -0.57 2017-12-21 08:53:00
#> 48   -0.64 2017-12-21 08:50:32
#> 49   -0.49 2017-12-21 08:48:04
#> 50   -0.56 2017-12-21 08:45:36
#> 51   -0.57 2017-12-21 08:43:09
#> 52   -0.75 2017-12-21 08:40:41
#> 53   -0.54 2017-12-21 08:38:13
#> 54   -0.47 2017-12-21 08:35:45
#> 55   -0.04 2017-12-21 08:33:18
#> 56   -0.11 2017-12-21 08:30:50
#> 57   -0.54 2017-12-21 08:28:22
#> 58   -0.65 2017-12-21 08:25:54
#> 59   -0.83 2017-12-21 08:23:27
#> 60   -0.76 2017-12-21 08:20:59
#> 61   -0.83 2017-12-21 08:18:31
#> 62   -1.47 2017-12-21 08:16:03
#> 63   -1.69 2017-12-21 08:13:36
#> 64   -1.68 2017-12-21 08:11:08
#> 65   -1.59 2017-12-21 08:08:40
#> 66   -1.79 2017-12-21 08:06:13
#> 67   -1.87 2017-12-21 08:03:45
#> 68   -1.94 2017-12-21 08:01:17
#> 69   -2.02 2017-12-21 07:58:49
#> 70   -2.13 2017-12-21 07:56:22
#> 71   -2.21 2017-12-21 07:53:54
#> 72   -2.13 2017-12-21 07:51:26
#> 73   -1.96 2017-12-21 07:48:58
#> 74   -2.23 2017-12-21 07:46:31
#> 75   -2.14 2017-12-21 07:44:03
#> 76   -2.15 2017-12-21 07:41:35
#> 77   -2.12 2017-12-21 07:39:08
#> 78   -2.23 2017-12-21 07:36:40
#> 79   -2.34 2017-12-21 07:34:11
#> 80   -2.33 2017-12-21 07:31:43
#> 81   -2.44 2017-12-21 07:29:15
#> 82   -2.50 2017-12-21 07:26:47
#> 83   -2.54 2017-12-21 07:24:19
#> 84   -2.69 2017-12-21 07:21:52
#> 85   -2.90 2017-12-21 07:19:23
#> 86   -3.03 2017-12-21 07:16:56
#> 87   -3.05 2017-12-21 07:14:28
#> 88   -3.04 2017-12-21 07:12:00
#> 89   -2.92 2017-12-21 07:09:32
#> 90   -2.90 2017-12-21 07:07:05
#> 91   -3.12 2017-12-21 07:04:37
#> 92   -3.33 2017-12-21 07:02:09
#> 93   -3.32 2017-12-21 06:59:41
#> 94   -3.41 2017-12-21 06:57:14
#> 95   -3.57 2017-12-21 06:54:46
#> 96   -3.54 2017-12-21 06:52:18
#> 97   -3.49 2017-12-21 06:49:50
#> 98   -3.49 2017-12-21 06:47:22
#> 99   -3.49 2017-12-21 06:44:54
#> 100  -3.33 2017-12-21 06:42:26
#> 101  -3.38 2017-12-21 06:39:59
#> 102  -3.51 2017-12-21 06:37:31
#> 103  -3.53 2017-12-21 06:35:02
#> 104  -3.52 2017-12-21 06:32:35
#> 105  -3.48 2017-12-21 06:30:07
#> 106  -3.63 2017-12-21 06:27:39
#> 107  -3.73 2017-12-21 06:25:11
#> 108  -3.88 2017-12-21 06:22:44
#> 109  -3.87 2017-12-21 06:20:16
#> 110  -3.79 2017-12-21 06:17:48
#> 111  -3.79 2017-12-21 06:15:20
#> 112  -3.97 2017-12-21 06:12:53
#> 113  -4.02 2017-12-21 06:10:25
#> 114  -3.94 2017-12-21 06:07:57
#> 115  -3.82 2017-12-21 06:05:28
#> 116  -3.72 2017-12-21 06:03:00
#> 117  -3.71 2017-12-21 06:00:31
#> 118  -3.73 2017-12-21 05:58:04
#> 119  -3.81 2017-12-21 05:55:36
#> 120  -3.85 2017-12-21 05:53:08
#> 121  -3.86 2017-12-21 05:50:40
#> 122  -3.88 2017-12-21 05:48:13
#> 123  -4.01 2017-12-21 05:45:44
#> 124  -4.02 2017-12-21 05:43:16
#> 125  -4.07 2017-12-21 05:40:47
#> 126  -4.01 2017-12-21 05:38:16
#> 127  -4.11 2017-12-21 05:35:49
#> 128  -4.12 2017-12-21 05:33:20
#> 129  -4.17 2017-12-21 05:30:51
#> 130  -4.27 2017-12-21 05:28:23
#> 131  -4.30 2017-12-21 05:25:56
#> 132  -4.32 2017-12-21 05:23:28
#> 133  -4.37 2017-12-21 05:21:00
#> 134  -4.30 2017-12-21 05:18:32
#> 135  -4.31 2017-12-21 05:16:05
#> 136  -4.37 2017-12-21 05:13:37
#> 137  -4.38 2017-12-21 05:11:08
#> 138  -4.39 2017-12-21 05:08:40
#> 139  -4.34 2017-12-21 05:06:12
#> 140  -4.33 2017-12-21 05:03:44
#> 141  -4.37 2017-12-21 05:01:15
#> 142  -4.38 2017-12-21 04:58:48
#> 143  -4.34 2017-12-21 04:56:20
#> 144  -4.30 2017-12-21 04:53:43
#> 145  -4.33 2017-12-21 04:51:15
#> 146  -4.26 2017-12-21 04:48:47
#> 147  -4.30 2017-12-21 04:46:18
#> 148  -4.35 2017-12-21 04:43:49
#> 149  -4.48 2017-12-21 04:41:21
#> 150  -4.45 2017-12-21 04:38:53
#> 151  -4.28 2017-12-21 04:36:17
#> 152  -4.42 2017-12-21 04:33:47
#> 153  -4.51 2017-12-21 04:31:10
#> 154  -4.73 2017-12-21 04:28:30
#> 155  -4.93 2017-12-21 04:26:02
#> 156  -5.03 2017-12-21 04:23:34
#> 157  -4.95 2017-12-21 04:21:06
#> 158  -4.96 2017-12-21 04:18:38
#> 159  -4.87 2017-12-21 04:16:09
#> 160  -4.85 2017-12-21 04:13:41
#> 161  -4.71 2017-12-21 04:10:58
#> 162  -4.65 2017-12-21 04:08:28
#> 163  -4.70 2017-12-21 04:05:57
#> 164  -4.66 2017-12-21 04:03:29
#> 165  -4.66 2017-12-21 04:00:59
#> 166  -4.67 2017-12-21 03:58:31
#> 167  -4.71 2017-12-21 03:56:02
#> 168  -4.70 2017-12-21 03:51:44
#> 169  -4.86 2017-12-21 03:44:44
#> 170  -4.89 2017-12-21 03:42:05
#> 171  -4.85 2017-12-21 03:39:29
#> 172  -4.84 2017-12-21 03:36:52
#> 173  -4.88 2017-12-21 03:31:34
#> 174  -4.96 2017-12-21 03:29:01
#> 175  -4.90 2017-12-21 03:25:26
#> 176  -4.97 2017-12-21 03:22:50
#> 177  -4.78 2017-12-21 03:17:42
#> 178  -4.71 2017-12-21 03:15:07
#> 179  -4.53 2017-12-21 03:12:29
#> 180  -4.44 2017-12-21 03:09:56
#> 181  -4.52 2017-12-21 03:07:27
#> 182  -4.51 2017-12-21 03:04:57
#> 183  -4.51 2017-12-21 03:02:25
#> 184  -4.48 2017-12-21 02:59:55
#> 185  -4.34 2017-12-21 02:57:23
#> 186  -4.44 2017-12-21 02:54:55
#> 187  -4.51 2017-12-21 02:52:27
#> 188  -4.39 2017-12-21 02:49:58
#> 189  -4.42 2017-12-21 02:47:29
#> 190  -4.44 2017-12-21 02:45:02
#> 191  -4.47 2017-12-21 02:42:34
#> 192  -4.47 2017-12-21 02:40:05
#> 193  -4.46 2017-12-21 02:37:37
#> 194  -4.25 2017-12-21 02:35:09
#> 195  -4.24 2017-12-21 02:32:41
#> 196  -4.17 2017-12-21 02:30:14
#> 197  -4.27 2017-12-21 02:27:45
#> 198  -4.20 2017-12-21 02:25:17
#> 199  -4.24 2017-12-21 02:22:50
#> 200  -4.23 2017-12-21 02:20:22
#> 201  -4.10 2017-12-21 02:17:54
#> 202  -4.06 2017-12-21 02:15:26
#> 203  -4.02 2017-12-21 02:12:58
#> 204  -4.13 2017-12-21 02:10:31
#> 205  -4.04 2017-12-21 02:08:03
#> 206  -3.95 2017-12-21 02:05:35
#> 207  -4.04 2017-12-21 02:03:07
#> 208  -3.93 2017-12-21 02:00:40
#> 209  -3.84 2017-12-21 01:58:12
#> 210  -3.73 2017-12-21 01:55:44
#> 211  -3.71 2017-12-21 01:53:16
#> 212  -3.65 2017-12-21 01:50:48
#> 213  -3.61 2017-12-21 01:48:20
#> 214  -3.51 2017-12-21 01:45:53
#> 215  -3.52 2017-12-21 01:43:25
#> 216  -3.59 2017-12-21 01:40:57
#> 217  -3.50 2017-12-21 01:38:29
#> 218  -3.50 2017-12-21 01:36:00
#> 219  -3.56 2017-12-21 01:33:33
#> 220  -3.49 2017-12-21 01:31:05
#> 221  -3.48 2017-12-21 01:28:37
#> 222  -3.44 2017-12-21 01:26:09
#> 223  -3.41 2017-12-21 01:23:41
#> 224  -3.34 2017-12-21 01:21:14
#> 225  -3.40 2017-12-21 01:18:46
#> 226  -3.42 2017-12-21 01:16:18
#> 227  -3.40 2017-12-21 01:13:50
#> 228  -3.35 2017-12-21 01:11:23
#> 229  -3.39 2017-12-21 01:08:55
#> 230  -3.29 2017-12-21 01:06:27
#> 231  -3.32 2017-12-21 01:04:00
#> 232  -3.07 2017-12-21 01:01:32
#> 233  -3.04 2017-12-21 00:59:04
#> 234  -3.08 2017-12-21 00:56:36
#> 235  -3.15 2017-12-21 00:54:09
#> 236  -3.26 2017-12-21 00:51:41
#> 237  -3.15 2017-12-21 00:49:13
#> 238  -3.19 2017-12-21 00:46:46
#> 239  -3.15 2017-12-21 00:44:18
#> 240  -3.29 2017-12-21 00:41:50
#> 241  -3.31 2017-12-21 00:39:22
#> 242  -3.25 2017-12-21 00:36:55
#> 243  -3.16 2017-12-21 00:34:27
#> 244  -3.13 2017-12-21 00:31:59
#> 245  -3.04 2017-12-21 00:29:31
#> 246  -3.01 2017-12-21 00:27:04
#> 247  -2.68 2017-12-21 00:24:36
#> 248  -2.56 2017-12-21 00:22:08
#> 249  -2.61 2017-12-21 00:19:40
#> 250  -2.56 2017-12-21 00:17:13
#> 251  -2.65 2017-12-21 00:14:45
#> 252  -2.65 2017-12-21 00:12:17
#> 253  -2.65 2017-12-21 00:09:50
#> 254  -2.71 2017-12-21 00:07:22
#> 255  -2.67 2017-12-21 00:04:54
#> 256  -2.68 2017-12-21 00:02:26
#> 257  -2.71 2017-12-20 23:59:59
#> 258  -2.65 2017-12-20 23:57:31
#> 259  -2.66 2017-12-20 23:55:03
#> 260  -2.70 2017-12-20 23:52:35
#> 261  -2.66 2017-12-20 23:50:08
#> 262  -2.55 2017-12-20 23:47:40
#> 263  -2.34 2017-12-20 23:45:12
#> 264  -2.30 2017-12-20 23:42:44
#> 265  -2.28 2017-12-20 23:40:17
#> 266  -2.30 2017-12-20 23:37:49
#> 267  -2.35 2017-12-20 23:35:21
#> 268  -2.45 2017-12-20 23:32:53
#> 269  -2.41 2017-12-20 23:30:25
#> 270  -2.41 2017-12-20 23:27:57
#> 271  -2.46 2017-12-20 23:25:29
#> 272  -2.55 2017-12-20 23:23:01
#> 273  -2.50 2017-12-20 23:20:33
#> 274  -2.42 2017-12-20 23:18:06
#> 275  -2.37 2017-12-20 23:15:38
#> 276  -2.24 2017-12-20 23:13:10
#> 277  -2.26 2017-12-20 23:10:42
#> 278  -2.04 2017-12-20 23:08:15
#> 279  -2.04 2017-12-20 23:05:47
#> 280  -2.00 2017-12-20 23:03:19
#> 281  -2.05 2017-12-20 23:00:51
#> 282  -2.06 2017-12-20 22:58:23
#> 283  -2.11 2017-12-20 22:55:56
#> 284  -2.12 2017-12-20 22:53:28
#> 285  -2.15 2017-12-20 22:51:00
#> 286  -2.20 2017-12-20 22:48:32
#> 287  -2.16 2017-12-20 22:46:05
#> 288  -2.14 2017-12-20 22:43:37
#> 289  -2.13 2017-12-20 22:41:09
#> 290  -2.18 2017-12-20 22:38:41
#> 291  -2.24 2017-12-20 22:36:13
#> 292  -2.29 2017-12-20 22:33:46
#> 293  -2.33 2017-12-20 22:31:18
#> 294  -2.24 2017-12-20 22:28:50
#> 295  -2.23 2017-12-20 22:26:22
#> 296  -2.23 2017-12-20 22:23:54
#> 297  -2.33 2017-12-20 22:21:27
#> 298  -2.32 2017-12-20 22:18:59
#> 299  -2.34 2017-12-20 22:16:31
#> 300  -2.31 2017-12-20 22:14:03
#> 301  -2.32 2017-12-20 22:11:36
#> 302  -2.30 2017-12-20 22:09:08
#> 303  -2.21 2017-12-20 22:06:40
#> 304  -2.13 2017-12-20 22:04:12
#> 305  -2.03 2017-12-20 22:01:44
#> 306  -1.92 2017-12-20 21:59:15
#> 307  -1.86 2017-12-20 21:56:47
#> 308  -1.77 2017-12-20 21:54:19
#> 309  -1.76 2017-12-20 21:51:51
#> 310  -1.80 2017-12-20 21:49:23
#> 311  -1.70 2017-12-20 21:46:55
#> 312  -1.69 2017-12-20 21:44:27
#> 313  -1.74 2017-12-20 21:41:59
#> 314  -1.67 2017-12-20 21:39:32
#> 315  -1.56 2017-12-20 21:37:04
#> 316  -1.47 2017-12-20 21:34:36
#> 317  -1.52 2017-12-20 21:32:08
#> 318  -1.53 2017-12-20 21:29:40
#> 319  -1.45 2017-12-20 21:27:13
#> 320  -1.34 2017-12-20 21:24:45
#> 321  -1.30 2017-12-20 21:22:17
#> 322  -1.35 2017-12-20 21:19:49
#> 323  -1.25 2017-12-20 21:17:22
#> 324  -1.23 2017-12-20 21:14:53
#> 325  -1.23 2017-12-20 21:12:25
#> 326  -1.24 2017-12-20 21:09:58
#> 327  -1.29 2017-12-20 21:07:30
#> 328  -1.04 2017-12-20 21:05:02
#> 329  -0.97 2017-12-20 21:02:34
#> 330  -0.93 2017-12-20 21:00:07
#> 331  -0.85 2017-12-20 20:57:38
#> 332  -0.90 2017-12-20 20:55:11
#> 333  -0.89 2017-12-20 20:52:42
#> 334  -0.89 2017-12-20 20:50:15
#> 335  -0.86 2017-12-20 20:47:47
#> 336  -0.85 2017-12-20 20:45:19
#> 337  -0.86 2017-12-20 20:42:51
#> 338  -0.85 2017-12-20 20:40:23
#> 339  -0.88 2017-12-20 20:37:56
#> 340  -0.83 2017-12-20 20:35:28
#> 341  -0.80 2017-12-20 20:33:00
#> 342  -0.82 2017-12-20 20:30:32
#> 343  -0.73 2017-12-20 20:28:04
#> 344  -0.61 2017-12-20 20:25:36
#> 345  -0.59 2017-12-20 20:23:08
#> 346  -0.54 2017-12-20 20:20:40
#> 347  -0.61 2017-12-20 20:18:13
#> 348  -0.58 2017-12-20 20:15:44
#> 349  -0.59 2017-12-20 20:13:16
#> 350  -0.62 2017-12-20 20:10:48
#> 351  -0.68 2017-12-20 20:08:21
#> 352  -0.71 2017-12-20 20:05:53
#> 353  -0.62 2017-12-20 20:03:25
#> 354  -0.70 2017-12-20 20:00:57
#> 355  -0.72 2017-12-20 19:58:30
#> 356  -0.70 2017-12-20 19:56:02
#> 357  -0.61 2017-12-20 19:53:34
#> 358  -0.53 2017-12-20 19:51:05
#> 359  -0.48 2017-12-20 19:48:38
#> 360  -0.46 2017-12-20 19:46:10
#> 361  -0.44 2017-12-20 19:43:42
#> 362  -0.47 2017-12-20 19:41:14
#> 363  -0.44 2017-12-20 19:38:46
#> 364  -0.46 2017-12-20 19:36:19
#> 365  -0.51 2017-12-20 19:33:51
#> 366  -0.54 2017-12-20 19:31:23
#> 367  -0.49 2017-12-20 19:28:55
#> 368  -0.41 2017-12-20 19:26:27
#> 369  -0.27 2017-12-20 19:23:59
#> 370  -0.26 2017-12-20 19:21:32
#> 371  -0.30 2017-12-20 19:19:04
#> 372  -0.22 2017-12-20 19:16:34
#> 373  -0.22 2017-12-20 19:14:07
#> 374  -0.42 2017-12-20 19:11:39
#> 375  -0.45 2017-12-20 19:09:10
#> 376  -0.42 2017-12-20 19:06:42
#> 377  -0.37 2017-12-20 19:04:14
#> 378  -0.35 2017-12-20 19:01:46
#> 379  -0.38 2017-12-20 18:59:18
#> 380  -0.39 2017-12-20 18:56:50
#> 381  -0.41 2017-12-20 18:54:22
#> 382  -0.42 2017-12-20 18:51:54
#> 383  -0.44 2017-12-20 18:49:25
#> 384  -0.44 2017-12-20 18:46:58
#> 385  -0.41 2017-12-20 18:44:29
#> 386  -0.34 2017-12-20 18:42:01
#> 387  -0.32 2017-12-20 18:39:32
#> 388  -0.37 2017-12-20 18:37:04
#> 389  -0.29 2017-12-20 18:34:36
#> 390  -0.27 2017-12-20 18:32:08
#> 391  -0.30 2017-12-20 18:29:40
#> 392  -0.33 2017-12-20 18:27:12
#> 393  -0.29 2017-12-20 18:24:44
#> 394  -0.35 2017-12-20 18:22:17
#> 395  -0.32 2017-12-20 18:19:49
#> 396  -0.36 2017-12-20 18:17:21
#> 397  -0.36 2017-12-20 18:14:53
#> 398  -0.32 2017-12-20 18:12:25
#> 399  -0.33 2017-12-20 18:09:57
#> 400  -0.35 2017-12-20 18:07:30
#> 401  -0.41 2017-12-20 18:05:01
#> 402  -0.44 2017-12-20 18:02:33
#> 403  -0.46 2017-12-20 18:00:06
#> 404  -0.49 2017-12-20 17:57:38
#> 405  -0.36 2017-12-20 17:55:10
#> 406  -0.15 2017-12-20 17:52:42
#> 407  -0.20 2017-12-20 17:50:14
#> 408  -0.11 2017-12-20 17:47:45
#> 409  -0.11 2017-12-20 17:45:18
#> 410  -0.15 2017-12-20 17:42:50
#> 411  -0.17 2017-12-20 17:40:21
#> 412  -0.28 2017-12-20 17:37:54
#> 413  -0.22 2017-12-20 17:35:26
#> 414  -0.03 2017-12-20 17:32:58
#> 415  -0.04 2017-12-20 17:30:30
#> 416  -0.06 2017-12-20 17:28:02
#> 417   0.00 2017-12-20 17:25:34
#> 418   0.04 2017-12-20 17:23:07
#> 419   0.06 2017-12-20 17:20:39
#> 420   0.07 2017-12-20 17:18:11
#> 421   0.00 2017-12-20 17:15:43
#> 422   0.04 2017-12-20 17:13:15
#> 423   0.03 2017-12-20 17:10:48
#> 424   0.04 2017-12-20 17:08:20
#> 425   0.00 2017-12-20 17:05:52
#> 426   0.06 2017-12-20 17:03:24
#> 427   0.05 2017-12-20 17:00:56
#> 428   0.09 2017-12-20 16:58:28
#> 429   0.09 2017-12-20 16:56:01
#> 430   0.12 2017-12-20 16:53:33
#> 431   0.07 2017-12-20 16:50:53
#> 432   0.14 2017-12-20 16:48:26
#> 433   0.15 2017-12-20 16:45:58
#> 434   0.26 2017-12-20 16:43:30
#> 435   0.21 2017-12-20 16:41:02
#> 436   0.21 2017-12-20 16:38:34
#> 437   0.26 2017-12-20 16:36:06
#> 438   0.28 2017-12-20 16:33:38
#> 439   0.34 2017-12-20 16:31:10
#> 440   0.40 2017-12-20 16:28:43
#> 441   0.46 2017-12-20 16:26:15
#> 442   0.42 2017-12-20 16:23:47
#> 443   0.42 2017-12-20 16:21:19
#> 444   0.46 2017-12-20 16:18:51
#> 445   0.45 2017-12-20 16:16:24
#> 446   0.46 2017-12-20 16:11:30
#> 447   0.49 2017-12-20 16:09:02
#> 448   0.48 2017-12-20 16:06:34
#> 449   0.53 2017-12-20 16:04:07
#> 450   0.64 2017-12-20 16:01:38
#> 451   0.65 2017-12-20 15:59:10
#> 452   0.64 2017-12-20 15:56:42
#> 453   0.54 2017-12-20 15:54:15
#> 454   0.67 2017-12-20 15:51:47
#> 455   0.86 2017-12-20 15:49:19
#> 456   0.90 2017-12-20 15:46:51
#> 457   0.92 2017-12-20 15:44:24
#> 458   0.92 2017-12-20 15:41:56
#> 459   0.97 2017-12-20 15:39:28
#> 460   1.01 2017-12-20 15:37:00
#> 461   1.02 2017-12-20 15:34:32
#> 462   1.02 2017-12-20 15:32:04
#> 463   1.10 2017-12-20 15:29:37
#> 464   1.12 2017-12-20 15:27:09
#> 465   1.02 2017-12-20 15:24:41
#> 466   1.01 2017-12-20 15:22:14
#> 467   0.97 2017-12-20 15:19:46
#> 468   0.95 2017-12-20 15:17:18
#> 469   1.06 2017-12-20 15:14:50
#> 470   1.05 2017-12-20 15:12:22
#> 471   1.04 2017-12-20 15:09:55
#> 472   1.09 2017-12-20 15:07:27
#> 473   1.16 2017-12-20 15:04:59
#> 474   1.31 2017-12-20 15:02:31
#> 475   1.26 2017-12-20 15:00:03
#> 476   1.27 2017-12-20 14:57:36
#> 477   1.40 2017-12-20 14:55:08
#> 478   1.55 2017-12-20 14:52:40
#> 479   1.43 2017-12-20 14:50:12
#> 480   1.86 2017-12-20 14:47:44
#> 481   1.82 2017-12-20 14:45:17
#> 482   1.82 2017-12-20 14:42:49
#> 483   1.89 2017-12-20 14:40:21
#> 484   1.93 2017-12-20 14:37:53
#> 485   2.26 2017-12-20 14:35:25
#> 486   2.36 2017-12-20 14:32:58
#> 487   2.41 2017-12-20 14:30:30
#> 488   2.46 2017-12-20 14:28:02
#> 489   2.49 2017-12-20 14:25:34
#> 490   2.59 2017-12-20 14:23:07
#> 491   2.54 2017-12-20 14:20:39
#> 492   2.66 2017-12-20 14:18:11
#> 493   2.80 2017-12-20 14:15:43
#> 494   2.87 2017-12-20 14:13:16
#> 495   2.82 2017-12-20 14:10:48
#> 496   2.78 2017-12-20 14:08:20
#> 497   2.85 2017-12-20 14:05:52
#> 498   3.21 2017-12-20 14:03:25
#> 499   3.19 2017-12-20 14:00:57
#> 500   3.18 2017-12-20 13:58:29
#> 501   3.20 2017-12-20 13:56:01
#> 502   3.28 2017-12-20 13:53:34
#> 503   3.29 2017-12-20 13:51:06
#> 504   3.30 2017-12-20 13:48:38
#> 505   3.40 2017-12-20 13:46:10
#> 506   3.37 2017-12-20 13:43:42
#> 507   3.42 2017-12-20 13:41:14
#> 508   3.48 2017-12-20 13:38:47
#> 509   3.51 2017-12-20 13:36:19
#> 510   3.43 2017-12-20 13:33:51
#> 511   3.33 2017-12-20 13:31:23
#> 512   3.33 2017-12-20 13:28:56
#> 513   3.22 2017-12-20 13:26:28
#> 514   3.28 2017-12-20 13:24:00
#> 515   3.40 2017-12-20 13:21:32
#> 516   3.45 2017-12-20 13:19:05
#> 517   3.48 2017-12-20 13:16:37
#> 518   3.25 2017-12-20 13:14:09
#> 519   3.23 2017-12-20 13:11:41
#> 520   3.17 2017-12-20 13:09:13
#> 521   3.18 2017-12-20 13:06:45
#> 522   2.98 2017-12-20 13:04:17
#> 523   2.97 2017-12-20 13:01:49
#> 524   3.20 2017-12-20 12:59:22
#> 525   3.35 2017-12-20 12:56:54
#> 526   3.16 2017-12-20 12:54:26
#> 527   2.92 2017-12-20 12:51:58
#> 528   3.17 2017-12-20 12:49:30
#> 529   3.39 2017-12-20 12:47:03
#> 530   3.76 2017-12-20 12:44:35
#> 531   3.60 2017-12-20 12:42:07
#> 532   3.32 2017-12-20 12:39:39
#> 533   3.30 2017-12-20 12:37:12
#> 534   3.70 2017-12-20 12:34:44
#> 535   3.57 2017-12-20 12:32:16
#> 536   3.54 2017-12-20 12:29:48
#> 537   3.42 2017-12-20 12:27:21
#> 538   3.74 2017-12-20 12:24:53
#> 539   4.17 2017-12-20 12:22:25
#> 540   4.95 2017-12-20 12:19:57
#> 541   5.54 2017-12-20 12:17:30
#> 542   6.73 2017-12-20 12:15:02
#> 543   6.56 2017-12-20 12:12:34
#> 544   6.52 2017-12-20 12:10:06
#> 545   6.04 2017-12-20 12:07:39
#> 546   5.74 2017-12-20 12:05:11
#> 547   5.16 2017-12-20 12:02:43
#> 548   4.55 2017-12-20 12:00:15
#> 549   4.16 2017-12-20 11:57:47
#> 550   4.39 2017-12-20 11:55:20
#> 551   4.19 2017-12-20 11:52:52
#> 552   4.15 2017-12-20 11:50:24
#> 553   4.28 2017-12-20 11:47:56
#> 554   4.43 2017-12-20 11:45:29
#> 555   3.87 2017-12-20 11:43:01
#> 556   3.28 2017-12-20 11:40:33
#> 557   2.97 2017-12-20 11:38:05
#> 558   3.16 2017-12-20 11:35:38
#> 559   2.84 2017-12-20 11:33:10
#> 560   2.47 2017-12-20 11:30:42
#> 561   2.39 2017-12-20 11:28:14
#> 562   2.34 2017-12-20 11:25:46
#> 563   2.56 2017-12-20 11:23:18
#> 564   2.78 2017-12-20 11:20:51
#> 565   2.32 2017-12-20 11:18:23
#> 566   2.33 2017-12-20 11:15:55
#> 567   2.72 2017-12-20 11:13:26
#> 568   2.64 2017-12-20 11:10:59
#> 569   2.38 2017-12-20 11:08:31
#> 570   2.08 2017-12-20 11:06:03
#> 571   2.12 2017-12-20 11:03:35
#> 572   2.07 2017-12-20 11:01:07
#> 573   2.37 2017-12-20 10:58:40
#> 574   1.95 2017-12-20 10:56:12
#> 575   1.67 2017-12-20 10:53:44
#> 576   2.54 2017-12-20 10:51:16
#> 577   2.69 2017-12-20 10:48:49
#> 578   2.56 2017-12-20 10:46:21
#> 579   2.67 2017-12-20 10:43:52
#> 580   2.57 2017-12-20 10:41:24
#> 581   2.38 2017-12-20 10:38:57
#> 582   2.22 2017-12-20 10:36:29
#> 583   2.15 2017-12-20 10:34:01
#> 584   2.10 2017-12-20 10:31:33
#> 585   1.94 2017-12-20 10:29:06
#> 586   2.07 2017-12-20 10:26:38
#> 587   1.83 2017-12-20 10:24:10
#> 588   1.58 2017-12-20 10:21:42
#> 589   1.31 2017-12-20 10:19:14
#> 590   1.31 2017-12-20 10:16:46
#> 591   1.39 2017-12-20 10:14:19
#> 592   1.40 2017-12-20 10:11:51
#> 593   1.59 2017-12-20 10:09:23
#> 594   1.42 2017-12-20 10:06:55
#> 595   1.39 2017-12-20 10:04:28
#> 596   1.17 2017-12-20 10:02:00
#> 597   1.34 2017-12-20 09:59:32
#> 598   1.10 2017-12-20 09:57:04
#> 599   1.24 2017-12-20 09:54:37
#> 600   1.04 2017-12-20 09:52:09
#> 601   1.06 2017-12-20 09:49:41
#> 602   1.10 2017-12-20 09:47:13
#> 603   1.36 2017-12-20 09:44:46
#> 604   1.24 2017-12-20 09:42:18
#> 605   1.27 2017-12-20 09:39:50
#> 606   1.17 2017-12-20 09:37:23
#> 607   1.11 2017-12-20 09:34:55
#> 608   1.06 2017-12-20 09:32:27
#> 609   0.84 2017-12-20 09:29:59
#> 610   0.54 2017-12-20 09:27:31
#> 611   0.50 2017-12-20 09:25:04
#> 612   0.33 2017-12-20 09:22:36
#> 613   0.32 2017-12-20 09:20:08
#> 614   0.15 2017-12-20 09:17:40
#> 615   0.42 2017-12-20 09:15:13
#> 616   0.54 2017-12-20 09:12:45
#> 617   0.45 2017-12-20 09:10:16
#> 618   0.33 2017-12-20 09:07:49
#> 619   0.50 2017-12-20 09:05:21
#> 620   0.38 2017-12-20 09:02:53
#> 621   0.35 2017-12-20 09:00:25
#> 622   0.42 2017-12-20 08:57:57
#> 623   0.28 2017-12-20 08:55:30
#> 624   0.21 2017-12-20 08:53:02
#> 625   0.18 2017-12-20 08:50:34
#> 626   0.05 2017-12-20 08:48:06
#> 627  -0.02 2017-12-20 08:45:39
#> 628   0.00 2017-12-20 08:43:11
#> 629  -0.08 2017-12-20 08:40:43
#> 630  -0.10 2017-12-20 08:38:15
#> 631  -0.18 2017-12-20 08:35:48
#> 632  -0.21 2017-12-20 08:33:20
#> 633  -0.24 2017-12-20 08:30:52
#> 634  -0.25 2017-12-20 08:28:24
#> 635  -0.34 2017-12-20 08:25:56
#> 636  -0.46 2017-12-20 08:23:28
#> 637  -0.53 2017-12-20 08:21:00
#> 638  -0.43 2017-12-20 08:18:32
#> 639  -0.42 2017-12-20 08:16:05
#> 640  -0.47 2017-12-20 08:13:37
#> 641  -0.49 2017-12-20 08:11:09
#> 642  -0.52 2017-12-20 08:08:41
#> 643  -0.49 2017-12-20 08:06:14
#> 644  -0.55 2017-12-20 08:03:46
#> 645  -0.63 2017-12-20 08:01:18
#> 646  -0.73 2017-12-20 07:58:50
#> 647  -0.78 2017-12-20 07:56:23
#> 648  -0.80 2017-12-20 07:53:55
#> 649  -0.74 2017-12-20 07:51:27
#> 650  -0.76 2017-12-20 07:48:59
#> 651  -0.71 2017-12-20 07:46:32
#> 652  -0.80 2017-12-20 07:44:04
#> 653  -0.77 2017-12-20 07:41:36
#> 654  -0.75 2017-12-20 07:39:08
#> 655  -0.75 2017-12-20 07:36:40
#> 656  -0.81 2017-12-20 07:34:12
#> 657  -0.85 2017-12-20 07:31:44
#> 658  -0.90 2017-12-20 07:29:16
#> 659  -0.92 2017-12-20 07:26:49
#> 660  -0.96 2017-12-20 07:24:21
#> 661  -0.96 2017-12-20 07:21:53
#> 662  -1.08 2017-12-20 07:19:25
#> 663  -1.06 2017-12-20 07:16:58
#> 664  -1.00 2017-12-20 07:14:30
#> 665  -1.01 2017-12-20 07:12:02
#> 666  -1.02 2017-12-20 07:09:34
#> 667  -1.03 2017-12-20 07:07:07
#> 668  -1.11 2017-12-20 07:04:39
#> 669  -1.13 2017-12-20 07:02:11
#> 670  -1.17 2017-12-20 06:59:43
#> 671  -1.20 2017-12-20 06:57:16
#> 672  -1.12 2017-12-20 06:54:48
#> 673  -1.10 2017-12-20 06:52:20
#> 674  -1.10 2017-12-20 06:49:52
#> 675  -1.13 2017-12-20 06:47:25
#> 676  -1.14 2017-12-20 06:44:57
#> 677  -1.15 2017-12-20 06:42:29
#> 678  -1.15 2017-12-20 06:40:01
#> 679  -1.18 2017-12-20 06:37:34
#> 680  -1.18 2017-12-20 06:35:06
#> 681  -1.15 2017-12-20 06:32:38
#> 682  -1.14 2017-12-20 06:30:11
#> 683  -1.11 2017-12-20 06:27:43
#> 684  -1.11 2017-12-20 06:25:15
#> 685  -1.11 2017-12-20 06:22:47
#> 686  -1.14 2017-12-20 06:20:20
#> 687  -1.11 2017-12-20 06:17:52
#> 688  -1.10 2017-12-20 06:15:24
#> 689  -1.13 2017-12-20 06:12:56
#> 690  -1.17 2017-12-20 06:10:29
#> 691  -1.17 2017-12-20 06:08:01
#> 692  -1.20 2017-12-20 06:05:33
#> 693  -1.26 2017-12-20 06:03:06
#> 694  -1.27 2017-12-20 06:00:38
#> 695  -1.28 2017-12-20 05:58:10
#> 696  -1.29 2017-12-20 05:55:42
#> 697  -1.22 2017-12-20 05:53:15
#> 698  -1.16 2017-12-20 05:50:47
#> 699  -1.26 2017-12-20 05:48:19
#> 700  -1.29 2017-12-20 05:45:51
#> 701  -1.34 2017-12-20 05:43:24
#> 702  -1.35 2017-12-20 05:40:56
#> 703  -1.41 2017-12-20 05:38:28
#> 704  -1.43 2017-12-20 05:36:00
#> 705  -1.43 2017-12-20 05:33:32
#> 706  -1.40 2017-12-20 05:31:05
#> 707  -1.44 2017-12-20 05:28:37
#> 708  -1.54 2017-12-20 05:26:09
#> 709  -1.56 2017-12-20 05:23:41
#> 710  -1.52 2017-12-20 05:21:13
#> 711  -1.50 2017-12-20 05:18:46
#> 712  -1.54 2017-12-20 05:16:18
#> 713  -1.55 2017-12-20 05:13:50
#> 714  -1.55 2017-12-20 05:11:22
#> 715  -1.58 2017-12-20 05:08:55
#> 716  -1.54 2017-12-20 05:06:27
#> 717  -1.56 2017-12-20 05:03:59
#> 718  -1.60 2017-12-20 05:01:31
#> 719  -1.49 2017-12-20 04:59:04
#> 720  -1.46 2017-12-20 04:56:36
#> 721  -1.43 2017-12-20 04:54:08
#> 722  -1.41 2017-12-20 04:51:41
#> 723  -1.43 2017-12-20 04:49:13
#> 724  -1.47 2017-12-20 04:46:45
#> 725  -1.47 2017-12-20 04:44:17
#> 726  -1.52 2017-12-20 04:41:50
#> 727  -1.49 2017-12-20 04:39:22
#> 728  -1.55 2017-12-20 04:36:54
#> 729  -1.54 2017-12-20 04:34:27
#> 730  -1.61 2017-12-20 04:31:59
#> 731  -1.65 2017-12-20 04:29:31
#> 732  -1.66 2017-12-20 04:27:03
#> 733  -1.51 2017-12-20 04:24:36
#> 734  -1.49 2017-12-20 04:22:08
#> 735  -1.53 2017-12-20 04:19:40
#> 736  -1.53 2017-12-20 04:17:12
#> 737  -1.52 2017-12-20 04:14:45
#> 738  -1.55 2017-12-20 04:12:17
#> 739  -1.44 2017-12-20 04:09:49
#> 740  -1.48 2017-12-20 04:07:21
#> 741  -1.53 2017-12-20 04:04:54
#> 742  -1.58 2017-12-20 04:02:26
#> 743  -1.63 2017-12-20 03:59:58
#> 744  -1.58 2017-12-20 03:57:30
#> 745  -1.59 2017-12-20 03:55:03
#> 746  -1.54 2017-12-20 03:52:35
#> 747  -1.57 2017-12-20 03:50:07
#> 748  -1.54 2017-12-20 03:47:39
#> 749  -1.61 2017-12-20 03:45:11
#> 750  -1.59 2017-12-20 03:42:44
#> 751  -1.60 2017-12-20 03:40:16
#> 752  -1.59 2017-12-20 03:37:48
#> 753  -1.69 2017-12-20 03:35:21
#> 754  -1.63 2017-12-20 03:32:53
#> 755  -1.54 2017-12-20 03:30:25
#> 756  -1.57 2017-12-20 03:27:57
#> 757  -1.46 2017-12-20 03:25:30
#> 758  -1.43 2017-12-20 03:23:02
#> 759  -1.44 2017-12-20 03:20:34
#> 760  -1.43 2017-12-20 03:18:06
#> 761  -1.43 2017-12-20 03:15:39
#> 762  -1.44 2017-12-20 03:13:11
#> 763  -1.43 2017-12-20 03:10:43
#> 764  -1.52 2017-12-20 03:08:15
#> 765  -1.55 2017-12-20 03:05:48
#> 766  -1.59 2017-12-20 03:03:20
#> 767  -1.54 2017-12-20 03:00:52
#> 768  -1.46 2017-12-20 02:58:24
#> 769  -1.48 2017-12-20 02:55:56
#> 770  -1.50 2017-12-20 02:53:29
#> 771  -1.48 2017-12-20 02:51:01
#> 772  -1.55 2017-12-20 02:48:33
#> 773  -1.52 2017-12-20 02:46:05
#> 774  -1.49 2017-12-20 02:43:37
#> 775  -1.53 2017-12-20 02:41:08
#> 776  -1.56 2017-12-20 02:38:37
#> 777  -1.55 2017-12-20 02:36:07
#> 778  -1.57 2017-12-20 02:33:38
#> 779  -1.54 2017-12-20 02:30:58
#> 780  -1.57 2017-12-20 02:28:31
#> 781  -1.65 2017-12-20 02:26:03
#> 782  -1.62 2017-12-20 02:23:35
#> 783  -1.56 2017-12-20 02:21:07
#> 784  -1.56 2017-12-20 02:18:39
#> 785  -1.56 2017-12-20 02:16:12
#> 786  -1.57 2017-12-20 02:13:44
#> 787  -1.61 2017-12-20 02:11:16
#> 788  -1.63 2017-12-20 02:08:48
#> 789  -1.60 2017-12-20 02:06:21
#> 790  -1.60 2017-12-20 02:03:52
#> 791  -1.54 2017-12-20 02:01:24
#> 792  -1.60 2017-12-20 01:58:57
#> 793  -1.60 2017-12-20 01:56:29
#> 794  -1.60 2017-12-20 01:54:01
#> 795  -1.62 2017-12-20 01:51:33
#> 796  -1.56 2017-12-20 01:49:05
#> 797  -1.62 2017-12-20 01:46:38
#> 798  -1.61 2017-12-20 01:44:10
#> 799  -1.68 2017-12-20 01:41:42
#> 800  -1.70 2017-12-20 01:39:15
#> 801  -1.69 2017-12-20 01:36:47
#> 802  -1.68 2017-12-20 01:34:19
#> 803  -1.65 2017-12-20 01:31:51
#> 804  -1.63 2017-12-20 01:29:24
#> 805  -1.68 2017-12-20 01:26:56
#> 806  -1.68 2017-12-20 01:24:28
#> 807  -1.63 2017-12-20 01:22:00
#> 808  -1.65 2017-12-20 01:19:33
#> 809  -1.70 2017-12-20 01:17:05
#> 810  -1.75 2017-12-20 01:14:37
#> 811  -1.74 2017-12-20 01:12:10
#> 812  -1.75 2017-12-20 01:09:42
#> 813  -1.67 2017-12-20 01:07:14
#> 814  -1.65 2017-12-20 01:04:46
#> 815  -1.65 2017-12-20 01:02:18
#> 816  -1.70 2017-12-20 00:59:51
#> 817  -1.69 2017-12-20 00:57:23
#> 818  -1.66 2017-12-20 00:54:55
#> 819  -1.64 2017-12-20 00:52:27
#> 820  -1.64 2017-12-20 00:50:00
#> 821  -1.66 2017-12-20 00:47:32
#> 822  -1.63 2017-12-20 00:45:04
#> 823  -1.64 2017-12-20 00:42:36
#> 824  -1.67 2017-12-20 00:40:09
#> 825  -1.67 2017-12-20 00:37:41
#> 826  -1.68 2017-12-20 00:35:13
#> 827  -1.70 2017-12-20 00:32:45
#> 828  -1.77 2017-12-20 00:30:17
#> 829  -1.80 2017-12-20 00:27:50
#> 830  -1.88 2017-12-20 00:25:22
#> 831  -1.94 2017-12-20 00:22:54
#> 832  -1.86 2017-12-20 00:20:26
#> 833  -1.87 2017-12-20 00:17:59
#> 834  -1.89 2017-12-20 00:15:31
#> 835  -1.90 2017-12-20 00:13:03
#> 836  -1.91 2017-12-20 00:10:36
#> 837  -1.94 2017-12-20 00:08:08
#> 838  -1.86 2017-12-20 00:05:40
#> 839  -1.89 2017-12-20 00:03:12
#> 840  -1.88 2017-12-20 00:00:45
#> 841  -1.85 2017-12-19 23:58:17
#> 842  -1.84 2017-12-19 23:55:49
#> 843  -1.84 2017-12-19 23:53:21
#> 844  -1.87 2017-12-19 23:50:54
#> 845  -1.86 2017-12-19 23:48:26
#> 846  -1.90 2017-12-19 23:45:58
#> 847  -1.89 2017-12-19 23:43:31
#> 848  -1.88 2017-12-19 23:41:03
#> 849  -1.81 2017-12-19 23:38:35
#> 850  -1.75 2017-12-19 23:36:07
#> 851  -1.56 2017-12-19 23:33:40
#> 852  -1.50 2017-12-19 23:31:12
#> 853  -1.72 2017-12-19 23:28:44
#> 854  -1.83 2017-12-19 23:26:16
#> 855  -1.82 2017-12-19 23:23:49
#> 856  -1.82 2017-12-19 23:21:20
#> 857  -1.89 2017-12-19 23:18:52
#> 858  -1.97 2017-12-19 23:16:25
#> 859  -1.96 2017-12-19 23:13:57
#> 860  -1.84 2017-12-19 23:11:29
#> 861  -2.04 2017-12-19 23:09:00
#> 862  -2.06 2017-12-19 23:06:33
#> 863  -2.07 2017-12-19 23:04:05
#> 864  -2.14 2017-12-19 23:01:37
#> 865  -2.21 2017-12-19 22:59:09
#> 866  -2.28 2017-12-19 22:56:42
#> 867  -2.29 2017-12-19 22:54:14
#> 868  -2.28 2017-12-19 22:51:46
#> 869  -2.32 2017-12-19 22:49:18
#> 870  -2.35 2017-12-19 22:46:50
#> 871  -2.39 2017-12-19 22:44:23
#> 872  -2.40 2017-12-19 22:41:55
#> 873  -2.45 2017-12-19 22:39:27
#> 874  -2.50 2017-12-19 22:36:59
#> 875  -2.54 2017-12-19 22:34:32
#> 876  -2.46 2017-12-19 22:32:04
#> 877  -2.49 2017-12-19 22:29:36
#> 878  -2.44 2017-12-19 22:27:08
#> 879  -2.46 2017-12-19 22:24:41
#> 880  -2.50 2017-12-19 22:22:12
#> 881  -2.49 2017-12-19 22:19:44
#> 882  -2.50 2017-12-19 22:17:16
#> 883  -2.53 2017-12-19 22:14:49
#> 884  -2.61 2017-12-19 22:12:21
#> 885  -2.59 2017-12-19 22:09:53
#> 886  -2.59 2017-12-19 22:07:25
#> 887  -2.65 2017-12-19 22:04:58
#> 888  -2.60 2017-12-19 22:02:30
#> 889  -2.59 2017-12-19 22:00:02
#> 890  -2.64 2017-12-19 21:57:34
#> 891  -2.60 2017-12-19 21:55:06
#> 892  -2.68 2017-12-19 21:52:37
#> 893  -2.64 2017-12-19 21:50:09
#> 894  -2.76 2017-12-19 21:47:41
#> 895  -2.72 2017-12-19 21:45:14
#> 896  -2.78 2017-12-19 21:42:46
#> 897  -2.64 2017-12-19 21:40:18
#> 898  -2.59 2017-12-19 21:37:50
#> 899  -2.67 2017-12-19 21:35:22
#> 900  -2.59 2017-12-19 21:32:55
#> 901  -2.69 2017-12-19 21:30:27
#> 902  -2.77 2017-12-19 21:27:59
#> 903  -2.78 2017-12-19 21:25:31
#> 904  -2.81 2017-12-19 21:23:03
#> 905  -2.84 2017-12-19 21:20:34
#> 906  -2.89 2017-12-19 21:18:06
#> 907  -2.97 2017-12-19 21:15:37
#> 908  -2.95 2017-12-19 21:13:09
#> 909  -2.82 2017-12-19 21:10:41
#> 910  -2.80 2017-12-19 21:08:11
#> 911  -2.72 2017-12-19 21:05:42
#> 912  -2.71 2017-12-19 21:03:13
#> 913  -2.77 2017-12-19 21:00:44
#> 914  -2.87 2017-12-19 20:58:15
#> 915  -2.99 2017-12-19 20:55:45
#> 916  -3.04 2017-12-19 20:53:14
#> 917  -3.07 2017-12-19 20:50:44
#> 918  -3.12 2017-12-19 20:48:15
#> 919  -3.09 2017-12-19 20:45:47
#> 920  -3.21 2017-12-19 20:43:18
#> 921  -3.18 2017-12-19 20:40:49
#> 922  -3.16 2017-12-19 20:38:20
#> 923  -3.09 2017-12-19 20:35:50
#> 924  -3.08 2017-12-19 20:33:20
#> 925  -3.11 2017-12-19 20:30:51
#> 926  -3.15 2017-12-19 20:28:21
#> 927  -3.09 2017-12-19 20:25:46
#> 928  -3.05 2017-12-19 20:23:18
#> 929  -3.12 2017-12-19 20:20:48
#> 930  -3.03 2017-12-19 20:18:19
#> 931  -3.06 2017-12-19 20:15:51
#> 932  -3.12 2017-12-19 20:13:23
#> 933  -3.10 2017-12-19 20:10:56
#> 934  -3.15 2017-12-19 20:08:28
#> 935  -3.20 2017-12-19 20:06:00
#> 936  -3.26 2017-12-19 20:03:32
#> 937  -3.26 2017-12-19 20:01:04
#> 938  -3.26 2017-12-19 19:58:36
#> 939  -3.34 2017-12-19 19:56:08
#> 940  -3.19 2017-12-19 19:53:40
#> 941  -3.00 2017-12-19 19:51:13
#> 942  -2.72 2017-12-19 19:48:45
#> 943  -3.08 2017-12-19 19:46:17
#> 944  -3.11 2017-12-19 19:43:49
#> 945  -3.11 2017-12-19 19:41:21
#> 946  -3.15 2017-12-19 19:38:52
#> 947  -3.22 2017-12-19 19:36:25
#> 948  -3.17 2017-12-19 19:33:56
#> 949  -3.05 2017-12-19 19:31:27
#> 950  -3.10 2017-12-19 19:28:58
#> 951  -3.14 2017-12-19 19:26:29
#> 952  -3.18 2017-12-19 19:24:01
#> 953  -3.26 2017-12-19 19:21:33
#> 954  -3.32 2017-12-19 19:19:05
#> 955  -3.39 2017-12-19 19:16:38
#> 956  -3.41 2017-12-19 19:14:10
#> 957  -3.38 2017-12-19 19:11:42
#> 958  -3.42 2017-12-19 19:09:14
#> 959  -3.44 2017-12-19 19:06:44
#> 960  -3.47 2017-12-19 19:04:17
#> 961  -3.45 2017-12-19 19:01:49
#> 962  -3.44 2017-12-19 18:59:21
#> 963  -3.44 2017-12-19 18:56:53
#> 964  -3.47 2017-12-19 18:54:24
#> 965  -3.50 2017-12-19 18:51:56
#> 966  -3.62 2017-12-19 18:49:28
#> 967  -3.59 2017-12-19 18:47:00
#> 968  -3.60 2017-12-19 18:44:31
#> 969  -3.57 2017-12-19 18:42:03
#> 970  -3.47 2017-12-19 18:39:36
#> 971  -3.52 2017-12-19 18:37:07
#> 972  -3.55 2017-12-19 18:34:39
#> 973  -3.60 2017-12-19 18:32:10
#> 974  -3.63 2017-12-19 18:29:43
#> 975  -3.64 2017-12-19 18:27:15
#> 976  -3.73 2017-12-19 18:24:47
#> 977  -3.74 2017-12-19 18:22:19
#> 978  -3.82 2017-12-19 18:19:50
#> 979  -3.85 2017-12-19 18:17:23
#> 980  -3.90 2017-12-19 18:14:55
#> 981  -3.92 2017-12-19 18:12:27
#> 982  -4.00 2017-12-19 18:09:59
#> 983  -4.14 2017-12-19 18:07:31
#> 984  -3.90 2017-12-19 18:05:04
#> 985  -3.94 2017-12-19 18:02:36
#> 986  -4.03 2017-12-19 18:00:08
#> 987  -4.01 2017-12-19 17:57:40
#> 988  -4.03 2017-12-19 17:55:12
#> 989  -4.00 2017-12-19 17:52:45
#> 990  -3.94 2017-12-19 17:50:17
#> 991  -3.85 2017-12-19 17:47:49
#> 992  -3.90 2017-12-19 17:45:21
#> 993  -3.96 2017-12-19 17:42:54
#> 994  -3.91 2017-12-19 17:40:26
#> 995  -3.77 2017-12-19 17:37:58
#> 996  -3.72 2017-12-19 17:35:30
#> 997  -3.66 2017-12-19 17:33:02
#> 998  -3.50 2017-12-19 17:30:34
#> 999  -3.71 2017-12-19 17:28:06
#> 1000 -3.78 2017-12-19 17:25:38
#> 1001 -3.86 2017-12-19 17:23:10
#> 1002 -3.89 2017-12-19 17:20:43
#> 1003 -3.91 2017-12-19 17:18:15
#> 1004 -3.89 2017-12-19 17:15:47
#> 1005 -3.93 2017-12-19 17:13:19
#> 1006 -3.91 2017-12-19 17:10:51
#> 1007 -3.87 2017-12-19 17:08:24
#> 1008 -3.99 2017-12-19 17:05:56
#> 1009 -4.03 2017-12-19 17:03:28
#> 1010 -3.99 2017-12-19 17:01:00
#> 1011 -4.10 2017-12-19 16:58:33
#> 1012 -4.02 2017-12-19 16:56:05
#> 1013 -3.97 2017-12-19 16:53:37
#> 1014 -4.01 2017-12-19 16:51:09
#> 1015 -3.97 2017-12-19 16:48:41
#> 1016 -4.04 2017-12-19 16:46:14
#> 1017 -4.06 2017-12-19 16:43:46
#> 1018 -4.12 2017-12-19 16:41:18
#> 1019 -4.08 2017-12-19 16:38:50
#> 1020 -4.13 2017-12-19 16:36:22
#> 1021 -4.06 2017-12-19 16:33:54
#> 1022 -4.14 2017-12-19 16:31:26
#> 1023 -4.10 2017-12-19 16:28:58
#> 1024 -4.11 2017-12-19 16:26:31
#> 1025 -4.00 2017-12-19 16:24:03
#> 1026 -3.94 2017-12-19 16:21:35
#> 1027 -3.85 2017-12-19 16:19:04
#> 1028 -3.87 2017-12-19 16:16:36
#> 1029 -3.93 2017-12-19 16:14:08
#> 1030 -3.93 2017-12-19 16:11:40
#> 1031 -3.96 2017-12-19 16:09:13
#> 1032 -3.96 2017-12-19 16:06:45
#> 1033 -4.04 2017-12-19 16:04:17
#> 1034 -3.92 2017-12-19 16:01:49
#> 1035 -4.06 2017-12-19 15:59:22
#> 1036 -3.93 2017-12-19 15:56:54
#> 1037 -3.90 2017-12-19 15:54:26
#> 1038 -3.99 2017-12-19 15:51:58
#> 1039 -3.99 2017-12-19 15:49:30
#> 1040 -4.09 2017-12-19 15:47:02
#> 1041 -4.04 2017-12-19 15:44:35
#> 1042 -4.11 2017-12-19 15:42:07
#> 1043 -4.05 2017-12-19 15:39:38
#> 1044 -4.04 2017-12-19 15:37:09
#> 1045 -3.91 2017-12-19 15:34:42
#> 1046 -3.97 2017-12-19 15:32:14
#> 1047 -3.99 2017-12-19 15:29:46
#> 1048 -4.00 2017-12-19 15:27:18
#> 1049 -3.85 2017-12-19 15:24:50
#> 1050 -3.92 2017-12-19 15:22:22
#> 1051 -3.96 2017-12-19 15:19:54
#> 1052 -3.97 2017-12-19 15:17:26
#> 1053 -3.92 2017-12-19 15:14:59
#> 1054 -3.87 2017-12-19 15:12:31
#> 1055 -3.84 2017-12-19 15:10:03
#> 1056 -3.71 2017-12-19 15:07:35
#> 1057 -3.72 2017-12-19 15:05:07
#> 1058 -3.89 2017-12-19 15:02:39
#> 1059 -3.60 2017-12-19 15:00:11
#> 1060 -3.56 2017-12-19 14:57:42
#> 1061 -3.57 2017-12-19 14:55:13
#> 1062 -3.54 2017-12-19 14:52:45
#> 1063 -3.62 2017-12-19 14:50:18
#> 1064 -3.66 2017-12-19 14:47:50
#> 1065 -3.71 2017-12-19 14:45:22
#> 1066 -3.46 2017-12-19 14:42:54
#> 1067 -3.49 2017-12-19 14:40:26
#> 1068 -3.61 2017-12-19 14:37:58
#> 1069 -3.67 2017-12-19 14:35:30
#> 1070 -3.74 2017-12-19 14:33:03
#> 1071 -3.71 2017-12-19 14:30:35
#> 1072 -3.56 2017-12-19 14:28:07
#> 1073 -3.37 2017-12-19 14:25:39
#> 1074 -3.12 2017-12-19 14:23:11
#> 1075 -3.22 2017-12-19 14:20:43
#> 1076 -3.15 2017-12-19 14:18:15
#> 1077 -3.22 2017-12-19 14:15:47
#> 1078 -3.18 2017-12-19 14:13:19
#> 1079 -3.13 2017-12-19 14:10:51
#> 1080 -2.73 2017-12-19 14:08:23
#> 1081 -2.78 2017-12-19 14:05:56
#> 1082 -2.82 2017-12-19 14:03:28
#> 1083 -2.71 2017-12-19 14:01:00
#> 1084 -2.60 2017-12-19 13:58:32
#> 1085 -2.66 2017-12-19 13:56:04
#> 1086 -2.80 2017-12-19 13:53:37
#> 1087 -2.86 2017-12-19 13:51:08
#> 1088 -2.54 2017-12-19 13:48:40
#> 1089 -2.50 2017-12-19 13:46:12
#> 1090 -2.59 2017-12-19 13:43:44
#> 1091 -2.60 2017-12-19 13:41:17
#> 1092 -2.76 2017-12-19 13:38:49
#> 1093 -2.64 2017-12-19 13:36:21
#> 1094 -2.88 2017-12-19 13:33:53
#> 1095 -2.66 2017-12-19 13:31:25
#> 1096 -2.66 2017-12-19 13:28:58
#> 1097 -2.63 2017-12-19 13:26:30
#> 1098 -2.67 2017-12-19 13:24:01
#> 1099 -2.75 2017-12-19 13:21:33
#> 1100 -2.75 2017-12-19 13:19:05
#> 1101 -2.76 2017-12-19 13:16:37
#> 1102 -3.00 2017-12-19 13:14:09
#> 1103 -2.81 2017-12-19 13:11:41
#> 1104 -2.90 2017-12-19 13:09:13
#> 1105 -3.25 2017-12-19 13:06:46
#> 1106 -2.99 2017-12-19 13:04:18
#> 1107 -2.73 2017-12-19 13:01:50
#> 1108 -2.66 2017-12-19 12:59:22
#> 1109 -2.77 2017-12-19 12:56:54
#> 1110 -3.12 2017-12-19 12:54:26
#> 1111 -3.20 2017-12-19 12:51:59
#> 1112 -3.30 2017-12-19 12:49:30
#> 1113 -3.29 2017-12-19 12:47:02
#> 1114 -3.11 2017-12-19 12:44:35
#> 1115 -3.14 2017-12-19 12:42:07
#> 1116 -3.25 2017-12-19 12:39:39
#> 1117 -3.18 2017-12-19 12:37:11
#> 1118 -2.90 2017-12-19 12:34:43
#> 1119 -2.94 2017-12-19 12:32:15
#> 1120 -2.92 2017-12-19 12:29:48
#> 1121 -3.19 2017-12-19 12:27:20
#> 1122 -3.31 2017-12-19 12:24:52
#> 1123 -2.92 2017-12-19 12:22:24
#> 1124 -3.03 2017-12-19 12:19:56
#> 1125 -2.81 2017-12-19 12:17:28
#> 1126 -3.16 2017-12-19 12:15:01
#> 1127 -3.41 2017-12-19 12:12:33
#> 1128 -3.20 2017-12-19 12:10:05
#> 1129 -3.56 2017-12-19 12:07:37
#> 1130 -3.21 2017-12-19 12:05:10
#> 1131 -3.05 2017-12-19 12:02:42
#> 1132 -3.37 2017-12-19 12:00:14
#> 1133 -3.49 2017-12-19 11:57:46
#> 1134 -3.22 2017-12-19 11:55:18
#> 1135 -3.43 2017-12-19 11:52:50
#> 1136 -3.68 2017-12-19 11:50:23
#> 1137 -3.77 2017-12-19 11:47:55
#> 1138 -3.06 2017-12-19 11:45:27
#> 1139 -2.84 2017-12-19 11:42:59
#> 1140 -3.06 2017-12-19 11:40:31
#> 1141 -3.28 2017-12-19 11:38:03
#> 1142 -3.66 2017-12-19 11:35:34
#> 1143 -3.92 2017-12-19 11:33:06
#> 1144 -3.14 2017-12-19 11:30:37
#> 1145 -3.13 2017-12-19 11:28:09
#> 1146 -2.86 2017-12-19 11:25:41
#> 1147 -2.52 2017-12-19 11:23:10
#> 1148 -2.47 2017-12-19 11:20:41
#> 1149 -2.47 2017-12-19 11:18:10
#> 1150 -2.86 2017-12-19 11:15:40
#> 1151 -2.54 2017-12-19 11:13:12
#> 1152 -2.92 2017-12-19 11:10:36
#> 1153 -4.07 2017-12-19 11:03:25
#> 1154 -3.84 2017-12-19 11:00:50
#> 1155 -4.02 2017-12-19 10:50:17
#> 1156 -4.55 2017-12-19 10:47:44
#> 
#> 

# NOT RUN {
get_senseBox_data(
  senseBoxId = c("5957b67494f0520011304cc1","5957b67494f0520011304cc1") ,
  sensorId = "all",
  fromDate = "2017-11-25-12:00:00",
  toDate = "2017-11-26-12:00:00")

get_senseBox_data(
  senseBoxId = "5957b67494f0520011304cc1",
  sensorId = "5957b67494f0520011304cc4",
  fromDate = "2017-11-25-12:00:00",
  toDate = "2017-11-26-12:00:00")
# }