Get data from a senseBox

If the arguments fromDate and toDate are NOT given, the measurements from the last 48 h from now are downloaded. The maximum numbers of downloaded measurements are 10,000. The maximum time frame is one month. It is possible to download the data directly as a csv file, when argument CSV is TRUE.

get_senseBox_data(senseBoxId, sensorId = rep("all", times =
  length(unlist(senseBoxId))), fromDate = NULL, toDate = NULL,
  parallel = FALSE, CSV = FALSE, POSIXct = TRUE)

Arguments

senseBoxId
characteror vector of character (required): senseBoxId or a vector of character of senseBoxIds

sensorId
character or list of character (required): sensorId or a list of senseBoxIds

fromDate
character (optional): Just show data with from date in format YYYY-mm-dd HH:MM:SS

toDate
character (optional): Just show data with up to date in format YYYY-mm-dd HH:MM:SS

parallel
logical (optional): Should the calculations be executed on multiple cores? At least 4 cores are necessary to use this feature.

CSV
logical (optional): Download data as csv file? NOT SUPPORTED UNTIL NOW!

POSIXct
logical (optional): Should the timestamp be translated into POSIXct?

Value

: A list with every entry is a sensBoxId. Every list entry inherits a data.frame with values and dates.

Function version

0.0.1

Examples

##=====================================
## Example: get data from sensorId
##=====================================

get_senseBox_data(
  senseBoxId = "5957b67494f0520011304cc1",
  sensorId = "5957b67494f0520011304cc4")

#> $`5957b67494f0520011304cc1`
#> $`5957b67494f0520011304cc1`$`Temperatur [°C]`
#>      value           createdAt
#> 1     2.23 2017-11-29 09:10:26
#> 2     2.41 2017-11-29 09:07:59
#> 3     2.06 2017-11-29 09:05:31
#> 4     2.17 2017-11-29 09:03:03
#> 5     2.46 2017-11-29 09:00:35
#> 6     1.87 2017-11-29 08:58:07
#> 7     1.85 2017-11-29 08:55:40
#> 8     1.62 2017-11-29 08:53:12
#> 9     1.82 2017-11-29 08:50:44
#> 10    1.94 2017-11-29 08:48:16
#> 11    1.93 2017-11-29 08:45:48
#> 12    1.62 2017-11-29 08:43:21
#> 13    1.17 2017-11-29 08:40:53
#> 14    0.98 2017-11-29 08:38:25
#> 15    1.03 2017-11-29 08:35:57
#> 16    0.91 2017-11-29 08:33:30
#> 17    0.96 2017-11-29 08:31:02
#> 18    0.91 2017-11-29 08:28:34
#> 19    0.87 2017-11-29 08:26:06
#> 20    0.85 2017-11-29 08:23:38
#> 21    0.81 2017-11-29 08:21:10
#> 22    0.84 2017-11-29 08:18:43
#> 23    0.80 2017-11-29 08:16:15
#> 24    0.90 2017-11-29 08:13:47
#> 25    0.93 2017-11-29 08:11:19
#> 26    0.82 2017-11-29 08:08:51
#> 27    0.71 2017-11-29 08:06:24
#> 28    0.57 2017-11-29 08:03:56
#> 29    0.51 2017-11-29 08:01:28
#> 30    0.47 2017-11-29 07:59:00
#> 31    0.43 2017-11-29 07:56:33
#> 32    0.46 2017-11-29 07:54:05
#> 33    0.45 2017-11-29 07:51:37
#> 34    0.25 2017-11-29 07:49:09
#> 35    0.19 2017-11-29 07:46:41
#> 36    0.10 2017-11-29 07:44:14
#> 37   -0.02 2017-11-29 07:41:46
#> 38   -0.05 2017-11-29 07:39:18
#> 39   -0.04 2017-11-29 07:36:50
#> 40   -0.06 2017-11-29 07:34:23
#> 41   -0.01 2017-11-29 07:31:55
#> 42    0.23 2017-11-29 07:29:27
#> 43    0.29 2017-11-29 07:26:58
#> 44    0.24 2017-11-29 07:24:30
#> 45    0.17 2017-11-29 07:22:03
#> 46    0.23 2017-11-29 07:19:35
#> 47    0.35 2017-11-29 07:17:07
#> 48    0.28 2017-11-29 07:14:39
#> 49    0.24 2017-11-29 07:12:12
#> 50    0.20 2017-11-29 07:09:44
#> 51    0.12 2017-11-29 07:07:16
#> 52    0.03 2017-11-29 07:04:48
#> 53   -0.07 2017-11-29 07:02:21
#> 54   -0.09 2017-11-29 06:59:53
#> 55   -0.10 2017-11-29 06:57:25
#> 56   -0.08 2017-11-29 06:54:57
#> 57   -0.13 2017-11-29 06:52:30
#> 58   -0.11 2017-11-29 06:50:02
#> 59   -0.17 2017-11-29 06:47:34
#> 60   -0.21 2017-11-29 06:45:06
#> 61   -0.19 2017-11-29 06:42:39
#> 62   -0.12 2017-11-29 06:40:11
#> 63   -0.15 2017-11-29 06:37:43
#> 64   -0.13 2017-11-29 06:35:15
#> 65   -0.15 2017-11-29 06:32:48
#> 66   -0.18 2017-11-29 06:30:20
#> 67   -0.16 2017-11-29 06:27:52
#> 68   -0.21 2017-11-29 06:25:24
#> 69   -0.21 2017-11-29 06:22:56
#> 70   -0.22 2017-11-29 06:20:29
#> 71   -0.19 2017-11-29 06:18:01
#> 72   -0.15 2017-11-29 06:15:33
#> 73   -0.10 2017-11-29 06:13:04
#> 74   -0.13 2017-11-29 06:10:37
#> 75   -0.13 2017-11-29 06:08:09
#> 76   -0.06 2017-11-29 06:05:41
#> 77   -0.06 2017-11-29 06:03:13
#> 78   -0.09 2017-11-29 06:00:45
#> 79   -0.02 2017-11-29 05:58:17
#> 80    0.04 2017-11-29 05:55:50
#> 81    0.08 2017-11-29 05:53:22
#> 82   -0.04 2017-11-29 05:50:54
#> 83   -0.19 2017-11-29 05:48:26
#> 84   -0.08 2017-11-29 05:45:58
#> 85   -0.09 2017-11-29 05:43:30
#> 86   -0.09 2017-11-29 05:41:02
#> 87   -0.06 2017-11-29 05:38:35
#> 88    0.00 2017-11-29 05:36:07
#> 89    0.03 2017-11-29 05:33:39
#> 90    0.11 2017-11-29 05:31:11
#> 91    0.11 2017-11-29 05:28:43
#> 92    0.05 2017-11-29 05:26:15
#> 93    0.05 2017-11-29 05:23:48
#> 94    0.04 2017-11-29 05:21:20
#> 95    0.04 2017-11-29 05:18:52
#> 96   -0.03 2017-11-29 05:16:24
#> 97    0.01 2017-11-29 05:13:55
#> 98   -0.19 2017-11-29 05:11:27
#> 99   -0.09 2017-11-29 05:09:00
#> 100  -0.05 2017-11-29 05:06:32
#> 101   0.01 2017-11-29 05:04:04
#> 102   0.02 2017-11-29 05:01:36
#> 103  -0.05 2017-11-29 04:59:08
#> 104   0.04 2017-11-29 04:56:40
#> 105   0.15 2017-11-29 04:54:12
#> 106   0.20 2017-11-29 04:51:45
#> 107   0.17 2017-11-29 04:49:16
#> 108   0.14 2017-11-29 04:46:48
#> 109   0.16 2017-11-29 04:44:20
#> 110   0.10 2017-11-29 04:41:52
#> 111   0.11 2017-11-29 04:39:25
#> 112   0.10 2017-11-29 04:36:57
#> 113   0.13 2017-11-29 04:34:29
#> 114  -0.02 2017-11-29 04:32:01
#> 115  -0.05 2017-11-29 04:29:34
#> 116   0.03 2017-11-29 04:27:06
#> 117   0.07 2017-11-29 04:24:38
#> 118   0.08 2017-11-29 04:22:10
#> 119   0.23 2017-11-29 04:19:43
#> 120   0.24 2017-11-29 04:17:15
#> 121   0.24 2017-11-29 04:14:47
#> 122   0.19 2017-11-29 04:12:19
#> 123   0.13 2017-11-29 04:09:51
#> 124   0.20 2017-11-29 04:07:24
#> 125   0.16 2017-11-29 04:04:56
#> 126   0.19 2017-11-29 04:02:28
#> 127   0.11 2017-11-29 04:00:00
#> 128   0.05 2017-11-29 03:57:33
#> 129   0.02 2017-11-29 03:55:05
#> 130   0.00 2017-11-29 03:52:37
#> 131   0.03 2017-11-29 03:50:09
#> 132   0.13 2017-11-29 03:47:42
#> 133   0.13 2017-11-29 03:45:14
#> 134   0.09 2017-11-29 03:42:46
#> 135   0.06 2017-11-29 03:40:18
#> 136   0.07 2017-11-29 03:37:51
#> 137   0.13 2017-11-29 03:35:23
#> 138   0.12 2017-11-29 03:32:55
#> 139   0.10 2017-11-29 03:30:28
#> 140   0.11 2017-11-29 03:28:00
#> 141   0.13 2017-11-29 03:25:32
#> 142   0.01 2017-11-29 03:23:03
#> 143   0.05 2017-11-29 03:20:36
#> 144   0.03 2017-11-29 03:18:08
#> 145   0.06 2017-11-29 03:15:40
#> 146   0.01 2017-11-29 03:13:13
#> 147  -0.09 2017-11-29 03:10:45
#> 148  -0.11 2017-11-29 03:08:17
#> 149  -0.20 2017-11-29 03:05:49
#> 150  -0.23 2017-11-29 03:03:21
#> 151  -0.10 2017-11-29 03:00:54
#> 152  -0.12 2017-11-29 02:58:26
#> 153  -0.07 2017-11-29 02:55:58
#> 154  -0.02 2017-11-29 02:53:30
#> 155   0.02 2017-11-29 02:51:02
#> 156  -0.02 2017-11-29 02:48:34
#> 157   0.10 2017-11-29 02:46:06
#> 158   0.10 2017-11-29 02:43:39
#> 159   0.00 2017-11-29 02:41:11
#> 160  -0.02 2017-11-29 02:38:31
#> 161  -0.05 2017-11-29 02:36:03
#> 162   0.04 2017-11-29 02:33:36
#> 163   0.16 2017-11-29 02:31:08
#> 164   0.14 2017-11-29 02:28:40
#> 165   0.14 2017-11-29 02:26:12
#> 166   0.15 2017-11-29 02:23:44
#> 167   0.16 2017-11-29 02:21:17
#> 168   0.07 2017-11-29 02:18:48
#> 169  -0.02 2017-11-29 02:16:21
#> 170  -0.04 2017-11-29 02:13:53
#> 171  -0.10 2017-11-29 02:11:25
#> 172  -0.13 2017-11-29 02:08:58
#> 173  -0.11 2017-11-29 02:06:30
#> 174  -0.06 2017-11-29 02:04:02
#> 175  -0.10 2017-11-29 02:01:34
#> 176  -0.05 2017-11-29 01:59:07
#> 177  -0.11 2017-11-29 01:56:39
#> 178  -0.15 2017-11-29 01:54:11
#> 179  -0.14 2017-11-29 01:51:43
#> 180  -0.07 2017-11-29 01:49:16
#> 181  -0.02 2017-11-29 01:46:48
#> 182   0.08 2017-11-29 01:44:20
#> 183   0.09 2017-11-29 01:41:52
#> 184   0.05 2017-11-29 01:39:25
#> 185  -0.07 2017-11-29 01:36:57
#> 186  -0.01 2017-11-29 01:34:29
#> 187   0.18 2017-11-29 01:32:01
#> 188   0.33 2017-11-29 01:29:34
#> 189   0.29 2017-11-29 01:27:06
#> 190   0.10 2017-11-29 01:24:38
#> 191   0.10 2017-11-29 01:22:10
#> 192   0.20 2017-11-29 01:19:43
#> 193   0.15 2017-11-29 01:17:15
#> 194   0.20 2017-11-29 01:14:47
#> 195   0.17 2017-11-29 01:12:19
#> 196   0.13 2017-11-29 01:09:51
#> 197   0.14 2017-11-29 01:07:24
#> 198   0.07 2017-11-29 01:04:56
#> 199   0.04 2017-11-29 01:02:28
#> 200  -0.02 2017-11-29 01:00:00
#> 201  -0.03 2017-11-29 00:57:32
#> 202  -0.16 2017-11-29 00:55:05
#> 203  -0.13 2017-11-29 00:52:37
#> 204   0.09 2017-11-29 00:50:09
#> 205   0.13 2017-11-29 00:47:41
#> 206   0.07 2017-11-29 00:45:14
#> 207  -0.04 2017-11-29 00:42:46
#> 208  -0.05 2017-11-29 00:40:18
#> 209   0.02 2017-11-29 00:37:50
#> 210   0.10 2017-11-29 00:35:23
#> 211   0.32 2017-11-29 00:32:55
#> 212   0.37 2017-11-29 00:30:27
#> 213   0.32 2017-11-29 00:27:59
#> 214   0.33 2017-11-29 00:25:31
#> 215   0.39 2017-11-29 00:23:03
#> 216   0.35 2017-11-29 00:20:36
#> 217   0.29 2017-11-29 00:18:08
#> 218   0.42 2017-11-29 00:15:40
#> 219   0.36 2017-11-29 00:13:13
#> 220   0.34 2017-11-29 00:10:45
#> 221   0.27 2017-11-29 00:08:17
#> 222   0.25 2017-11-29 00:05:49
#> 223   0.27 2017-11-29 00:03:21
#> 224   0.31 2017-11-29 00:00:54
#> 225   0.25 2017-11-28 23:58:26
#> 226   0.21 2017-11-28 23:55:58
#> 227   0.20 2017-11-28 23:53:30
#> 228   0.23 2017-11-28 23:51:03
#> 229   0.15 2017-11-28 23:48:35
#> 230   0.22 2017-11-28 23:46:07
#> 231   0.13 2017-11-28 23:43:39
#> 232   0.17 2017-11-28 23:41:11
#> 233   0.27 2017-11-28 23:38:44
#> 234   0.37 2017-11-28 23:36:16
#> 235   0.47 2017-11-28 23:33:48
#> 236   0.48 2017-11-28 23:31:20
#> 237   0.44 2017-11-28 23:28:53
#> 238   0.41 2017-11-28 23:26:25
#> 239   0.33 2017-11-28 23:23:57
#> 240   0.37 2017-11-28 23:21:29
#> 241   0.42 2017-11-28 23:19:01
#> 242   0.33 2017-11-28 23:16:34
#> 243   0.31 2017-11-28 23:14:06
#> 244   0.18 2017-11-28 23:11:38
#> 245   0.24 2017-11-28 23:09:10
#> 246   0.19 2017-11-28 23:06:42
#> 247   0.12 2017-11-28 23:04:14
#> 248   0.17 2017-11-28 23:01:47
#> 249   0.22 2017-11-28 22:59:19
#> 250   0.30 2017-11-28 22:56:51
#> 251   0.31 2017-11-28 22:54:23
#> 252   0.31 2017-11-28 22:51:55
#> 253   0.31 2017-11-28 22:49:27
#> 254   0.31 2017-11-28 22:46:59
#> 255   0.28 2017-11-28 22:44:31
#> 256   0.34 2017-11-28 22:42:03
#> 257   0.36 2017-11-28 22:39:35
#> 258   0.26 2017-11-28 22:37:07
#> 259   0.28 2017-11-28 22:34:39
#> 260   0.23 2017-11-28 22:32:12
#> 261   0.28 2017-11-28 22:29:43
#> 262   0.39 2017-11-28 22:27:15
#> 263   0.41 2017-11-28 22:24:47
#> 264   0.33 2017-11-28 22:22:19
#> 265   0.37 2017-11-28 22:19:51
#> 266   0.31 2017-11-28 22:17:24
#> 267   0.31 2017-11-28 22:14:56
#> 268   0.28 2017-11-28 22:12:28
#> 269   0.33 2017-11-28 22:10:00
#> 270   0.46 2017-11-28 22:07:32
#> 271   0.40 2017-11-28 22:05:05
#> 272   0.45 2017-11-28 22:02:37
#> 273   0.49 2017-11-28 22:00:09
#> 274   0.48 2017-11-28 21:57:41
#> 275   0.49 2017-11-28 21:55:13
#> 276   0.46 2017-11-28 21:52:45
#> 277   0.48 2017-11-28 21:50:18
#> 278   0.48 2017-11-28 21:47:50
#> 279   0.47 2017-11-28 21:45:22
#> 280   0.47 2017-11-28 21:42:54
#> 281   0.51 2017-11-28 21:40:26
#> 282   0.42 2017-11-28 21:37:59
#> 283   0.55 2017-11-28 21:35:31
#> 284   0.66 2017-11-28 21:33:03
#> 285   0.68 2017-11-28 21:30:35
#> 286   0.67 2017-11-28 21:28:07
#> 287   0.68 2017-11-28 21:25:40
#> 288   0.71 2017-11-28 21:23:12
#> 289   0.74 2017-11-28 21:20:44
#> 290   0.76 2017-11-28 21:18:16
#> 291   0.77 2017-11-28 21:15:48
#> 292   0.82 2017-11-28 21:13:21
#> 293   0.88 2017-11-28 21:10:53
#> 294   0.84 2017-11-28 21:08:25
#> 295   0.85 2017-11-28 21:05:57
#> 296   0.82 2017-11-28 21:03:29
#> 297   0.86 2017-11-28 21:01:01
#> 298   0.89 2017-11-28 20:58:33
#> 299   0.89 2017-11-28 20:56:05
#> 300   0.92 2017-11-28 20:53:37
#> 301   0.88 2017-11-28 20:51:10
#> 302   0.83 2017-11-28 20:48:42
#> 303   0.88 2017-11-28 20:46:14
#> 304   0.94 2017-11-28 20:43:45
#> 305   0.91 2017-11-28 20:41:17
#> 306   0.84 2017-11-28 20:38:49
#> 307   0.79 2017-11-28 20:36:21
#> 308   0.74 2017-11-28 20:33:53
#> 309   0.72 2017-11-28 20:31:24
#> 310   0.72 2017-11-28 20:28:56
#> 311   0.67 2017-11-28 20:26:28
#> 312   0.71 2017-11-28 20:23:59
#> 313   0.77 2017-11-28 20:21:31
#> 314   0.78 2017-11-28 20:19:03
#> 315   0.83 2017-11-28 20:16:33
#> 316   0.75 2017-11-28 20:14:05
#> 317   0.78 2017-11-28 20:11:37
#> 318   0.82 2017-11-28 20:09:09
#> 319   0.77 2017-11-28 20:06:41
#> 320   0.73 2017-11-28 20:04:13
#> 321   0.86 2017-11-28 20:01:44
#> 322   0.95 2017-11-28 19:59:17
#> 323   0.90 2017-11-28 19:56:48
#> 324   0.87 2017-11-28 19:54:20
#> 325   0.81 2017-11-28 19:51:50
#> 326   0.86 2017-11-28 19:49:21
#> 327   0.88 2017-11-28 19:46:53
#> 328   0.93 2017-11-28 19:44:25
#> 329   0.80 2017-11-28 19:41:57
#> 330   0.82 2017-11-28 19:39:27
#> 331   0.88 2017-11-28 19:36:59
#> 332   0.92 2017-11-28 19:34:30
#> 333   0.93 2017-11-28 19:31:59
#> 334   0.95 2017-11-28 19:29:31
#> 335   0.96 2017-11-28 19:27:02
#> 336   1.04 2017-11-28 19:24:31
#> 337   0.99 2017-11-28 19:22:03
#> 338   1.01 2017-11-28 19:19:34
#> 339   1.07 2017-11-28 19:17:06
#> 340   0.99 2017-11-28 19:14:38
#> 341   0.91 2017-11-28 19:12:07
#> 342   0.89 2017-11-28 19:09:39
#> 343   0.88 2017-11-28 19:07:11
#> 344   1.12 2017-11-28 19:04:43
#> 345   1.15 2017-11-28 19:02:15
#> 346   1.15 2017-11-28 18:59:47
#> 347   1.22 2017-11-28 18:57:20
#> 348   1.20 2017-11-28 18:54:52
#> 349   1.09 2017-11-28 18:52:24
#> 350   1.04 2017-11-28 18:49:56
#> 351   1.06 2017-11-28 18:47:28
#> 352   1.12 2017-11-28 18:45:00
#> 353   1.21 2017-11-28 18:42:32
#> 354   1.25 2017-11-28 18:40:02
#> 355   1.25 2017-11-28 18:37:33
#> 356   1.26 2017-11-28 18:35:05
#> 357   1.31 2017-11-28 18:32:38
#> 358   1.27 2017-11-28 18:30:09
#> 359   1.26 2017-11-28 18:27:40
#> 360   1.25 2017-11-28 18:25:12
#> 361   1.22 2017-11-28 18:22:43
#> 362   1.15 2017-11-28 18:20:14
#> 363   1.23 2017-11-28 18:17:46
#> 364   1.18 2017-11-28 18:15:18
#> 365   1.25 2017-11-28 18:12:49
#> 366   1.39 2017-11-28 18:10:19
#> 367   1.35 2017-11-28 18:07:51
#> 368   1.34 2017-11-28 18:05:24
#> 369   1.36 2017-11-28 18:02:56
#> 370   1.34 2017-11-28 18:00:28
#> 371   1.39 2017-11-28 17:58:00
#> 372   1.37 2017-11-28 17:55:32
#> 373   1.36 2017-11-28 17:53:04
#> 374   1.38 2017-11-28 17:50:36
#> 375   1.42 2017-11-28 17:48:07
#> 376   1.44 2017-11-28 17:45:39
#> 377   1.45 2017-11-28 17:43:12
#> 378   1.45 2017-11-28 17:40:44
#> 379   1.41 2017-11-28 17:38:16
#> 380   1.41 2017-11-28 17:35:47
#> 381   1.35 2017-11-28 17:33:19
#> 382   1.49 2017-11-28 17:30:51
#> 383   1.51 2017-11-28 17:28:23
#> 384   1.51 2017-11-28 17:25:55
#> 385   1.51 2017-11-28 17:23:27
#> 386   1.62 2017-11-28 17:20:59
#> 387   1.69 2017-11-28 17:18:31
#> 388   1.61 2017-11-28 17:16:03
#> 389   1.53 2017-11-28 17:13:35
#> 390   1.75 2017-11-28 17:11:07
#> 391   1.76 2017-11-28 17:08:40
#> 392   1.76 2017-11-28 17:06:12
#> 393   1.78 2017-11-28 17:03:43
#> 394   1.78 2017-11-28 17:01:15
#> 395   1.78 2017-11-28 16:58:47
#> 396   1.80 2017-11-28 16:56:19
#> 397   1.88 2017-11-28 16:53:51
#> 398   1.85 2017-11-28 16:51:23
#> 399   1.82 2017-11-28 16:48:56
#> 400   1.89 2017-11-28 16:46:28
#> 401   1.86 2017-11-28 16:44:00
#> 402   1.86 2017-11-28 16:41:30
#> 403   1.95 2017-11-28 16:39:02
#> 404   2.00 2017-11-28 16:36:35
#> 405   2.05 2017-11-28 16:34:07
#> 406   2.03 2017-11-28 16:31:39
#> 407   2.04 2017-11-28 16:29:11
#> 408   2.02 2017-11-28 16:26:42
#> 409   2.06 2017-11-28 16:24:15
#> 410   2.11 2017-11-28 16:21:47
#> 411   2.12 2017-11-28 16:19:19
#> 412   2.13 2017-11-28 16:16:51
#> 413   2.14 2017-11-28 16:14:23
#> 414   2.15 2017-11-28 16:11:55
#> 415   2.18 2017-11-28 16:09:27
#> 416   2.14 2017-11-28 16:06:59
#> 417   2.15 2017-11-28 16:04:32
#> 418   2.18 2017-11-28 16:02:04
#> 419   2.13 2017-11-28 15:59:36
#> 420   2.20 2017-11-28 15:57:08
#> 421   2.18 2017-11-28 15:54:40
#> 422   2.22 2017-11-28 15:52:12
#> 423   2.33 2017-11-28 15:49:45
#> 424   2.26 2017-11-28 15:47:16
#> 425   2.23 2017-11-28 15:44:49
#> 426   2.14 2017-11-28 15:42:21
#> 427   2.29 2017-11-28 15:39:53
#> 428   2.33 2017-11-28 15:37:25
#> 429   2.37 2017-11-28 15:34:57
#> 430   2.32 2017-11-28 15:32:29
#> 431   2.30 2017-11-28 15:30:01
#> 432   2.37 2017-11-28 15:27:33
#> 433   2.48 2017-11-28 15:25:05
#> 434   2.48 2017-11-28 15:19:33
#> 435   2.52 2017-11-28 15:17:04
#> 436   2.64 2017-11-28 15:14:36
#> 437   2.68 2017-11-28 15:08:48
#> 438   2.69 2017-11-28 15:06:14
#> 439   2.72 2017-11-28 15:03:41
#> 440   2.73 2017-11-28 15:00:52
#> 441   2.85 2017-11-28 14:55:17
#> 442   2.83 2017-11-28 14:52:44
#> 443   2.77 2017-11-28 14:50:15
#> 444   2.82 2017-11-28 14:47:48
#> 445   2.75 2017-11-28 14:45:20
#> 446   2.77 2017-11-28 14:39:48
#> 447   2.82 2017-11-28 14:37:20
#> 448   2.89 2017-11-28 14:34:52
#> 449   3.07 2017-11-28 14:29:22
#> 450   3.13 2017-11-28 14:26:54
#> 451   3.08 2017-11-28 14:24:06
#> 452   3.32 2017-11-28 14:21:33
#> 453   3.50 2017-11-28 14:19:04
#> 454   3.52 2017-11-28 14:16:36
#> 455   3.57 2017-11-28 14:14:09
#> 456   3.55 2017-11-28 14:11:41
#> 457   3.63 2017-11-28 14:09:13
#> 458   3.52 2017-11-28 14:06:45
#> 459   3.48 2017-11-28 14:04:17
#> 460   3.54 2017-11-28 14:01:49
#> 461   3.53 2017-11-28 13:59:22
#> 462   3.52 2017-11-28 13:56:54
#> 463   3.51 2017-11-28 13:54:26
#> 464   3.36 2017-11-28 13:51:58
#> 465   3.51 2017-11-28 13:49:30
#> 466   3.70 2017-11-28 13:47:02
#> 467   3.79 2017-11-28 13:44:35
#> 468   3.95 2017-11-28 13:42:07
#> 469   4.25 2017-11-28 13:39:39
#> 470   4.21 2017-11-28 13:37:11
#> 471   4.27 2017-11-28 13:34:43
#> 472   4.37 2017-11-28 13:32:15
#> 473   4.36 2017-11-28 13:29:47
#> 474   4.34 2017-11-28 13:27:19
#> 475   4.36 2017-11-28 13:24:50
#> 476   4.51 2017-11-28 13:22:22
#> 477   4.58 2017-11-28 13:19:55
#> 478   4.60 2017-11-28 13:17:25
#> 479   4.59 2017-11-28 13:14:57
#> 480   4.48 2017-11-28 13:12:29
#> 481   4.34 2017-11-28 13:10:01
#> 482   4.54 2017-11-28 13:07:34
#> 483   4.57 2017-11-28 13:05:05
#> 484   4.31 2017-11-28 13:02:38
#> 485   4.26 2017-11-28 13:00:09
#> 486   4.39 2017-11-28 12:57:41
#> 487   4.28 2017-11-28 12:55:13
#> 488   4.74 2017-11-28 12:52:45
#> 489   4.50 2017-11-28 12:50:18
#> 490   4.26 2017-11-28 12:47:50
#> 491   4.64 2017-11-28 12:45:22
#> 492   5.17 2017-11-28 12:42:54
#> 493   5.10 2017-11-28 12:40:26
#> 494   5.07 2017-11-28 12:37:58
#> 495   5.02 2017-11-28 12:35:30
#> 496   4.52 2017-11-28 12:33:03
#> 497   4.94 2017-11-28 12:30:35
#> 498   5.05 2017-11-28 12:28:07
#> 499   4.98 2017-11-28 12:25:39
#> 500   4.70 2017-11-28 12:23:11
#> 501   4.50 2017-11-28 12:20:43
#> 502   4.58 2017-11-28 12:18:16
#> 503   4.07 2017-11-28 12:15:48
#> 504   3.96 2017-11-28 12:13:20
#> 505   3.79 2017-11-28 12:10:52
#> 506   4.03 2017-11-28 12:08:24
#> 507   4.02 2017-11-28 12:05:57
#> 508   4.32 2017-11-28 12:03:29
#> 509   4.98 2017-11-28 12:01:01
#> 510   5.14 2017-11-28 11:58:33
#> 511   5.09 2017-11-28 11:56:06
#> 512   4.58 2017-11-28 11:53:38
#> 513   4.44 2017-11-28 11:51:10
#> 514   5.45 2017-11-28 11:48:42
#> 515   4.93 2017-11-28 11:46:14
#> 516   4.49 2017-11-28 11:43:47
#> 517   4.23 2017-11-28 11:41:19
#> 518   4.35 2017-11-28 11:38:51
#> 519   4.70 2017-11-28 11:36:23
#> 520   5.28 2017-11-28 11:33:56
#> 521   4.94 2017-11-28 11:31:28
#> 522   4.09 2017-11-28 11:29:00
#> 523   3.64 2017-11-28 11:26:31
#> 524   3.80 2017-11-28 11:24:03
#> 525   3.77 2017-11-28 11:21:36
#> 526   3.81 2017-11-28 11:19:08
#> 527   4.38 2017-11-28 11:16:40
#> 528   5.31 2017-11-28 11:14:12
#> 529   6.09 2017-11-28 11:11:44
#> 530   5.87 2017-11-28 11:09:17
#> 531   5.47 2017-11-28 11:06:49
#> 532   5.28 2017-11-28 11:04:21
#> 533   5.09 2017-11-28 11:01:53
#> 534   4.70 2017-11-28 10:59:25
#> 535   3.94 2017-11-28 10:56:57
#> 536   3.90 2017-11-28 10:54:30
#> 537   4.03 2017-11-28 10:52:02
#> 538   3.93 2017-11-28 10:49:34
#> 539   4.34 2017-11-28 10:47:06
#> 540   4.61 2017-11-28 10:42:11
#> 541   4.14 2017-11-28 10:39:44
#> 542   2.88 2017-11-28 10:37:16
#> 543   3.45 2017-11-28 10:34:48
#> 544   3.33 2017-11-28 10:32:20
#> 545   3.46 2017-11-28 10:29:52
#> 546   3.49 2017-11-28 10:27:24
#> 547   3.00 2017-11-28 10:24:56
#> 548   3.31 2017-11-28 10:22:29
#> 549   4.41 2017-11-28 10:20:01
#> 550   4.15 2017-11-28 10:17:33
#> 551   4.45 2017-11-28 10:15:05
#> 552   3.98 2017-11-28 10:12:37
#> 553   3.20 2017-11-28 10:10:10
#> 554   3.20 2017-11-28 10:07:42
#> 555   2.96 2017-11-28 10:05:14
#> 556   3.24 2017-11-28 10:02:46
#> 557   4.13 2017-11-28 10:00:19
#> 558   3.68 2017-11-28 09:57:50
#> 559   3.50 2017-11-28 09:55:22
#> 560   3.46 2017-11-28 09:52:54
#> 561   3.32 2017-11-28 09:50:27
#> 562   3.16 2017-11-28 09:47:59
#> 563   2.77 2017-11-28 09:45:31
#> 564   2.53 2017-11-28 09:43:03
#> 565   2.14 2017-11-28 09:40:36
#> 566   2.26 2017-11-28 09:38:08
#> 567   2.42 2017-11-28 09:35:40
#> 568   2.47 2017-11-28 09:33:12
#> 569   2.11 2017-11-28 09:30:44
#> 570   1.69 2017-11-28 09:28:17
#> 571   1.76 2017-11-28 09:25:49
#> 572   1.77 2017-11-28 09:23:21
#> 573   1.87 2017-11-28 09:20:53
#> 574   1.77 2017-11-28 09:18:25
#> 575   1.84 2017-11-28 09:15:57
#> 576   1.81 2017-11-28 09:13:30
#> 577   1.43 2017-11-28 09:11:02
#> 578   1.07 2017-11-28 09:08:34
#> 579   0.79 2017-11-28 09:06:06
#> 580   0.47 2017-11-28 09:03:39
#> 581   0.26 2017-11-28 09:01:11
#> 582  -0.06 2017-11-28 08:58:43
#> 583  -0.30 2017-11-28 08:56:15
#> 584  -0.20 2017-11-28 08:53:47
#> 585  -0.49 2017-11-28 08:51:19
#> 586  -0.64 2017-11-28 08:48:51
#> 587  -0.76 2017-11-28 08:46:24
#> 588  -0.87 2017-11-28 08:43:56
#> 589  -1.20 2017-11-28 08:41:28
#> 590  -1.32 2017-11-28 08:39:00
#> 591  -1.35 2017-11-28 08:36:32
#> 592  -1.29 2017-11-28 08:34:04
#> 593  -1.15 2017-11-28 08:31:36
#> 594  -0.83 2017-11-28 08:29:08
#> 595  -1.12 2017-11-28 08:26:41
#> 596  -1.32 2017-11-28 08:24:13
#> 597  -1.07 2017-11-28 08:21:45
#> 598  -0.85 2017-11-28 08:19:17
#> 599  -1.09 2017-11-28 08:16:49
#> 600  -1.32 2017-11-28 08:14:21
#> 601  -1.26 2017-11-28 08:11:54
#> 602  -0.75 2017-11-28 08:09:26
#> 603  -0.66 2017-11-28 08:06:58
#> 604  -0.82 2017-11-28 08:04:30
#> 605  -0.97 2017-11-28 08:02:02
#> 606  -1.11 2017-11-28 07:59:34
#> 607  -0.98 2017-11-28 07:57:06
#> 608  -0.89 2017-11-28 07:54:39
#> 609  -1.19 2017-11-28 07:52:11
#> 610  -1.31 2017-11-28 07:49:43
#> 611  -1.28 2017-11-28 07:47:15
#> 612  -1.37 2017-11-28 07:44:47
#> 613  -1.47 2017-11-28 07:42:20
#> 614  -1.62 2017-11-28 07:39:52
#> 615  -1.59 2017-11-28 07:37:24
#> 616  -1.54 2017-11-28 07:34:56
#> 617  -1.54 2017-11-28 07:32:28
#> 618  -1.58 2017-11-28 07:30:00
#> 619  -1.59 2017-11-28 07:27:32
#> 620  -1.64 2017-11-28 07:25:04
#> 621  -1.68 2017-11-28 07:22:35
#> 622  -1.69 2017-11-28 07:20:07
#> 623  -1.72 2017-11-28 07:17:39
#> 624  -1.71 2017-11-28 07:15:10
#> 625  -1.80 2017-11-28 07:12:42
#> 626  -1.84 2017-11-28 07:10:15
#> 627  -1.83 2017-11-28 07:07:47
#> 628  -1.93 2017-11-28 07:05:19
#> 629  -1.96 2017-11-28 07:02:51
#> 630  -1.95 2017-11-28 07:00:23
#> 631  -2.00 2017-11-28 06:57:55
#> 632  -2.16 2017-11-28 06:55:28
#> 633  -2.35 2017-11-28 06:53:00
#> 634  -2.23 2017-11-28 06:50:32
#> 635  -2.11 2017-11-28 06:48:04
#> 636  -2.22 2017-11-28 06:45:36
#> 637  -2.24 2017-11-28 06:43:07
#> 638  -2.24 2017-11-28 06:40:39
#> 639  -2.16 2017-11-28 06:38:11
#> 640  -2.18 2017-11-28 06:35:44
#> 641  -2.21 2017-11-28 06:33:16
#> 642  -2.13 2017-11-28 06:30:48
#> 643  -2.14 2017-11-28 06:28:20
#> 644  -2.11 2017-11-28 06:25:52
#> 645  -2.21 2017-11-28 06:23:24
#> 646  -2.28 2017-11-28 06:20:56
#> 647  -2.22 2017-11-28 06:18:28
#> 648  -2.28 2017-11-28 06:16:00
#> 649  -2.30 2017-11-28 06:13:32
#> 650  -2.39 2017-11-28 06:11:04
#> 651  -2.39 2017-11-28 06:08:36
#> 652  -2.39 2017-11-28 06:06:08
#> 653  -2.33 2017-11-28 06:03:41
#> 654  -2.41 2017-11-28 06:01:13
#> 655  -2.44 2017-11-28 05:58:44
#> 656  -2.35 2017-11-28 05:56:16
#> 657  -2.37 2017-11-28 05:53:48
#> 658  -2.45 2017-11-28 05:51:21
#> 659  -2.30 2017-11-28 05:48:53
#> 660  -2.23 2017-11-28 05:46:25
#> 661  -2.30 2017-11-28 05:43:57
#> 662  -2.25 2017-11-28 05:41:29
#> 663  -2.19 2017-11-28 05:39:01
#> 664  -2.23 2017-11-28 05:36:33
#> 665  -2.27 2017-11-28 05:34:06
#> 666  -2.28 2017-11-28 05:31:38
#> 667  -2.27 2017-11-28 05:29:10
#> 668  -2.30 2017-11-28 05:26:41
#> 669  -2.44 2017-11-28 05:24:13
#> 670  -2.43 2017-11-28 05:21:45
#> 671  -2.28 2017-11-28 05:19:17
#> 672  -2.14 2017-11-28 05:16:49
#> 673  -2.20 2017-11-28 05:14:21
#> 674  -2.28 2017-11-28 05:11:53
#> 675  -2.17 2017-11-28 05:09:26
#> 676  -2.13 2017-11-28 05:06:58
#> 677  -2.16 2017-11-28 05:04:30
#> 678  -2.20 2017-11-28 05:02:02
#> 679  -2.22 2017-11-28 04:59:34
#> 680  -2.20 2017-11-28 04:57:06
#> 681  -2.20 2017-11-28 04:54:39
#> 682  -2.13 2017-11-28 04:52:11
#> 683  -2.07 2017-11-28 04:49:43
#> 684  -2.13 2017-11-28 04:47:15
#> 685  -2.20 2017-11-28 04:44:47
#> 686  -2.11 2017-11-28 04:42:19
#> 687  -2.14 2017-11-28 04:39:51
#> 688  -2.15 2017-11-28 04:37:24
#> 689  -2.22 2017-11-28 04:34:56
#> 690  -2.18 2017-11-28 04:32:28
#> 691  -2.10 2017-11-28 04:30:00
#> 692  -2.03 2017-11-28 04:27:32
#> 693  -1.98 2017-11-28 04:25:04
#> 694  -1.99 2017-11-28 04:22:37
#> 695  -1.99 2017-11-28 04:20:09
#> 696  -2.04 2017-11-28 04:17:41
#> 697  -1.87 2017-11-28 04:15:13
#> 698  -1.95 2017-11-28 04:12:45
#> 699  -1.99 2017-11-28 04:10:17
#> 700  -2.04 2017-11-28 04:07:49
#> 701  -2.00 2017-11-28 04:05:22
#> 702  -1.99 2017-11-28 04:02:54
#> 703  -1.88 2017-11-28 04:00:26
#> 704  -1.85 2017-11-28 03:57:55
#> 705  -1.76 2017-11-28 03:55:28
#> 706  -1.77 2017-11-28 03:53:00
#> 707  -1.72 2017-11-28 03:50:32
#> 708  -1.68 2017-11-28 03:48:04
#> 709  -1.50 2017-11-28 03:42:38
#> 710  -1.49 2017-11-28 03:40:10
#> 711  -1.41 2017-11-28 03:37:42
#> 712  -1.30 2017-11-28 03:35:14
#> 713  -1.32 2017-11-28 03:32:47
#> 714  -1.37 2017-11-28 03:30:19
#> 715  -1.35 2017-11-28 03:27:40
#> 716  -1.30 2017-11-28 03:21:59
#> 717  -1.38 2017-11-28 03:19:28
#> 718  -1.38 2017-11-28 03:16:58
#> 719  -1.42 2017-11-28 03:11:33
#> 720  -1.39 2017-11-28 03:09:05
#> 721  -1.37 2017-11-28 03:06:36
#> 722  -1.38 2017-11-28 03:04:09
#> 723  -1.36 2017-11-28 03:01:41
#> 724  -1.35 2017-11-28 02:59:06
#> 725  -1.34 2017-11-28 02:56:23
#> 726  -1.38 2017-11-28 02:53:55
#> 727  -1.48 2017-11-28 02:51:27
#> 728  -1.49 2017-11-28 02:48:59
#> 729  -1.47 2017-11-28 02:46:19
#> 730  -1.55 2017-11-28 02:43:50
#> 731  -1.71 2017-11-28 02:41:22
#> 732  -1.66 2017-11-28 02:38:45
#> 733  -1.57 2017-11-28 02:36:18
#> 734  -1.44 2017-11-28 02:33:50
#> 735  -1.48 2017-11-28 02:31:22
#> 736  -1.49 2017-11-28 02:28:54
#> 737  -1.46 2017-11-28 02:26:26
#> 738  -1.51 2017-11-28 02:23:58
#> 739  -1.53 2017-11-28 02:21:31
#> 740  -1.49 2017-11-28 02:19:03
#> 741  -1.41 2017-11-28 02:16:35
#> 742  -1.44 2017-11-28 02:14:07
#> 743  -1.47 2017-11-28 02:11:39
#> 744  -1.49 2017-11-28 02:09:12
#> 745  -1.46 2017-11-28 02:06:44
#> 746  -1.52 2017-11-28 02:04:16
#> 747  -1.52 2017-11-28 02:01:48
#> 748  -1.53 2017-11-28 01:59:20
#> 749  -1.53 2017-11-28 01:56:52
#> 750  -1.57 2017-11-28 01:54:25
#> 751  -1.61 2017-11-28 01:51:57
#> 752  -1.67 2017-11-28 01:49:29
#> 753  -1.67 2017-11-28 01:47:01
#> 754  -1.71 2017-11-28 01:44:33
#> 755  -1.75 2017-11-28 01:42:06
#> 756  -1.72 2017-11-28 01:39:38
#> 757  -1.71 2017-11-28 01:37:10
#> 758  -1.62 2017-11-28 01:34:42
#> 759  -1.61 2017-11-28 01:32:15
#> 760  -1.61 2017-11-28 01:29:47
#> 761  -1.62 2017-11-28 01:27:19
#> 762  -1.52 2017-11-28 01:24:51
#> 763  -1.53 2017-11-28 01:22:23
#> 764  -1.53 2017-11-28 01:19:56
#> 765  -1.53 2017-11-28 01:17:28
#> 766  -1.58 2017-11-28 01:15:00
#> 767  -1.53 2017-11-28 01:12:32
#> 768  -1.57 2017-11-28 01:10:04
#> 769  -1.59 2017-11-28 01:07:37
#> 770  -1.57 2017-11-28 01:05:09
#> 771  -1.58 2017-11-28 01:02:41
#> 772  -1.59 2017-11-28 01:00:13
#> 773  -1.64 2017-11-28 00:57:46
#> 774  -1.65 2017-11-28 00:55:18
#> 775  -1.67 2017-11-28 00:52:50
#> 776  -1.69 2017-11-28 00:50:22
#> 777  -1.70 2017-11-28 00:47:55
#> 778  -1.74 2017-11-28 00:45:27
#> 779  -1.77 2017-11-28 00:42:59
#> 780  -1.85 2017-11-28 00:40:31
#> 781  -1.94 2017-11-28 00:38:03
#> 782  -1.94 2017-11-28 00:35:36
#> 783  -1.82 2017-11-28 00:33:08
#> 784  -1.70 2017-11-28 00:30:40
#> 785  -1.71 2017-11-28 00:28:12
#> 786  -1.67 2017-11-28 00:25:45
#> 787  -1.67 2017-11-28 00:23:17
#> 788  -1.71 2017-11-28 00:20:49
#> 789  -1.76 2017-11-28 00:18:21
#> 790  -1.79 2017-11-28 00:15:53
#> 791  -1.85 2017-11-28 00:13:26
#> 792  -1.92 2017-11-28 00:10:58
#> 793  -2.00 2017-11-28 00:08:30
#> 794  -2.03 2017-11-28 00:06:02
#> 795  -2.08 2017-11-28 00:03:34
#> 796  -2.08 2017-11-28 00:01:07
#> 797  -2.21 2017-11-27 23:58:38
#> 798  -2.36 2017-11-27 23:56:10
#> 799  -2.36 2017-11-27 23:53:42
#> 800  -2.41 2017-11-27 23:51:14
#> 801  -2.40 2017-11-27 23:48:47
#> 802  -2.48 2017-11-27 23:46:19
#> 803  -2.50 2017-11-27 23:43:51
#> 804  -2.60 2017-11-27 23:41:23
#> 805  -2.70 2017-11-27 23:38:55
#> 806  -2.81 2017-11-27 23:36:28
#> 807  -2.84 2017-11-27 23:34:00
#> 808  -2.90 2017-11-27 23:31:32
#> 809  -2.97 2017-11-27 23:29:04
#> 810  -2.92 2017-11-27 23:26:37
#> 811  -2.81 2017-11-27 23:24:09
#> 812  -2.54 2017-11-27 23:21:41
#> 813  -2.34 2017-11-27 23:19:13
#> 814  -2.27 2017-11-27 23:16:45
#> 815  -2.43 2017-11-27 23:14:18
#> 816  -2.55 2017-11-27 23:11:50
#> 817  -2.60 2017-11-27 23:09:22
#> 818  -2.60 2017-11-27 23:06:54
#> 819  -2.59 2017-11-27 23:04:26
#> 820  -2.70 2017-11-27 23:01:59
#> 821  -2.54 2017-11-27 22:59:31
#> 822  -2.55 2017-11-27 22:57:03
#> 823  -2.61 2017-11-27 22:54:35
#> 824  -2.71 2017-11-27 22:52:07
#> 825  -2.80 2017-11-27 22:49:40
#> 826  -2.83 2017-11-27 22:47:12
#> 827  -2.90 2017-11-27 22:44:44
#> 828  -2.90 2017-11-27 22:42:16
#> 829  -2.93 2017-11-27 22:39:48
#> 830  -2.91 2017-11-27 22:37:21
#> 831  -2.93 2017-11-27 22:34:52
#> 832  -2.93 2017-11-27 22:32:25
#> 833  -2.94 2017-11-27 22:29:57
#> 834  -2.85 2017-11-27 22:27:29
#> 835  -2.90 2017-11-27 22:25:01
#> 836  -2.88 2017-11-27 22:22:33
#> 837  -2.96 2017-11-27 22:20:06
#> 838  -3.04 2017-11-27 22:17:38
#> 839  -3.20 2017-11-27 22:15:10
#> 840  -3.12 2017-11-27 22:12:42
#> 841  -2.86 2017-11-27 22:10:14
#> 842  -2.96 2017-11-27 22:07:46
#> 843  -2.96 2017-11-27 22:05:19
#> 844  -2.99 2017-11-27 22:02:51
#> 845  -2.98 2017-11-27 22:00:23
#> 846  -2.88 2017-11-27 21:57:55
#> 847  -2.80 2017-11-27 21:55:28
#> 848  -2.72 2017-11-27 21:53:00
#> 849  -2.72 2017-11-27 21:50:32
#> 850  -2.67 2017-11-27 21:48:04
#> 851  -2.68 2017-11-27 21:45:36
#> 852  -2.74 2017-11-27 21:43:08
#> 853  -2.69 2017-11-27 21:40:40
#> 854  -2.73 2017-11-27 21:38:12
#> 855  -2.77 2017-11-27 21:35:44
#> 856  -2.79 2017-11-27 21:33:16
#> 857  -2.76 2017-11-27 21:30:49
#> 858  -2.65 2017-11-27 21:28:21
#> 859  -2.59 2017-11-27 21:25:53
#> 860  -2.64 2017-11-27 21:23:25
#> 861  -2.62 2017-11-27 21:20:57
#> 862  -2.78 2017-11-27 21:18:30
#> 863  -2.63 2017-11-27 21:16:02
#> 864  -2.66 2017-11-27 21:13:33
#> 865  -2.54 2017-11-27 21:11:05
#> 866  -2.49 2017-11-27 21:08:37
#> 867  -2.49 2017-11-27 21:06:09
#> 868  -2.44 2017-11-27 21:03:41
#> 869  -2.50 2017-11-27 21:01:14
#> 870  -2.42 2017-11-27 20:58:46
#> 871  -2.34 2017-11-27 20:56:18
#> 872  -2.37 2017-11-27 20:53:50
#> 873  -2.39 2017-11-27 20:51:22
#> 874  -2.52 2017-11-27 20:48:54
#> 875  -2.48 2017-11-27 20:46:26
#> 876  -2.07 2017-11-27 20:43:59
#> 877  -2.01 2017-11-27 20:41:31
#> 878  -2.00 2017-11-27 20:39:02
#> 879  -1.97 2017-11-27 20:36:34
#> 880  -1.96 2017-11-27 20:34:06
#> 881  -2.01 2017-11-27 20:31:38
#> 882  -1.87 2017-11-27 20:29:10
#> 883  -1.84 2017-11-27 20:26:41
#> 884  -1.84 2017-11-27 20:24:13
#> 885  -1.81 2017-11-27 20:21:45
#> 886  -1.90 2017-11-27 20:19:16
#> 887  -1.97 2017-11-27 20:16:48
#> 888  -2.01 2017-11-27 20:14:20
#> 889  -1.93 2017-11-27 20:11:52
#> 890  -1.83 2017-11-27 20:09:24
#> 891  -1.93 2017-11-27 20:06:56
#> 892  -1.98 2017-11-27 20:04:27
#> 893  -1.89 2017-11-27 20:01:59
#> 894  -1.81 2017-11-27 19:59:31
#> 895  -1.80 2017-11-27 19:57:03
#> 896  -1.74 2017-11-27 19:54:35
#> 897  -1.70 2017-11-27 19:52:07
#> 898  -1.72 2017-11-27 19:49:39
#> 899  -1.59 2017-11-27 19:47:11
#> 900  -1.56 2017-11-27 19:44:42
#> 901  -1.60 2017-11-27 19:42:12
#> 902  -1.54 2017-11-27 19:39:42
#> 903  -1.38 2017-11-27 19:37:10
#> 904  -1.26 2017-11-27 19:34:41
#> 905  -1.22 2017-11-27 19:32:11
#> 906  -1.18 2017-11-27 19:29:42
#> 907  -1.14 2017-11-27 19:27:13
#> 908  -1.11 2017-11-27 19:24:43
#> 909  -1.06 2017-11-27 19:22:14
#> 910  -1.03 2017-11-27 19:19:45
#> 911  -1.03 2017-11-27 19:17:15
#> 912  -1.00 2017-11-27 19:14:46
#> 913  -0.93 2017-11-27 19:12:18
#> 914  -0.94 2017-11-27 19:09:50
#> 915  -0.96 2017-11-27 19:07:21
#> 916  -0.94 2017-11-27 19:04:54
#> 917  -0.83 2017-11-27 19:02:26
#> 918  -0.89 2017-11-27 18:59:58
#> 919  -0.89 2017-11-27 18:57:30
#> 920  -0.84 2017-11-27 18:55:02
#> 921  -0.78 2017-11-27 18:52:34
#> 922  -0.73 2017-11-27 18:50:05
#> 923  -0.71 2017-11-27 18:47:37
#> 924  -0.72 2017-11-27 18:45:09
#> 925  -0.84 2017-11-27 18:42:40
#> 926  -0.77 2017-11-27 18:40:13
#> 927  -0.78 2017-11-27 18:37:45
#> 928  -0.75 2017-11-27 18:35:17
#> 929  -0.66 2017-11-27 18:32:49
#> 930  -0.57 2017-11-27 18:30:21
#> 931  -0.37 2017-11-27 18:27:53
#> 932  -0.39 2017-11-27 18:25:25
#> 933  -0.47 2017-11-27 18:22:57
#> 934  -0.56 2017-11-27 18:20:29
#> 935  -0.40 2017-11-27 18:18:01
#> 936  -0.35 2017-11-27 18:15:33
#> 937  -0.30 2017-11-27 18:13:05
#> 938  -0.46 2017-11-27 18:10:37
#> 939  -0.42 2017-11-27 18:08:09
#> 940  -0.48 2017-11-27 18:05:41
#> 941  -0.39 2017-11-27 18:03:13
#> 942  -0.50 2017-11-27 18:00:45
#> 943  -0.47 2017-11-27 17:58:18
#> 944  -0.46 2017-11-27 17:55:50
#> 945  -0.35 2017-11-27 17:53:22
#> 946  -0.43 2017-11-27 17:50:54
#> 947  -0.24 2017-11-27 17:48:26
#> 948  -0.24 2017-11-27 17:45:58
#> 949  -0.17 2017-11-27 17:43:30
#> 950  -0.07 2017-11-27 17:41:03
#> 951  -0.21 2017-11-27 17:38:32
#> 952  -0.11 2017-11-27 17:36:05
#> 953  -0.18 2017-11-27 17:33:36
#> 954  -0.22 2017-11-27 17:31:08
#> 955  -0.09 2017-11-27 17:28:40
#> 956  -0.08 2017-11-27 17:26:12
#> 957   0.08 2017-11-27 17:23:44
#> 958   0.07 2017-11-27 17:21:17
#> 959   0.04 2017-11-27 17:18:49
#> 960   0.25 2017-11-27 17:16:21
#> 961   0.25 2017-11-27 17:13:53
#> 962   0.26 2017-11-27 17:11:25
#> 963   0.37 2017-11-27 17:08:58
#> 964   0.44 2017-11-27 17:06:30
#> 965   0.36 2017-11-27 17:04:02
#> 966   0.40 2017-11-27 17:01:34
#> 967   0.62 2017-11-27 16:59:06
#> 968   0.68 2017-11-27 16:56:37
#> 969   0.61 2017-11-27 16:54:09
#> 970   0.68 2017-11-27 16:51:42
#> 971   0.81 2017-11-27 16:49:14
#> 972   0.86 2017-11-27 16:46:46
#> 973   0.80 2017-11-27 16:44:18
#> 974   0.86 2017-11-27 16:41:50
#> 975   0.84 2017-11-27 16:39:21
#> 976   0.94 2017-11-27 16:36:53
#> 977   1.20 2017-11-27 16:34:25
#> 978   1.23 2017-11-27 16:31:57
#> 979   1.27 2017-11-27 16:29:30
#> 980   1.27 2017-11-27 16:27:02
#> 981   1.37 2017-11-27 16:24:34
#> 982   1.39 2017-11-27 16:22:06
#> 983   1.47 2017-11-27 16:19:38
#> 984   1.47 2017-11-27 16:17:11
#> 985   1.51 2017-11-27 16:14:43
#> 986   1.51 2017-11-27 16:12:15
#> 987   1.54 2017-11-27 16:09:47
#> 988   1.57 2017-11-27 16:07:19
#> 989   1.55 2017-11-27 16:04:52
#> 990   1.53 2017-11-27 16:02:24
#> 991   1.55 2017-11-27 15:59:56
#> 992   1.49 2017-11-27 15:57:27
#> 993   1.45 2017-11-27 15:54:59
#> 994   1.46 2017-11-27 15:52:31
#> 995   1.49 2017-11-27 15:50:03
#> 996   1.54 2017-11-27 15:47:35
#> 997   1.54 2017-11-27 15:45:07
#> 998   1.59 2017-11-27 15:42:39
#> 999   1.67 2017-11-27 15:40:10
#> 1000  1.49 2017-11-27 15:37:42
#> 1001  1.85 2017-11-27 15:35:14
#> 1002  1.98 2017-11-27 15:32:46
#> 1003  1.97 2017-11-27 15:30:18
#> 1004  2.09 2017-11-27 15:27:50
#> 1005  2.13 2017-11-27 15:25:21
#> 1006  2.11 2017-11-27 15:22:53
#> 1007  2.05 2017-11-27 15:20:26
#> 1008  2.08 2017-11-27 15:17:58
#> 1009  2.12 2017-11-27 15:15:30
#> 1010  2.08 2017-11-27 15:13:02
#> 1011  1.98 2017-11-27 15:10:34
#> 1012  2.07 2017-11-27 15:08:06
#> 1013  1.98 2017-11-27 15:05:38
#> 1014  2.18 2017-11-27 15:03:11
#> 1015  2.28 2017-11-27 15:00:43
#> 1016  2.32 2017-11-27 14:58:15
#> 1017  2.43 2017-11-27 14:55:47
#> 1018  2.51 2017-11-27 14:53:19
#> 1019  2.69 2017-11-27 14:50:52
#> 1020  2.58 2017-11-27 14:48:24
#> 1021  2.64 2017-11-27 14:45:56
#> 1022  2.78 2017-11-27 14:43:28
#> 1023  2.83 2017-11-27 14:41:00
#> 1024  2.73 2017-11-27 14:38:32
#> 1025  2.79 2017-11-27 14:36:05
#> 1026  2.83 2017-11-27 14:33:37
#> 1027  3.18 2017-11-27 14:31:08
#> 1028  3.26 2017-11-27 14:28:40
#> 1029  3.55 2017-11-27 14:26:13
#> 1030  3.53 2017-11-27 14:23:44
#> 1031  3.61 2017-11-27 14:21:16
#> 1032  3.43 2017-11-27 14:18:48
#> 1033  3.53 2017-11-27 14:16:20
#> 1034  3.70 2017-11-27 14:13:53
#> 1035  3.89 2017-11-27 14:11:25
#> 1036  4.14 2017-11-27 14:08:57
#> 1037  3.93 2017-11-27 14:06:29
#> 1038  4.02 2017-11-27 14:04:01
#> 1039  3.93 2017-11-27 14:01:34
#> 1040  4.03 2017-11-27 13:59:06
#> 1041  4.14 2017-11-27 13:56:38
#> 1042  4.26 2017-11-27 13:54:10
#> 1043  4.25 2017-11-27 13:51:42
#> 1044  4.39 2017-11-27 13:49:14
#> 1045  4.23 2017-11-27 13:46:46
#> 1046  4.17 2017-11-27 13:44:18
#> 1047  4.28 2017-11-27 13:41:50
#> 1048  4.17 2017-11-27 13:39:22
#> 1049  4.13 2017-11-27 13:36:55
#> 1050  4.24 2017-11-27 13:34:27
#> 1051  4.16 2017-11-27 13:31:59
#> 1052  4.38 2017-11-27 13:29:31
#> 1053  4.69 2017-11-27 13:27:03
#> 1054  4.64 2017-11-27 13:24:36
#> 1055  4.72 2017-11-27 13:22:08
#> 1056  4.70 2017-11-27 13:19:40
#> 1057  4.88 2017-11-27 13:17:12
#> 1058  4.85 2017-11-27 13:14:44
#> 1059  5.20 2017-11-27 13:12:16
#> 1060  5.02 2017-11-27 13:09:49
#> 1061  4.81 2017-11-27 13:07:20
#> 1062  4.72 2017-11-27 13:04:52
#> 1063  4.63 2017-11-27 13:02:24
#> 1064  4.53 2017-11-27 12:59:56
#> 1065  4.57 2017-11-27 12:57:29
#> 1066  4.53 2017-11-27 12:55:01
#> 1067  4.38 2017-11-27 12:52:33
#> 1068  4.16 2017-11-27 12:50:03
#> 1069  4.40 2017-11-27 12:47:36
#> 1070  4.39 2017-11-27 12:45:08
#> 1071  4.34 2017-11-27 12:42:40
#> 1072  4.58 2017-11-27 12:40:12
#> 1073  4.49 2017-11-27 12:37:44
#> 1074  4.46 2017-11-27 12:35:17
#> 1075  4.53 2017-11-27 12:32:49
#> 1076  4.38 2017-11-27 12:30:21
#> 1077  4.26 2017-11-27 12:27:53
#> 1078  4.24 2017-11-27 12:25:26
#> 1079  4.91 2017-11-27 12:22:58
#> 1080  4.91 2017-11-27 12:20:30
#> 1081  4.99 2017-11-27 12:18:02
#> 1082  4.85 2017-11-27 12:15:34
#> 1083  4.91 2017-11-27 12:13:07
#> 1084  4.82 2017-11-27 12:10:39
#> 1085  4.67 2017-11-27 12:08:11
#> 1086  4.93 2017-11-27 12:05:43
#> 1087  4.68 2017-11-27 12:03:16
#> 1088  4.41 2017-11-27 12:00:48
#> 1089  4.41 2017-11-27 11:58:20
#> 1090  3.93 2017-11-27 11:55:52
#> 1091  3.98 2017-11-27 11:53:23
#> 1092  3.79 2017-11-27 11:50:55
#> 1093  4.32 2017-11-27 11:48:28
#> 1094  4.65 2017-11-27 11:46:00
#> 1095  4.42 2017-11-27 11:43:32
#> 1096  4.58 2017-11-27 11:41:04
#> 1097  4.53 2017-11-27 11:38:37
#> 1098  4.97 2017-11-27 11:36:09
#> 1099  4.74 2017-11-27 11:33:41
#> 1100  4.89 2017-11-27 11:31:13
#> 1101  4.63 2017-11-27 11:28:46
#> 1102  4.22 2017-11-27 11:26:18
#> 1103  4.57 2017-11-27 11:23:50
#> 1104  4.61 2017-11-27 11:21:23
#> 1105  4.74 2017-11-27 11:18:55
#> 1106  5.29 2017-11-27 11:16:27
#> 1107  5.59 2017-11-27 11:13:59
#> 1108  6.38 2017-11-27 11:11:31
#> 1109  5.41 2017-11-27 11:09:04
#> 1110  6.11 2017-11-27 11:06:36
#> 1111  6.13 2017-11-27 11:04:08
#> 1112  7.41 2017-11-27 11:01:40
#> 1113  7.52 2017-11-27 10:59:13
#> 1114  7.04 2017-11-27 10:56:45
#> 1115  7.62 2017-11-27 10:54:17
#> 1116  7.52 2017-11-27 10:51:49
#> 1117  7.30 2017-11-27 10:49:22
#> 1118  7.00 2017-11-27 10:46:54
#> 1119  6.31 2017-11-27 10:44:26
#> 1120  5.88 2017-11-27 10:41:58
#> 1121  5.14 2017-11-27 10:39:31
#> 1122  4.67 2017-11-27 10:37:03
#> 1123  4.35 2017-11-27 10:34:35
#> 1124  4.35 2017-11-27 10:32:07
#> 1125  4.17 2017-11-27 10:29:39
#> 1126  4.22 2017-11-27 10:27:12
#> 1127  4.33 2017-11-27 10:24:44
#> 1128  4.25 2017-11-27 10:22:16
#> 1129  4.35 2017-11-27 10:19:48
#> 1130  4.33 2017-11-27 10:17:21
#> 1131  4.27 2017-11-27 10:14:53
#> 1132  4.28 2017-11-27 10:12:25
#> 1133  4.33 2017-11-27 10:09:57
#> 1134  4.57 2017-11-27 10:07:29
#> 1135  4.45 2017-11-27 10:05:01
#> 1136  4.20 2017-11-27 10:02:33
#> 1137  3.97 2017-11-27 10:00:05
#> 1138  4.03 2017-11-27 09:57:38
#> 1139  4.59 2017-11-27 09:55:10
#> 1140  5.04 2017-11-27 09:52:40
#> 1141  4.63 2017-11-27 09:50:12
#> 1142  4.25 2017-11-27 09:47:44
#> 1143  3.44 2017-11-27 09:45:16
#> 1144  3.31 2017-11-27 09:42:48
#> 1145  3.59 2017-11-27 09:40:21
#> 1146  3.90 2017-11-27 09:37:53
#> 1147  3.52 2017-11-27 09:35:25
#> 1148  3.18 2017-11-27 09:32:57
#> 1149  3.07 2017-11-27 09:30:30
#> 1150  2.83 2017-11-27 09:28:02
#> 1151  2.79 2017-11-27 09:25:34
#> 1152  2.85 2017-11-27 09:23:06
#> 1153  2.54 2017-11-27 09:20:39
#> 1154  2.70 2017-11-27 09:18:11
#> 1155  2.33 2017-11-27 09:15:43
#> 1156  2.70 2017-11-27 09:13:15
#> 
#> 

# NOT RUN {
get_senseBox_data(
  senseBoxId = c("5957b67494f0520011304cc1","5957b67494f0520011304cc1") ,
  sensorId = "all",
  fromDate = "2017-11-25-12:00:00",
  toDate = "2017-11-26-12:00:00")

get_senseBox_data(
  senseBoxId = "5957b67494f0520011304cc1",
  sensorId = "5957b67494f0520011304cc4",
  fromDate = "2017-11-25-12:00:00",
  toDate = "2017-11-26-12:00:00")
# }